Add getRecipes() to getAllPosts() and update generatePostIndex() to work
Working version of generatePostIndex()
