+++ public/scripts/postIndex.js
@@ -1,1 +1,1 @@
-export default [{ id: 0, directory: "Musings", title: "10 beginners mistakes to avoid when designing for print", subtitle: "I made them too and will probably make them again", slug: "/posts/10-beginners-mistakes-to-avoid-when-designing-for-print", dateCreated: "2019-05-27",  content: " If you are an innocent graphic designer (a.k.a. mostly used to the web), print design might seem like a super scary and intimidating world. It is. When designing for print, you are bound to make mistakes as a beginner. There are just  too many  things to think about. But don't fret. The world of print design is also fascinating and deeply satisfying — just imagine holding your freshly printed [insert your project name] 😍 Get started, experiment, and welcome to the world of print design! (But try to avoid these mistakes if you can… 🤞) 1. Don't use RGB colors, design in CMYK.    The RGB colors you've been using (almost) all your life on screen are based on light. Red, Green and Blue lights are added together to reproduce other colors. However, your printer is much more limited and won't be able to reproduce all the colors you see on screen. It's because it has to mix the four CMYK inks (Cyan, Magenta, Yellow and Key) to obtain a new color. So what?  Make sure to select the CMYK color mode, when creating your artwork. And if possible, always base your color decisions on a physical CMYK or Pantone color chart. 2. Your images resolution is too low.    Simply inserting this super cool image you found on the web into your print design can often lead to disaster. If you don't want to get a very pixelated result, your image resolution should always be between 300 ppi (pixels per inch) and 400 ppi. So what?  Even if the image quality seems to look good on screen, it might not be the case in print. Always check that your artwork resolution is above the 300 ppi threshold. 3. You haven't thought about the bleed and quiet borders.    Need to print to the border very edge? Set up a  bleed  before creating your design. The bleed is an outer-margin that you set around your document borders. All the graphics that touch the borders will be slightly extended to the bleed limit, before being trimmed to the sheet size after print. Thanks to the bleed, you won't have any unwanted white border appearing along the edges of the artwork. The  quiet border  is a safety margin inside your document, where no essential elements — such as text or logo — should be placed. This border is necessary as the trimming process is not always precise and can cut off important information. So what?  Set up a bleed of 3–5 mm and a quiet border of 10–15 mm for a regular A4 paper. These limits should be adjusted to your printer and paper size. As a rule of thumb, the bigger the sheet, the wider the bleed / quiet border. 4. You don't really know how to use black.    You are writing down your copy, selecting black ⬛ as a text fill color and are ready to print! Well,  maybe not . Double-click on the black fill you used and check the CMYK code. It might be something like C75-M68-Y67-K90. Using all four inks to write in black can result in blurry text — and that's bad. What you really need is to use the following code: C0-M0-Y0-K100. However, when applied on a larger area that black tone will look almost grey. For best results, switch to a  rich black  by mixing several ink colors to get a darker tone. There is not one recipe, but try to use the C40-M30-Y30-K100 CMYK code for those large areas. So what?  Use a crisp black C0-M0-Y0-K100 for small text, and a rich black for large black areas C40-M30-Y30-K100. If possible, directly check with your printer what their specifications for rich black are. 5. Your characters are too thin.    Depending on your printer capacities, your very small and thin characters might not be printed at all. So be careful with fine hairlines and tiny text! Another common mistake is to use small text on a rich black background. The slight misalignment of printing plates might make your characters illegible. When you have white text over black, use a C0-M0-Y0-K100 simple black to avoid this possible blur 🏃‍♂️🏃‍♂️🏃‍♂️ So what?  Don't use tiny text in your artwork. 6pt character size is a good rule of thumb — even though it all depends on your typeface style! When using small white text over black, choose a C0-M0-Y0-K100 simple black. 6. Haven't proofread your file?    Yes, proofreading is boring. It is so easy to publish a blog post, notice a typo, and correct it afterwards. However, spelling mistakes are the worst in print design. When printing hundreds of posters, that small typo reproduced over and over again might not seem so trivial anymore 😨 So what?  Always and systematically proofread and review your artwork before sending it to the printer. As a designer, also check for spelling mistakes and typos even if you haven’t written the original copy. 7. You haven’t embedded your graphics and text.    Your artwork is looking good and ready for print. You are especially happy about this super fancy font that you’ve found on Creative Market! 💁‍♀️ Just sending your design for print (even in PDF format) is not always a good idea. In all probability, the printers you are working with won’t have that fancy font and some problems might occur — some characters might be replaced, or disappear altogether. Make sure to embed all fonts in your artwork, before sending it to your printer. Outlining any stroke in your artwork to convert them into shapes and flattening all your work layers are also good practices. So what?  Don’t forget to embed your fonts, outline your artwork strokes and flatten your layers. 8. You forgot about ink coverage.    To print an artwork using CMYK colors, your printer will use four plates of inks that combined together in dot patterns will make up an image. Your printer and paper will often have an ink coverage limit. For example, let’s say you apply a rich black C70-M70-Y70-K100 on a large area. That area ink coverage would amount to 70+70+70+100=310%. Most standard printers can handle an ink coverage of 260%. Going over this limit can result in quality problems and warped paper. So what?  Check with your printer what their ink coverage is, and make sure that the colors you use fit this limit. 9. You’ve been using the wrong format.    Are you about to export your artwork into PNG or GIF files for print? 😱 These formats were developed for screen only. They have a resolution that is too low for print, as well as a reduced color capacity. So what?  For best results, export your file into a PDF format before sending it to your printer. 10. You thought a hard-copy proof wasn’t necessary.    Getting a hard-copy proof from your printer is not always necessary. But the bigger your project is, the more critical it becomes to see a physical example of your work before printing 10,000 other copies. Also make sure to use your printer color chart if they have one! So what?  Get a hard-copy proof of your work and check with your client that everything is perfect before scaling up.  Alright, you now know the basics and are ready to go — all by yourself 🤧 Do your best {end-key}"},{ id: 1, directory: "Musings", title: "#10 On obsessions", subtitle: "Update #10", slug: "/posts/10-on-obsessions", dateCreated: "2021-01-10",  content: "    Meh print  by Ryan Putnam   Life peek: On obsessions  Since the last week of December, I have been  obsessively  working on my website. Being pretty new to coding, I am always discovering exciting things to try out in this new playground. The downside is when the obsession gets out of control. It is not uncommon that I stay up until late in the evening building a non-essential feature. Feeling excited to work on my website is good. The problem is that it replaces other things that I deem important: thinking, reading, cooking, writing, AFK time. In the coming week, I will set up a few healthy habit trackers. I have — unsuccessfully — done so countless times in the past. The difference this time is that it will be public to keep me accountable. Until next week for more info! Read &#x26; Watched   Maniac (miniseries, 2018)   I am pretty picky in terms of series and it is not unusual for me to stop one a few episodes in.  Maniac  hooked me from the start. In two words, I would describe it as  retro futuristic . It was directed by Cary Joji Fukunaga and I just love his attention to detail. Antilibrary   Straw Dogs (2003)    Straw Dogs  by the English philosopher John Gray challenged Norbauer's view on  pretty much everything . The book is a critique of philosophical humanism.  Recommended by Ryan Norbauer in  Dense Discovery   Discovered   Dsxyliea   This website simulates the reading experience of a person with dyslexia. The letters  seem to jump around . I think that it looks a bit too extreme — first and last letters of a long word swapping, but it is an important reminder of how different our experiences of the world can be. Overheard   HTTP status codes give us all the vocab to talk about dating 300 multiple choices, 303 see other, 400 bad request, 404 not found, 408 request timeout, 409 conflict, 410 gone, 412 precondition failed, 413 request entity too large, 417 expectation failed, 500 internal server error — Tatiana Mac  {end-key}"},{ id: 2, directory: "Musings", title: "#11 Habits in the oven", subtitle: "Update #11", slug: "/posts/11-habits-in-the-oven", dateCreated: "2021-01-17",  content: "    Addiction 1/2 — Scroll  by Timothée Bart   Life peek: Habit tracking  Instead of proactively working on my new habits this last week, I procrastinated — again — by setting up overkill habit trackers. I call them overkill because my habit goals are pretty simple:   Easy : I want to be up by 7am every day.  Medium : Take at least two 15-minute breaks on work days.  Hard : Turn off laptop by 7pm every day.  I told you about my  recent obsession  with building this website. A consequence is that I spend way too much time on my computer. To gather some stats about my screen time, I installed  TimeFibers  on my laptop and the  StayFree app  on my phone.  1   Every month now, I will try to give a completely transparent habit report to keep me accountable. Most importantly, I will start  actually  doing some habits this week — pinky swear. Antilibrary   1984 (graphic novel, 2020)   While browsing books on Open Library, I stumbled upon this graphic novel version of  1984  by Fido Nesti. This being a graphic novel, I pre-judge it by its cover and don't think twice before adding it to my antilibrary.  The Art of Fermentation (book)   In  The Preserve Journal Issue 4 , journalist Kaszmarzyk says  I believe everyone goes through the same thing when they receive their first copy of  The Art of Fermentation : you spend days with it. You bond and form a relationship like no other with it, because it's neither a recipe book nor a nonfiction novel, it is its own special world.   Classic Food of China (book)   I have been digitally browsing a lot of cookbooks recently. Chinese cooking expert Fuchsia Dunlop warmly recommends  Classic Food of China  by Yan-Kit So. What I like about this book is that it's not only about food as it also gives insights about culture and history. Discovered   Practical Typography   I have just stumbled upon this gem again as I was sorting through my old bookmarks. I immediately added it to my  bookmarks . It is not a website, nor a book, but a treasure chest of practical information about typography. {end-key}  Footnotes    A digital solution to a digital problem? Yep, I'm not sure it's a good idea either.  ↩   "},{ id: 3, directory: "Musings", title: "#12 Culinary experiments", subtitle: "Update #12", slug: "/posts/12-culinary-experiments", dateCreated: "2021-01-23",  content: "    Breakfast  by Essi Kimpimäki   Life peek: Culinary experiments  I love trying out new recipes — especially the ones that are about creating things that you usually buy. This week, I baked some sourdough croissants. For a first time, it wasn't too bad: not airy, but not dense either, a bit too buttery, but so satisfying. Robin and I had some plain croissants for breakfast, but we also tried a grilled cheese version which was deliciously decadent. Another recipe I tried this week are those  vegetarian Polish cabbage rolls . It is not easy finding out winter dishes that are seasonal and that make me excited. These rolls did the job perfectly.   They also exist in Germany as  Kohlrouladen , but I stumbled upon the vegetarian Polish version first.  Read &#x26; Watched   Crazy Rich Asians (movie, 2018)   A light rom com that gives just what you are expecting from it. It is cheesy, overacted, but very entertaining. (And you just have to love Awkwafina.)  Blade Runner (movie, 1982)   I finally watched this classic that wonderfully depicts the year 2019: dark and polluted, people from all nationalities mixing, not one dominant language, strong Japanese video game feel. Human-like and life-like androids are a thing. People still have retro haircuts and clothes — of course. I loved the strong vibrations of the soundtrack. Tonight, we are watching the 2049 edition. Discovered   Musixmatch   Musixmatch is another one of these tools that have been lying around forever and that I just discovered recently. It connects with your Spotify account and shows you the synced song lyrics. We have already had a few spontaneous karaokes and I love it.  Five Books   I discovered Five Books through  Dunlop's interview  about Chinese food books. The concept is amazing: asking experts to recommend five books in their subject. The result is a curated list of very good books in any fields you can think of. Overheard   Bier ist so viel mehr als nur ein Frühstücksgetränk. — Kiezbalkon in Friedrichshain, Berlin  {end-key}"},{ id: 4, directory: "Musings", title: "#13 Holiday at home", subtitle: "Update #13", slug: "/posts/13-holiday-at-home", dateCreated: "2021-01-30",  content: "    PREPARATION  by Meghdeep Sarkar   What a week! I feel like so many things happened since last week. Life peek: Holiday at home  This week was a bit special as both Robin and I took time off work without any holiday plans. We spent most of it having outdoor walks, doing pizza-related activities and working on our personal projects.  Yes , I did spend a disproportionate amount of time working on my website.  1   I love adding new features to my website and learning to code. The problem is that this website should be a  side -side-project at best. The reality is that it takes time away from things I would like to focus on. Something I stopped doing is my routine of reading essays / working on my digital garden. Lately, I am learning a lot coding-wise but not much outside of that. In that direction,  The Knitting Club  that Mirha Masala launched with two of her friends inspired me a lot. Every three weeks, they choose a topic, a skill or an activity. The three friends then share their experiences and discuss their learnings. I love this idea of having a  learning club  where you grow together and keep each other accountable. Read &#x26; Watched   Kubo and the two strings (2016)   Beautifully crafted stop-motion by the Laika studio, combining traditional stop-motion techniques with CGI. It is funny and poetic with just the right amount of darkness.  Kubo and the two strings  is everything but just a kid movie. Antilibrary   Defining Decade (2013)   The complete title is the very concise  Defining Decade: Why Your Twenties Matter And How To Make The Most Of Them Now  (by Meg Jay). Mirha Masala recommends it as the one book that marked 2020 for her.  We think that by avoiding decisions now, we keep all of our options open for later — but not making choices is a choice all the same.  T(W)IL  This week I learned a few new things: To use an API  The Movies page looks pretty much like the existing Books page. The difference? I coded it all by myself and had to learn to use the TMDb API on the way there. I had been trying to do it for weeks, so I am quite happy to have finally managed it! To make a sourdough pizza  Robin and I decided that this past week was going to be  Pizza Week . We tried lots of good pizzas, made some pizza-related powerpoints and also learned to make our own with Andrew Janjigian's  sourdough recipe   2  . The result was delicious but it could definitely be improved — even with our tiny microwave oven. {end-key}  Footnotes    Like that night I couldn't sleep and coded until 1am.  ↩   Janjigian is one of these “cooking creators” who don't just give you a flat recipe — but also the story or science behind it. I just love these new cooking channels and blogs that go further than necessary.  ↩   "},{ id: 5, directory: "Musings", title: "#14 Books, books, books", subtitle: "Update #14", slug: "/posts/14-books-books-books", dateCreated: "2021-02-07",  content: "    Mariah Idrissi, Badass. by Maggie Bignell    Life peek: Books, books, books  Since the start of the year, I haven't made much time for reading. I guess I just couldn't find a book that really hooked me — a book that I was looking forward to going back to. I finally found it this last week. It is called  Girl, Woman, Other  by Bernadine Evaristo. I had never heard of it before, never read anything about it, and simply began reading it with my company book club. I was sucked in and finished it in a few days time.  1   Another book that marked this week is  The Witches  by Roald Dahl. In the evening, I put on some background music, turn on some warm lights and Robin and I take turns, reading the book aloud to each other. We used to do the same with  Le Petit Nicholas  by Jean-Jacques Sempé, a series of French children's books — that are also great for adults. One last book-related update: I have finally imported all my books from Goodreads to my own personal bookshelves. It feels liberating to completely own my book data for the first time in a while! Read & Watched   Girl, Woman, Other (2019)   Beautifully written in a stream of consciousness style. We jump into the lives of 12 girls, women — others — in England. From one mind to the next, we get to know and understand them through each other's perceptions, discovering how it is never as simple as it seems. Strong recurring themes are feminism, politics, sexuality and otherness. The characters are “othered” by society for their skin color, sexuality, gender — sometimes even othering one another.  Do the right thing (1989)   I am almost ashamed to say that we chose to watch this movie because it had a pizza theme. Our  Pizza Week  brought us many joys, and  Do the right thing  was definitely a surprise hightlight. Spike Lee directed it in the 80s, but its topics are still — sadly — very relevant today. For me, this movie was  fluid . Everything glides and moves, up, up, keeping you in your seat until the end. One recommendation: choose your movies based on what you eat. Food for Thought   Environmental Impacts of Food (2020)   This might be the most thought-provoking article I have read in the past few months.  Environmental Impacts of Food  was written based on the largest meta-analysis on global food systems to date — with data from 38,000 farms and 119 countries. From the analysis, it is pretty clear that what you eat is much more important than how local the produce is. Transportation and packaging account for a very small percentage of the food emissions. What surprised me most is that this is also the case for low-impact meat producers. As a rule of thumb, it is usually better to buy vegetables from across the globe, rather than meat and dairy produces from your local farm. T(W)IL  To bake an amazing banana bread  Robin lived in London for one year and introduced me to this amazing tea-time snack: banana bread. It is warm, slightly toasted, topped with a slice of melted butter. This weekend, I tried baking some banana bread, following  the recipe  a friend from work shared with me. We don't have baking powder at home and made a sourdough version instead. Conclusion: 5 stars of deliciousness that will certainly become a kitchen-regular. My first steps with D3.js  November last year, I gave into the Cyber Week craze and got a discounted version of the  D3.js &#x26; Firebase  MOOC. For context, I knew almost nothing about JavaScript at that time and realized that the course was way above my level. I started the course this last week and now understand all of it without much difficulty. It is incredible to see how far I've come since the end of 2020. I am quite excited to finally learn more about creating dynamic and interactive data visualizations! {end-key}  Footnotes    More about it in the  Read &#x26; Watched  section below.  ↩   "},{ id: 6, directory: "Musings", title: "#15 Walk on the frozen canal", subtitle: "Update #15", slug: "/posts/15-walk-on-the-frozen-canal", dateCreated: "2021-02-14",  content: "    Coffee shop during winter by Jenny Lelong    Life peek: Frozen canal &#x26; digital garden  The last week in Berlin has been freezing cold — think -10 °C. It was so cold that we were able to walk on the frozen canal. This was the first time in my life doing this kind of thing in a city center. The air was biting, the sun was shining, while we walked on the ice eating pastries.  It has now been a few months I am trying to build my personal knowledge management system — a.k.a. my digital garden as the trend calls it. I was first inspired by  Andy Matuschak  and  Maggie Appleton 's gardens. The idea is to have a place to record my learnings, connect thoughts and hopefully see a thinking companion emerge from it — not there yet, but fingers crossed. Yesterday, I was finally able to turn my offline digital garden into a live one. It has long been a private project, so it is currently not the easiest to navigate. I am building the habit of nurturing it, so come back in a bit and you might see more! Read &#x26; Watched   Bacurau (2019)   Bacurau was... unexpected. I watched it as some Brazilian friends recommended it to me, without knowing anything about its plot. It starts a few years from now in a tiny isolated village in rural Brazil. It is a direct reference to  quilombos  — historical places of resistance where slaves ran away from their masters during Portuguese colonial times. Midway through the movie, we suddenly realize we are in a dystopian future, mixing genres from action to sci-fi. Very original, slow and tense, a critic of its times. Food for thought   The Great Unbundling (2020)   In this interview of Casper ter Kuile, they talk about the unbundling process — or how spirituality, community and rituals that were once part of religious institutions, are taking on new contemporary forms. People now find community in their fitness communities, justice groups or workplace. Young people are less and less religious, but still yearn for spirituality — moments infused with meaning and intention. I have just started his book  The Power of Ritual . In it, Casper explains how you can bring back meaning to things you do everyday—it might be your brewing coffee routine, meditation or journalling practice. This idea of making things meaningful and intentional really speaks to me and I am looking forward to going back to the book — which doesn't happen often with nonfiction. T(W)IL  To use D3.js with React  This last week, I tried to integrate what I first learned of D3.js with React. Here is the  example of a treemap  I created using my movies data. Each box represents a movie genre. It is obviously still a work in progress, but I am already quite pleased with the first results. {end-key}"},{ id: 7, directory: "Musings", title: "#16 Sunbathing on planet Earth", subtitle: "Update #16", slug: "/posts/16-sunbathing-on-planet-earth", dateCreated: "2021-02-22",  content: "    The Hiding-in-the-Wrong-Place Club by Andy Wang    Life peek: Sunbathing on planet Earth  Last week, Robin and I were  walking on the frozen canal . This weekend, we took off our hats, coats and gloves, had a long walk outside that finished in the best possible way — with a cup of coffee in the sun. On February 18, 2021, NASA's Perseverance rover landed on Mars. At the exact same time, something else was happening. I was completing the first illustration of my  Tiny Worlds  project. Have a look at the  final SVG . Read &#x26; Watched   Sourdough (2017)}   Robin Sloan has a way of mixing genres.  Sourdough  is both very modern and slightly fantastical. Even though, the ending left me a bit  meh , I still really liked the overall book and plot. This is my favorite book by Robin Sloan so far.  Onward (2020)   Many movie gems passed right under our noses in 2020,  Onward  is one of them. It is set in a fantasy world where magic has been forgotten, where unicorns are another kind of rats and elves use electricity. Due to some circumstances, an unassured young elf goes on an epic treasure hunt with his brother. It is a journey of self discovery, bonding and coming of age. T(W)IL  To make Biang Biang noodles  Biang Biang noodles come from the Shaanxi province in China. They are deliciously chewy and are super simple to make. This weekend, Robin and I followed  this recipe  by  My Name is Andong . We had a lot of fun making them, and even more eating them. To bake some pastéis de nata  After the noodles, we needed a dessert. A friend had recommended  this recipe  to me by  Babish Culinary Universe . I like how he goes through the extra effort of making the  laminated dough  — just learned that English word — from scratch. The result was much better than I had expected. It was maybe slightly too sweet for me (and let's not even talk about how my microwave-oven stops at 230 °C only). {end-key}"},{ id: 8, directory: "Musings", title: "#17 Back on tracks", subtitle: "Update #17", slug: "/posts/17-back-on-tracks", dateCreated: "2022-11-18",  content: "    Bouldering illustration by Karolin Chen    Oh my, oh my. So I kind of took the decision to stop these weeknotes more than a year ago. Nowadays, I am slowly feeling the need to come back to them. I remember that they sometimes felt like chores, but they also helped to anchor me in the present. Well, in any case, I am giving this habit a second (or third/fourth?) chance. I'll keep on calling them weeknotes even though I am not really expecting to follow a strict schedule. Life peek  These last few weeks have been getting colder in Berlin. Tonight, we are expecting the temperature to drop under 0 °C. When we went bouldering this morning, we could see a thin layer of snow on cars and terrace tables. My routine of going bouldering in the early morning with friends, followed by pastries at   La Maison    1  , is still one of my great highlights. It will probably be the thing I'll miss the most when we'll move out of Berlin next February.  Otherwise, I finally completed my first ever D&#x26;D campaign last weekend — what an epilogue! I learned a lot, and we are ready to kick start the next campaign straight away. I still need to think about my character, but am excited to try to get inside the head of someone new. Read &#x26; Watched   Lives of the Mayfair Witches (1990)   What a ride! I finally finished reading this supernatural horror trilogy of +2000 pages. Honestly, I don't know how I did it. It was half fascinating, half disturbing. In the end, I wouldn't even particularly recommend it (the disturbing parts were too much for me). However, once you are caught up in the story, you cannot really stop as you  need  to know what happens next.  Adieu Bohème (2017)    Adieu Bohème  is a beautiful short-film that will leave you awed. As the film comes to an end, you realize that it was shot  and  broadcast live from the 6th basement of the Bastille Opera House. Outstanding and surreal performance. T(W)IL  To make some Italian-approved carbonara  A few weeks ago, we invited over some Italian friends for a wine-tasting and carbonara-making session. It was super fun to see them argue over the technicalities of making this very simple — oh but so controversial — dish. This is more or less the  recipe we followed  — even though they definitely didn't need a recipe. To lead-climb  Last weekend, Robin and I completed our lead climbing course at  Magic Mountain . It is actually not that complicated, but we did learn a number of safety habits that we'll need to integrate in the long term. The weight difference between Robin and I (about 40 kg) was also totally fine, as we learned to use the Ohm that was made for this specific reason. We'll go back to lead climbing by ourselves on Monday to keep practicing. {end-key}  Footnotes    Amazing French bakery in Berlin with a super nice staff  ↩   "},{ id: 9, directory: "Musings", title: "#18 Friendsgiving and D&D", subtitle: "Update #18", slug: "/posts/18-friendsgiving-and-dnd", dateCreated: "2022-11-28",  content: "    Man and goose by Max Hodlevskyi    Life peek  On Saturday, I was part of my first ever Friendsgiving  1  . In Europe, Thanksgiving is just this remote Christmas-like event that you only see in movies, so I didn't really know what to expect. In short, my American (obviously) friend took upon herself to organize the most incredible homemade menu I have ever seen on a dinner table. We had everything that you can expect for Thanksgiving, even a goose — couldn't find a Turkey—that we named Larry (David) Goose. I feel so grateful to be surrounded by close friends in a city where I have only lived for a couple of years.  On Sunday, we started a new D&#x26;D mini-campaign! It is my second one only, so it feels very exciting to play a new race, class, and personality. I am playing a Tiefling Rogue and already love all the special powers that come just from the race.  2   Read &#x26; Watched  The Truffle Hunters (2020)   The Truffle Hunters  is an Italian documentary film following a group of old men hunting truffles with their dogs. The shots are incredibly beautiful — they are staged, yet very authentic. The documentary shows the decline of truffle hunting, but doesn't make a clear argument as to which side it roots for. I would definitely recommend it if you want to witness an emotional and bittersweet slice of life. Werewolf by night (2022)  So... I am into Marvel. During the first lockdown, Robin and I proceeded to watch all the MCU movies in an order that made sense plot-wise. Now that we are up to speed, we just  cannot  miss a new Marvel movie anymore. More often than not, those movies will be bad  *cough* Black Widow *cough* . But even the bad ones add a new puzzle piece to the entire MCU universe.  Werewolf by night  is a short-ish movie (53 minutes) that came out this year. I found the plot quite straightforward, but I did love the black and white, very Hitchcock-like rendition. If you want to have a short but entertaining evening, I can only recommend it! T(W)IL  To make a green bean casserole  For my friend's Thanksgiving dinner, I prepared a green bean casserole — even though I had never tried one and didn't know what it was supposed to taste like. Reading through  the recipe , it looked simple enough and surprisingly close to... a green bean grâtin? I guess that might just be my French upbringing lense distortion, but the recipe felt a lot more approachable once I made this connection in my mind. I would definitely try to make more of these veggie-heavy casseroles / grâtins in the future! {end-key}  Footnotes    I also, just discovered that  Friendsgiving  is apparently a word.  ↩   Trying to intimidate a baddie with advantage by using my loud voice boomed up 3 times — done.  ↩   "},{ id: 10, directory: "Musings", title: "#19 Sprained ankle and interactive fiction", subtitle: "Update #19", slug: "/posts/19-sprained-ankle-and-interactive-fiction", dateCreated: "2022-12-05",  content: "    Arcade Astronaut by Jordan Jenkins    Life peek  So, my life just took an unexpected turn on Friday morning. As usual, I started the day bouldering with friends. As I was doing a very non-dangerous move  1  , I slipped and fell on the side of my ankle with a lot of speed and strength. It hurt  a lot . Of course, we still had to follow up with our normal routine of getting breakfast first at a local café — because  priorities  — before taking any drastic decisions, like going to the doctor. Stuffed with pastries and coffee, Robin finally took me on a piggyback ride to the closest doctor. Diagnostic: a very badly sprained ankle and 4–6 weeks without sport. And that'd only be if the ankle was  not  fractured. Which luckily it was not.  2   Needless to say, a lot of my plans for this month have been reviewed after the  catastrophe . Fortunately, I have a local wine bar five minutes from home where we are welcomed like friends. It is nice to build these local connections in the place you live in, from the coffee shop waiter to the grocery lady. It makes me a bit sad to leave them behind once we'll move away from Berlin. Discovered  An immobilised kind of lifestyle calls for more quiet activities. Fortunately, the list of winners for this year  Interactive Fiction Competition  is out! I proceeded to play three games, which totaled to about 5–6 hours. The Absence of Miriam Lane   Sometimes people give pieces of themselves away. Sometimes they give too much and who they are wears thin. They become an absence. A hole in the world. And a terrible Light shines through.    The Absence of Miriam Lane   is a choice-based game  3  . The way it was implemented on its plaform is amazing. The black and white imagery, the soundtrack create a very atmospheric environment, in which you are enclosed until the end of the story  4  . A Long Way to the Nearest Star   A thief on the run from the galactic police finds refuge on an abandoned spaceship. A lonely ship AI finds unexpected company.  Similarly,   A Long Way to the Nearest Star   is a choice-based game. The use of the platform is not as creative, but definitely helps with the setting. The difference between your thoughts and the things you read on digital devices (usually the AI talking to you) are stylized perfectly. The game is also quite lengthy — if you are a slow player like me — and left me with this wonderful feeling of coming out of a very good and absorbing story. And Then You Come to a House Not Unlike the Previous One   Wander through an apparently infinite number of spooky mansions, solving a basic puzzle in each one.  Finally,   And Then You Come to a House Not Unlike the Previous One   is a parser-based game  5  . The official synopsis doesn't tell you much and actually hides the fact that the adventure actually  has a meaning . Starting this game, I was more or less ready to solve a hundred problems, one after the other. However you quickly realize that the story is more complex than that... T(W)IL  To make Knödel  One of the things that inspire me to cook at home is going to the restaurant, eating something delicious and being like: I want to know how to make this. Well, that happened just last week with  Knödel . A few days later, I had made a fresh batch of my own that was not bad at all. Basically, they felt like big gnocchi where you simply swap the potatoes for bread instead.  6   To make lasagna  Yay, lasagna! How come I had never tried before? Not sure, but I will definitely have another go at it very soon. It was much easier than I expected it to be, and the result is just so good! {end-key}  Footnotes    For those interested in the technicalities, it was one of those run along the wall on a long volume move. However, as I kept my eyes on the prize — the final hand hold — my foot just  missed  the volume at some point and I fell to the floor... 15 cm below.  ↩   But we found that out the long way around after sticking around in the emergency room for 7 hours, yay.  ↩   In choice-based game you click on existing links to make your decisions.  ↩   I lost the game, but it is still worth playing.  ↩   In parser-based games, you need to spell out your actions through specifc commands.  ↩   I feel like both Germans and Italians would kill me for saying that.  ↩   "},{ id: 11, directory: "Musings", title: "#20 Getting ready for Christmas", subtitle: "Update #20", slug: "/posts/20-getting-ready-for-christmas", dateCreated: "2022-12-16",  content: "    Chojny Train Station by Krzysztof Nowak    Life peek  So, I'm writing this from  Switzerland … or actually 4 km across Switzerland border on the French side.  1   Robin and I are back in cows land for family reunions time. Yesterday was a long train day. We left our Berlin home at 7am in the morning and arrived at Robin's parents' place 12 hours later. Thankfully, German trains are equipped with wifi, so we have plenty of things to do during the ride  2  . The last week has felt intense. We had a bunch of stuff to prep and do before leaving Berlin, and I was pretty busy with work as well — in a good way. All that rush culminated to a wonderful note on Wednesday evening. My good friend Valentina prepared for us the most delicious bolognese ever. The strict recipe comes straight from her Italian grandmother — so we cannot get much more authentic than that. Read &#x26; Watched  Pinocchio (2022)  Guillermo del Toro's Pinocchio is one of the best animated movies I have watched recently. The stop-motion work is super lively, beautifully made, and close to perfection  3  . The storyline is also much darker than the Disney version, which gives it more depth overall. The Queen's Gambit (2020)  Yes, I'm two years late to the party. Robin watched this miniseries when it came out during the first COVID-19 lockdown. I only found time recently to start  The Queen's Gambit  on my own, and was able to finish it in the 12-hour train ride. And what a ride! I am not sure if it deserves the level of hype it got, but the series definitely gripped me. T(W)IL  To bake a tarte tatin  I tried to bake my first apple tarte tatin this week. We had a bunch of apples and butter we had to finish before they went bad, so it felt like the most fitting recipe. There were a few challenges on the way. First, the puff pastry. I followed  this recipe , and I have to say that it worked quite well! The result didn't have as many layers as a store-bought dough, but it definitely did the job! Second, the caramelized apples. I struggled to make a caramel base that would not burn, but would still hold together when poaching the apples in it. Mine became too liquid, and I removed some of it before baking the pie. My concern was that it would be completely liquid when flipping the pie at the end of the process. The final tarte tatin was still delicious, but the apples were not as beautifully caramelized as I would have wanted them to be. For the next try, I will try to use apples that have less water content  4  , as well as make a dry caramel first before adding the melted butter. {end-key}  Footnotes    We are at a bouldering gym with friends, but I am forced to stay on the bench as I sprained my ankle two weeks ago.  ↩   All in all, I wrote a new post, added a recipe to my website and watched four episodes of  The Queen's Gambit .  ↩   Del Toro apparently asked his animators to keep the mistakes in the final rendition to make it look more like stop-motion!  ↩   I might actually even let them dry for a night after cutting them.  ↩   "},{ id: 12, directory: "Musings", title: "2022 doodling", subtitle: "Playing around with my graphics tablet", slug: "/posts/2022-doodling", dateCreated: "2022-12-29",  content: " In October, I got a super cool graphics tablet. I have been doodling on and off since then — and I should definitely try to create a habit around it. In any case, here are already a few experimental drawings.               {end-key}"},{ id: 13, directory: "Musings", title: "#21 Berlin post-Christmas", subtitle: "Update #21", slug: "/posts/21-berlin-post-christmas", dateCreated: "2022-12-27",  content: "    Margot 1980–2012 Reykavic by Irina Werning    Life peek  After about ten days spent with family and friends in Switzerland, we took the train back to Berlin on December 25th. On the ride back, I continued restructuring the way  my recipes  are coded  1  , as well as watched two heist movies with Robin. I don't visit people I love in Switzerland as often as I would like to. Days always feel full and intense, as visits and appointments are scheduled one day after another. Yet, I don't miss Lausanne or Geneva. When not meeting with friends or family, I usually get bored in these cities I used to live in. Anyway, we are having one more laid back Christmas party today with close friends from Germany, the US and Brazil. Let's see how this international party deals with the traditional Swiss cheese fondue! I'll personally allow for non-traditional toppings, spices and sides, but let's see what Robin's views are on such blasphemy. At least, it'll be fun to watch. Read &#x26; Watched  Army of Thieves (2021)  Matthias Schweighöfer directed, produced and had the leading role in  Army of Thieves . We watched it on our 12-hour ride back from Switzerland. The plot was unoriginal, but entertaining nonetheless. Schweighöfer's character and performance is pretty amazing, he made us laugh and cry. From my perspective, the movie also gets bonus points for showing locations in Berlin and other parts of Germany. I felt right at home! Army of the Dead (2021)  After watching  Army of Thieves , we discovered that it actually had a sequel that came out the same year! We still had 3 hours left before arriving in Berlin… so we continued our streak and watched  Army of the Dead  as well. The prequel was more focused on the heist and comedy part of the show, whereas the sequel was much more zombie-action-packed, as well as serious overall. Discovered   Back to the Future   Irina Werning, based in Argentina, looks at people old photographs and helps them reenact these past scenes. In her own words:  I love old photos. I admit being a nosey photographer. As soon as I step into someone else’s house, I start sniffing for them. Most of us are fascinated by their retro look but to me, it's imagining how people would feel and look like if they were to reenact them today… In 2011, I decided to actually do this. So, with my camera, I started inviting people to go back to their future.  Robin and I stumbled upon the book documenting this project in a coffee shop  2  . The before / after pictures are funny, shocking and endearing. We will definitely try to do the same with some pictures of our own! T(W)IL  To play chess again  Yesterday, we went to one of our favorite coffee shops in Berlin. We usually go there to work as they have a very tolerant laptop policy. They also have a massive wooden chessboard. Even though Robin is a big chess fan, I am quite reluctant to play with him. On one hand because I don't like to lose  3  , and on the other hand because I just feel lost most of the time. For once, I agreed to play a game with him (and did feel a bit crappy losing badly at the end). However, as we went back home and played some more on  lichess.org , I started to enjoy the puzzles, post-game analysis and trainings available on the platform. Losing hard still leave me quite sore. But seeing it more like a competition with my own abilities (a.k.a. my chess rating), rather than a competitive game against another player, makes me feel much better and excited about the whole thing. {end-key}  Footnotes    I used to directly code the HTML of each recipe. This worked quite well when my recipe layouts were more straightforward. The HTML template I am using is becoming more complex, as it includes  microformats  and needs more styling. Basically, I am now dividing my recipes more clearly between the recipes content (e.g. ingredient list, instructions, etc.) and their shared HTML structure.  ↩   Where else?  ↩   One on one competitive games often make me feel crappy if I am losing or even winning too much.  ↩   "},{ id: 14, directory: "Musings", title: "#22 Gently sliding into 2023", subtitle: "Update #22", slug: "/posts/22-gently-sliding-into-2023", dateCreated: "2023-01-02",  content: "    House by the sea by Jessamy Hawke    Life peek  Guten Rutsch ins neue Jahr! Or have a good slide into the new year, as Germans say. I am back to work after one week of relaxed vacation in Berlin. It feels like much longer though, as we had been to in Switzerland the two previous weeks as well. Nothing much happened since my  last weeknote . Robin and I met with friends for some holiday dinners, we started playing chess, we did our Year in Review  1  , we caught up with our current D&#x26;D campaign, we created a photo album for 2022, and I started to do some rehabilitation exercises for my sprained ankle. Writing this last paragraph, I realize that nothing much was actually a lot. It just felt very laid-back and nice. The perfect vibe for the end of year. People around me are talking a lot about New Year's resolutions. For once, I am looking at the coming year with a very open mind. My main goal is to get acclimated with our new life in the French countryside. Of course, I have a  ton of things I'd love to do there . But I see these more like wishes and ideas than resolutions.  27 days left before the move  Read &#x26; Watched  On body and soul (2017)   On body and soul  is the Hungarian movie that won the Golden Bear at Berlin International Film Festival in 2017. It is just the right amount of weird / sad / heartwarming that keeps you watching. Glass onion (2022)  We had a cozy New Year's Eve with friends, eating ravioli and watching  Glass Onion  on Netflix. The movie is longer than what we expected (2h20), and we finished it only 10 minutes before the midnight deadline. I don't think it will live on as the first movie  Knives Out  might. I still laughed and enjoyed it a lot, watching it only two days ago. However, I will probably not recommend it in one year anymore as it already feels slightly aged. Bad times at the El Royale (2018)  After watching  Glass Onion , we were still in the mood for a murder mystery. We decided to give a chance to  Bad times at the El Royale  that came out a few years back. To be honest, it surprised me in the best way possible. I found the plot enticing, the characters interesting and rather well explored. As far as murder mysteries go, I would definitely recommend this movie for an entertaining night! Discovered  Jessamy Hawke  You can see one of Jessamy Hawke's artwork at the start of this post. She is a UK-based illustrator who mainly works with watercolor and gouache. She is also super active  on Instagram  where I first discovered her. She has a way of illustrating landscapes and scenes, using simple shapes and lines with a lot of different textures. There is a roughness to her work that I just love. T(W)IL  To generate an RSS feed for my posts  It is less about me learning to generate a new RSS feed than copying  Robin's setup . I did learn a few things on the way and realized that I had misunderstood the whole GitHub / Cloudflare connection. At the moment, you can find my  RSS feed here . As I use Brut as a static site generator, I have created a  prebuild.js  file that is automatically run  before  the main build in Cloudflare. {end-key}  Footnotes    It's now been three years I write a Year in Review to reflect on the past and plan for the future. I am tweaking its structure year after year to fit my needs. I'll see if I find the time — or more like the motivation — to share it here these next few weeks.  ↩   "},{ id: 15, directory: "Musings", title: "#23 Galette des rois and IndieWeb nuggets", subtitle: "Update #23", slug: "/posts/23-galette-des-rois-and-indieweb-nuggets", dateCreated: "2023-01-09",  content: "   Who needs artwork when a galette des rois looks like this?   Life peek  Once Christmas and New Year are over, there is still one good thing left:  la galette des rois , or the  king cake . We got a mini-one from one of our favorite bakeries, and Robin got the  fève  this time (a cute tiny porcelain cow). As we are getting closer to our moving date, it feels like our days are packed with things to do and think of. Everything is also a lot more emotional. Each encounter with friends or trip to a well-known bar feels like it might be the last one.  20 days left before the move  Read &#x26; Watched  The lost city (2022)  By the Nee brothers So, as it often is, we were in the mood for a light and entertaining movie.  The lost city  was all of that, plus extra dumb as well. It was perfect. It stars Sandra Bullock and Channing Tatum in a romantic comedy-adventure. The movie doesn't take itself seriously and is just what you need after a long day of work. Also, did you know that there are three series and three other movies called exactly  The lost city  as well?  1   Aftersun (2022)  By Charlotte Wells On the other extreme of the spectrum, there is  Aftersun . What a wonderful debut by director Charlotte Wells! There is so much to love in this movie. To begin with, the cast is incredible. The father-daughter relationship is beautifully depicted, the emotions are hinted at, without being shoved at you. The use of vintage camera imagery is brilliant. This is a movie you might want to rewatch to uncover all the details Wells left there for us. Discovered  I am not a hundred-pinned-tabs kind of person. I don't really know why, it just doesn't happen for me.  2   These last few days, I consciously pinned websites I had stumbled on. If I like the process, this might be one of my new ways to nurture this  Discovered  section.  Blogroll by Ray   On Ray's website I love how old websites used to have a blogroll page. For someone who had no digital life at that time, it feels like it might have been a wonderful way to browse the web. This blogroll page is just that. Ray's curated bookmarked links, shared with the world.  Cool things people do with their blogs   A Brain Baking's article I am not really sure how I stumbled upon this article anymore. I didn't know about  Brain Baking  at all (and still haven't had time to explore it further), but I saw this specific article and really love the idea of it. Once again, it is a list, but a list of cool things people do with their personal blogs / websites.  Devils   A Neocities personal website Talking about personal websites, I just love how different and original this one is! Actually, when you dive a bit more into the websites generated with  Neocities , you'll realize most of them are super special and personal. Let's get away from marketing design and experiences when creating our own spaces on the internet!  Plot generator   On Big Huge Thesaurus I was searching for a synonym and I discovered this website... 5 minutes ago. And what a delight to discover that they have a plot generator! The page basically generates new story ideas every time you reload it.  A deranged pharmacist flies home. A naval officer comes to town. A senator meets a strange homeless woman in Australia. A cursed child must escape from an evil wizard in a lifeless desert. A sick child is kidnapped. Rival baseball coaches take a wrong turn after they die.  {end-key}  Footnotes     Wikipedia source   ↩   Maybe I get stressed out when having too many pinned tabs?  ↩   "},{ id: 16, directory: "Musings", title: "#24 Playing with games and typography", subtitle: "Update #24", slug: "/posts/24-playing-with-games-and-typography", dateCreated: "2023-01-19",  content: "    Berlin, 2014 by Magnus Hagdorn    Life peek  On Sunday, I went bouldering again after six weeks of not moving much.  I had hurt my ankle  on one of our bouldering-pastries morning days. My ankle range of movement, strength, and balance are still limited, but I try to do some targeted exercises to help it heal faster  1  . Back to the bouldering session, I was very careful to not do any move I was not completely in control of. It is a bit frustrating to be climbing super easy routes only, but feel like my body is not ready for more. Even though we still have a lot to do before our move, our days are packed with social events and activities. It feels like we only have a few days left to take with us as much of Berlin as we can.  Ten days left before the move  Read &#x26; Watched  Mommy (2014)  By Xavier Dolan This movie was — and was not — on my watch list for some time. Simply by looking at the poster, you know that the  Mommy  is going to be disturbing on some level. I finally watched it last week, and indeed, it was disturbing. However, I wasn't expecting it to also feel touching and wholesome. It made me feel a lot and I would definitely recommend it! John Wick (2014)  By Chad Stahelski Aaah, the John Wick franchise! I am not sure how I managed to never watch such a big blockbuster. I was in need of some light entertainment. So, Robin and I watched the first three movies last week. It was everything I thought it would be. Action-packed, not very smart, very entertaining. Plus, it stars some puppies and a network of assassins who all know each other's on a first name basis. Discovered   Return to Monkey Island   A point-and-click adventure game Most of the things I discover or get into are usually introduced to me by Robin. This game is no exception to this rule. He played the first two pixelated games that came out in 1990 and 1991. More than thirty years later, the original creator, Ron Gilbert is teaming up with designers to create a gorgeous new addition to the  Monkey Island series . This point-and-click adventure is full of quirky dialogues and fun puzzles to solve. It had been a while I hadn't enjoyed a game that much!  Keep talking and nobody explodes   A party game Yesterday, we met with some friends for dinner (pizza) and fun activities. We discovered  Keep talking and nobody explodes . In this collaborative party game, you need to solve puzzles to deactivate a bomb. The only problem is that you control the bomb, your friends have the instructions to defuse it, and you cannot see each other's parts of the game. In short, they'll have to talk you through the instructions fast if you don't want the bomb to explode! It is a super fun game, but very intense and stress-inducing as well. I don't think I could play it for too long or with too many people at once.  Chris Burnell's styleguide   A personal website styleguide I stumbled on Chris Burnell's website while browsing the IndieWeb ring (once again). And one thing I love from it is his personal website styleguide. I really like the concept and will definitely try to make one for my website!  Utopia   A responsive CSS approach I discovered Utopia through a link on Chris Burnell's styleguide. I haven't had time to dive into it properly, but I already like what I saw so far. Through CSS variables and calculations, they created a way to design responsive websites without the use of media-queries. I like how systematic and simple the concept is — the calculations are hard on the eye though. Once reviewed properly, I will see if I want to adapt my website to this kind of system. T(W)IL  To make a font  As explained in details  in this post , yes, I made my first font last week! {end-key}  Footnotes    I now feel like an ankle sprain expert after reading so much about it.  ↩   "},{ id: 17, directory: "Musings", title: "#25 Unpacking and wood stove fires", subtitle: "Update #25", slug: "/posts/25-unpacking-and-wood-stove-fires", dateCreated: "2023-02-04",  content: "    Campfire by Gillian Levine    Life peek  Robin and I are finally in Ardèche. We were welcomed on Monday afternoon with a beautiful weather: sunny, cold, and crisp. It was actually colder than expected, as the house needs to be manually heated with a wood stove. A few days in, we are getting the hang of it. We start the day by lighting a fire in the stove and keep it alive until the temperature is warm enough inside. We spend our days tending to our fire, going to local businesses, visiting neighbors, unpacking our boxes, cleaning, cooking, reading by the fireplace, eating, drinking  1  , creating a space we can call our own. I am still waiting to see how I'll like it eventually, but so far, I love this new pace of life. Read &#x26; Watched  I don't want to spend too much time on my laptop this week and will keep this section short and sweet. Eo (2022)  By Jerzy Skolimowski A Polish movie from the point of view of a donkey. I liked it better than  Au Hasard Balthazar  (1966) even though I didn't love it either. The Umbrellas of Cherbourg (1964)  By Jacques Demy A French classic that is surprisingly better than what I was expecting. And look at these wallpapers! Return to Seoul (2022)  By Davy Chou Pretty impressive to know that the main actress — Ji-Min Park — had never acted before. Not my favorite movie of the year, but it dives into the adoption topic that we don't see much on screen. Discovered   Cain's Jawbone   A murder mystery puzzle A month ago,  I watched  Glass Onion  . After watching a movie, I usually enjoy reading articles or watching videos analyzing it. In one of these explanation videos, I discovered that Benoit Blanc's introduction scene (in the bathroom) shows that he owns  Cain's Jawbone . It is an infamous murder mystery puzzle printed as a book, first published in 1930. Wikipedia will explain the concept better than I could:  The puzzle consists of a 100-page prose narrative with its pages arranged in the wrong order. [...] To solve the puzzle, the reader must determine the correct order of the pages and also the names of the murderers and victims within the story. The story's text includes a large number of quotations, references, puns, Spoonerisms and other word games. The pages can be arranged in 9.33×10157 (factorial of 100) possible combinations, but there is only one correct order. The solution to the puzzle has never been made public.  Needless to say, I really want this book.  Accessible palette   I haven't had time to find a practical use for this tool yet. But well, I am a sucker for rainbow colors on my screen. T(W)IL  That you can use ashes to clean stuff  As explained above, we need to heat our main living room with a stove. This means that we produce a lot of ashes. Robin's grandma — France — usually uses it for the garden. However, we now have too much of it and need to find other uses. Well, lucky us, ashes can be used — a bit like baking soda — to do practically anything! Some of these uses include: cleaning anything kitchen-related (an oven, a pot, a kitchen plate), making washing liquid, repelling insects.  2   {end-key}  Footnotes    ...drinking wine produced in the vineyards next door. Can you believe it?  ↩    Source (in French)   ↩   "},{ id: 18, directory: "Musings", title: "#26 Taking the time", subtitle: "Update #26", slug: "/posts/26-taking-the-time", dateCreated: "2023-02-26",  content: "    Island Houses by Jessica Smith    Life peek  It's now been almost a month Robin and I have moved to our new home. We are slowly building new routines again. Unsurprisingly, they tend to be quite different from what we had in the city. Everything takes more time, which makes me much more intentional about the things I choose to do. Waking up in the morning, I first need to start the wood stove fire. If I'm lucky, it only takes 5 minutes. However, the fire sometimes needs more baby-sitting before properly burning. The weekly market closest to our place is a 15-min drive away. As it is close to other stores, we usually plan in advance if we need something else and get it all at once. Meeting new people is different as well. Everyone is very welcoming and makes me feel like I belong very quickly. It will however probably take time for us to build meaningful relationships. Surprise, surprise: we started bouldering again by joining a local club. It was a bit of an admin hassle to integrate it, but now that we are in, we couldn't be happier! The members are super friendly and passionate about climbing. I'm looking forward to the outdoor events that'll come with the sunny days. One last big change is that I have been spending my free time much more offline. It is quite invigorating to step away from the online world a bit. On that note, I'll keep this weeknote short and stop it here. {end-key}"},{ id: 19, directory: "Musings", title: "#27 (Real) rock climbing and more", subtitle: "Update #27", slug: "/posts/27-real-rock-climbing-and-more", dateCreated: "2023-03-06",  content: "    Yerevan night life at Kong Cocktail Bar by Sarah Mari Shaboyan    Life peek  What a week! Lots of new discoveries these last few days. For the first time, I climbed my first crag with Robin and a few people we met at our local bouldering club. So far, we had only been climbing indoors — what a change! I do feel a lot more queasy than indoors, afraid to fall. I am also climbing a lot less efficiently as I spend precious minutes trying out tons of different holds before settling on one. On the opposite side, it is really liberating to be able to reach for any hold you want and decide to use it. The weather was also perfect. We arrived around 1PM and immediately removed jackets and sweaters. It was so sunny that it felt like the start of summer. I also had my first car driving experience. I never learned to drive, but the region really demands it. So far, I just learned to start the car, and I finally realize how much coordination it requires! I'll try to get through my car driving lessons as fast as possible — let's see if I can manage to drive a car by the end of the year. Read &#x26; Watched  The second week after our arrival in France, we subscribed to the big library in Aubenas. I love French-speaking libraries, as Belgium and France have a strong comics / graphic novels culture. We used to live next to a library, where we would go every two days to refresh our graphic novel storage. Our current library being 30 minutes from home, we now go every two weeks only. Nevertheless, I am happy to come back to this forgotten routine.  La maison de la plage (2019)   By Séverine Vidal and Victor L. Pinel  La maison de la plage  —  The house on the beach  — is a beautiful graphic novel by Séverine Vidal (scenario) and Victor L. Pinel (drawings). It is really well structured and made me travel through time, with an old family house at its center. Pinel masterly uses lighting and colors in his drawings to create distinct atmospheres and give us a sense of time. At the end of the day, this graphic novel made me feel a lot of different emotions and I would highly recommend it.  Natsuko no Sake (1988)   By Akira Oze  Natsuko no Sake  is my first manga in a while! It is a real love letter to the Japanese tradition of brewing sake. So far, I am only two volumes in — there are six volumes in total. I quite like it, but do feel like some of the content has aged a bit — nice characters being very sexist, saying inapproppriate things, and so on. I am also not completely taken by the main character whom I find a bit annoying. But the good part is that you learn a lot about sake and the traditions that surround it. I would recommend this series to people interested in learning more about sake in an entertaining way.  Dans la combi de Thomas Pesquet (2017)   By Marion Montaigne In the same vibe as her other scientific vulgarization graphic novels, Montaigne does a great job explaining the astronaut life in  Dans la combi de Thomas Pesquet . We quickly realize that being an astronaut is not as glamourous as we may think. Reading this book makes you feel like you are visiting the secret NASA backstage — worth it! {end-key}"},{ id: 20, directory: "Musings", title: "#28 On learning", subtitle: "Update #28", slug: "/posts/28-on-learning", dateCreated: "2023-03-24",  content: " Since we arrived in Ardèche, I find it hard to keep up with this weeknote. I think that one of the reasons is that I invest more time on  Eau de poisson  — the blog I share with Robin. I still like having this space that is only mine. However, it will have to adapt to these new needs, to this new rhythm. Let's see how this evolves in the future months.  It is now 10PM, we are just back from the local astronomy club meeting. The members average age must be about 50-year-old. (Robin and I are definitely the youngest there.) The club was founded in 1993 — practically my age. Members are quite active, but only few of them really know the sky. The most knowledgeable members were suggesting giving classes on how to manipulate the club telescopes and material. This really motivated me to get started with proactively learning the sky. I can already identify a few constellations, but would love to know more about their origin stories, the planets and stars, the physics rules that connect them all.  Talking about learning new things, I feel really inspired at the moment to  learn more . 3D modeling on Blender, the night sky, edible wild plants, gardening, driving. The motivation is there... but for how long? I know that these quick passions tend to die out. I might need to focus my energy on a few selected topics if I want to do some real progress. As they say: quality over quantity. I'll sleep on that. {end-key}"},{ id: 21, directory: "Musings", title: "#29 Time for a break", subtitle: "Update #29", slug: "/posts/29-time-for-a-break", dateCreated: "2023-04-04",  content: " It's official: I am leaving my job of almost four years by the end of May. Many feelings are battling inside my head — excitement, anxiety (a lot of it), wonder, a bit of nostalgia. I don't have clear-cut plans yet, but I would like these next few months to be free experimentation grounds. I want to learn new things — digital and not, to get dirty with my hands, to spend more time outdoors, to improve at crag climbing, to create more things from scratch, to preserve food, to kick-start personal projects  1  . And by the start of 2024, the grand evil plan is to go on a long traveling journey again. Robin and I have loved making our (slow) way through Southeast Asia in 2018. The still-to-be-defined idea would be to do something similar — a big shock and dive into (an)other culture(s) for a sustained amount of time. This week I learned  ...to scrape a simple web page with Node.js.  Using a library, but that's fine, right?  I used the  Cheerio library  and it was surprisingly easy to implement. All of that was to create my new  Movies pages . I first wanted to simply use the Letterboxd API, but they don't seem to want to open its access yet. ...to lead climb on a crag! What an exhilarating experience. I already knew the indoors lead climbing theory, but had never tried doing it outdoors — there are a few more technics to know in order not to die. Robin and I went to Mazet with our local climbing club. In the end, I climbed twice the  same 6A+–6B  with a break on the way. The next step is now to overcome my fear of falling... {end-key}  Footnotes    and completing my driving license (fingers crossed)  ↩   "},{ id: 22, directory: "Musings", title: "#3 Friends and digital gardening", subtitle: "Update #3", slug: "/posts/3-friends-and-digital-gardening", dateCreated: "2020-11-24",  content: "    Among us  by Pooria LFA   What a week! I feel like so many things happened since last week. Heart-to-heart interview  First of all, I finally had my interview with three people from my current company. For the first time in my life, I have answered to these interview questions with 100% honesty. It felt quite liberating to be completely myself and not give half-truths that are more aligned with people expectations. An app to feel closer to your friends  Robin introduced our small group of friends to the  Cappuccino App . How does it work? Every day, you have the possibility to record one  bean  that will be mixed with the beans from your friends in a podcast-like format — with background music and everything! You can also give prompts or ask questions if you don't like going on about your day. I love it. This app arrived right on time, as you currently don't have the possibility to meet your friends anymore. I love the fact that there is no possible small talk. No one is in front of you. You get straight to what is important. I feel like I have learned so much about my friends in just a few cappuccini. It is also true the other way around. I am usually quite reserved, but feel like talking about myself much more within these short 3 minutes. Reset of my digital garden  Middle of the week, I did it. I completely resetted my digital garden. I had first started it without much of a system. It rapidly got out of hands and began to feel chaotic and a bit overwhelming — far from the thinking-buddy-second-brain-thingy I had imagined. I started it again following  The Zettelkasten Method: Examples to help you get started  by Rebecca Williams to the letter. It felt great to work on a blank slate with a clear system in mind, knowing where I was headed. {end-key}"},{ id: 23, directory: "Musings", title: "#30 Slow start", subtitle: "Update #30", slug: "/posts/30-slow-start", dateCreated: "2023-04-20",  content: " After a week of holiday with family coming over to visit, I am back to my routine. Getting back to work felt intense. The things I am currently working on are quite interesting, but I felt exhausted at the end of each day. I'm not complaining though, I only have 3-day-work weeks from now on until the end of my contract. 3-day-work weeks means that I have a lot of free time. I have many DIY projects I want to pursue  1  , but often feel blocked by one thing or another. For instance, finding the material is a big hurdle. It might be because I am trying to upcycle too much instead of just buying new raw material. Another hurdle is that the woodworking workshop is  right next to  the neighbors' house  2  . One of their doors actually connects to the workshop. I find myself tiptoeing and trying to make as little noise as possible whenever I go there. This makes it hard to use that room at all. I might need to be more patient and simply progress one step at a time, wherever I have control. These shelves, clothes hangers and curtain holders are not that urgent to build. Experiments  Making cereal bars  As the weather is getting better, we are doing more and more sporty outdoorsy activities. Whenever we have cereal bars at home, I usually bring them as snacks. For the first time last week, I tried making my own cereal bars. So far, I am not 100% convinced by the recipe I tried — they were a bit too soft and didn't contain enough honey to my taste. However, I am pretty sure I can get it right after a few tries. I don't know why I didn't think of making these cereal bars earlier, as the ingredients are so simple  3  . Gardening  Robin did a lot of gardening work these last few weeks —  and I helped him here and there. I can already see that he is going to be the gardening-person out of us two. I do like it, but prefer the cooking-and-eating-the-veggies part. So far, we have a few strawberry and arugula plants on our terrace, six zucchini plants in the lower garden, 12 salads in the greenhouse. There are some more plants in France's garden that we are helping with as well. {end-key}  Footnotes    You should see my to-do list.  ↩   It's actually another couple that France kindly took in, a bit like us.  ↩   Basically, you just need to use granola ingredients.  ↩   "},{ id: 24, directory: "Musings", title: "#31 Friends, sci-fi and Coppola", subtitle: "Update #31", slug: "/posts/31-friends-sci-fi-and-coppola", dateCreated: "2023-05-14",  content: " These past two weeks, Jillian and Connor came to visit us in Joux! We have now been in the area long enough to have developed a few routines. It was great being able to share some parts of this new life with friends from Berlin. Nothing fancy, but heartwarming nonetheless: a lot of gardening, reading, outdoor climbing, village hopping, wine drinking, cheese, sandwiches and cheese-sandwiches eating. I am now writing this weeknote as we are all on a train  1   from Valence to Berlin. This is the start of 4 weeks away from Ardèche, our first time since we moved in. We are currently sitting across a four-year-old kid and his mother, and I just don't know how she does it. The level of patience and kindness she uses to deal with the little monster is insane.  2   Parenting seems like an exhausting full-time job, not sure whether I could do it. Read &#x26; Watched  A small streak of Becky Chambers  After one lunch discussion about sci-fi, I suddenly wanted to dive into something sci-fi-ish again. So I did. I went on and read two novellas by Becky Chambers back-to-back — and I fell in love with her writing. The first novella is called  To Be Taught, If Fortunate  (2019) and is set in the 22nd century. Space exploration has been revolutionized as astronauts bodies can now adapt to their environments through a new biotechnology. The second novella,  A Psalm for the Wild-Built  (2021), is drastically different. It seems to be set in a far future, where humans left fifty percent of the world untouched and protected from human intervention. The singularity is also a thing of the past, robots have gained self-awareness and have now wandered off into the wilderness. The main characters of each novella could not be more different: the first one is an astronaut and the second one a monk. However, they also share a strong common trait: wherever they go, there is a will to leave a mark as small and temporary as possible, be it on another planet or just on a stretch of Earth soil. Sicilian night  As a taster of our Sicily trip in one week, we watched  The Godfather  directed by Francis Ford Coppola. It is a three-hour-long movie from 1972, and  wow , it does deserve its classic label. I — surprisingly — really enjoyed the movie and didn't get bored at all.  Consiglio questo film!  Experiments  Dev  Robin and I have been working some more on our little side-project. We are creating our own private Pocket-like tool  3  . So far, we had been using  Hypothesis  a lot to annotate web pages. The highlighting and annotation experience is fairly good, but the tool is just not designed to bookmark future articles to-read. Other tools also exist, but they are either too expensive, not standardized enough  4  , or have a shitty reading / annotating experience. In the end, those  might be  are certainly excuses we use to just dive into yet another side-project. Gathering flowers  May is flower-month. In other words: time for thyme! When Jillian and Connor just arrived, we gathered a lot of flowering thyme around Balazuc. We let it dry for two weeks and spent approximately 1 million hours yesterday separating the flowers and seeds from the branches. It is a truly meditative experience, but be prepared to (1) use up the skin of your fingers, and (2) smell wonderful afterward. We also gathered some elderflowers from the garden. I tried to make an elderflower-pear pudding following  this recipe . It came out quite good, but almost too sweet and thick to my taste. Robin made some elderflower syrup that we are preciously keeping in the fridge until we come back from our month away. T(W)IL  I started to follow the  Typescript course by the codecademy . I feel like I have reached a comfortable level in JavaScript — at least, comfortable for what I usually need to do with it. Hopefully, this course will help me feel a little bit less lost when needing to read some Typescript files. I'm also entertaining the idea of migrating my website to Typescript for practice. (Well, now that I've written that sentence, I will do it for certain.) {end-key}  Footnotes    or more accurately, three train connections  ↩   I wonder if I were as insufferable as a kid and just don't remember it. Chances are that, on some level, yes.  ↩    Pocket  is a tool by Mozilla that you can use to bookmark web pages, articles, blog posts to read later.  ↩   What if we want to export our data at some point?  ↩   "},{ id: 25, directory: "Musings", title: "#32 On the train to Sicily", subtitle: "Update #32", slug: "/posts/32-on-the-train-to-sicily", dateCreated: "2023-05-21",  content: " Just like the last weeknote, I am writing from a long train ride. My days and mind are so busy lately, that it seems to be the only extended length of time I have with nothing going on.  Nothing  is actually not very accurate, as these long travels always come with a decent amount of stress, fatigue and attention-grabbing things. Anyway, Robin and I just left Berlin yesterday after spending a little under one week at Jillian and Connor's place. Our first night there was both very familiar and completely foreign. We have surprisingly gotten used to our new countryside pace in just a few months. Our short Berlin trip was also synonymous with our last week of work. By quitting, I know — or anxiously wish(?) — that I am making the right decision, but I still feel sad to leave this team that shared my daily life for the last few years.  Oh wait, the train is getting on the ferry(!) to Sicily. I'll go have a walk on the deck and come back to these thoughts.  Wow, we actually just learned that the Etna just erupted less than an hour ago. We tried to see the smoke, but are not sure we could distinguish it among the other clouds. There is such a special atmosphere in the air: the grey cloudy sky, the dark opaque and ink-like sea, the fresh breeze. It feels a bit like the ominous start of an apocalyptic movie. Hopefully, we'll still be able to eat some good Sicilian food in that post-apocalytic story. {end-key}"},{ id: 26, directory: "Musings", title: "#33 Post-Sicily", subtitle: "Update #33", slug: "/posts/33-post-sicily", dateCreated: "2023-06-05",  content: " Our first trip to Sicily came to an end yesterday, as we took a 13-hour ferry from Palermo (Sicily) to Cagliari (Sardinia). The two last weeks in Sicily were  full of discoveries . We are bringing back home some new recipes —  caponata, pasta alla Trapanese, panelle, pistacchio pesto  — and also many bottles of local specialties  1  . Another thing I am bringing back home is inspiration from people I've met and things I've seen on the way:  Isabel and Baptiste, singing and dancing their hearts out during our roadtrip The rich sea flora / fauna that can be observed with a simple snorkeling mask Sicilians love for good produces Davide and his wife starting to make their own wine, honey and olive oil Those winemakers offering us a bottle of wine out of nowhere That little creative guide of Palermo, randomly found in a wine bar Georgia and Davide moving to the heart of Sicily to open their B&#x26;B for climbers Old ladies feeding stray cats in the street Alice, traveling alone through Italy with only a backpack and camping gears The close-knit community in Bosa (Sardinia)  Even though we will only be home in a few days  2  , it already feels like the end of a vacation. I am both anxious and excited to be back in our little Ardèche corner. I am looking forward to seeing friends, climbing, cooking, and gardening again. I am however a bit anxious about being jobless  for real  and having days full of  nothing planned  for the first time in years. In the end, I guess it all comes down to taking responsibility for how I choose to spend my days and how I live my life.  3   {end-key}  Footnotes    alcoholic ones, that is  ↩   We are crossing Sardinia to get a ferry to France in three days.  ↩   Ugh, I definitely need to learn to take the pressure off.  ↩   "},{ id: 27, directory: "Musings", title: "#34 Mushroom hunting and playing with words", subtitle: "Update #34", slug: "/posts/34-mushroom-hunting-and-playing-with-words", dateCreated: "2023-06-19",  content: " I haven't been super consistent with these updates for the past few weeks. One specific highlight since we got home: we went mushroom hunting for the first time with Cricri! When we came back, it had been raining for a week straight, so the chanterelles and ceps were just  waiting for us  to be gathered. Robin already wrote extensively about  our mushrooming adventures , so I won't repeat it all here. We also gathered a few of the vegetables from our garden — so green, so lush! It feels like every meal now contains a bit of our work in it: super satisfying. Read &#x26; Watched    The Road (2006) , a novel by Cormac McCarthy — Visiting the classics with this short postapocalyptic novel, that is focussed on the relationship between a boy and his father in this collapsing world.  Dungeons &#x26; Dragons: Honor Among Thieves (2023) , a movie directed by Jonathan Goldstein and John Francis Daley — Loved watching this after having played a few games!  Howl's Moving Castle (2004) , a movie directed by Hayao Miyazaki — Rewatching it after years, I had forgotten how it was full of poetry and humour. It made me want to read the original novel soon!  Decision to leave (2022) , a movie directed by Park Chan-wook — Much more subtle than some of his previous movies. Striking photography, plot focussed on a forbidden love story (as usual).  Experiments  Swedish painting  We wanted to repaint the main door to our house, and discovered  this Swedish painting technique  through France. The paint is made by mixing flour, color pigments, iron oxides, linseed oil, and a bit of black soap. It can then be used as a regular outdoor paint, as it protects our wooden door against rain and deterioration. Getting creative with words  For a few years now, I have been writing on and off in my numerous diaries, travel journals, blogs, and so on. All of these projects are very down-to-earth, as I usually talk about lived experiences, current thoughts or future plans. I have been trying to bring back some creativity in my life in different ways. One of them is playing more with words and poetry. I published  my first try at it in a previous post (in French) . I also tried to do some  caviardage  exercises, that are quite fun and relaxing!  1   {end-key}  Footnotes    In French,  caviarder  means blacking out lines or sentences. You can do it creatively from a source text to create a poem, that has a completely different meaning.  ↩   "},{ id: 28, directory: "Musings", title: "#35 From one project to the next", subtitle: "Update #35", slug: "/posts/35-from-one-project-to-the-next", dateCreated: "2023-07-12",  content: " Apparently, the fewer obligations I have in life, the less I keep this blog updated. One reason might be that I find it hard to focus. As I don't have any real obligations, my days are spread out across multiple projects. I find myself having days filled with varied snippets of many little things. It feels great to have the freedom of waking up in the morning and be like nah, not today. However, I do miss the continuity of working on one big project consistently. Until I manage to focus my brain, here is a list of the small ongoing projects I have at the moment.  learning about plants and creating a herbarium home improvements (starting with building a bookshelf) getting better at JavaScript (usually because I am working on yet another unnecessary feature for my website  1  ) gardening with France learning ways to preserve food climbing, climbing, climbing working on a private Pocket alternative with Robin planning our 2024 big trip  Special climbing note  Since we've arrived in Ardèche, we had a number of climbing first times. First, we had to learn to climb a real mountain wall and get used to how varied rock types will feel different. Being used to bouldering only, our stamina was also challenged. After feeling comfortable climbing top rope, we practiced lead climbing again. And most importantly: lead falling. (I am still in the middle of properly learning that one.) Last week, we tried our first multi-pitch climbing ( Alien vs Predators , 50m). And this week, we directly went on with our second multi-pitch ( Biotone , 110m). Multi-pitch climbing adds a thick layer of new techniques to regular sport climbing. Suddenly, you have to be extremely organized with how you manage your rope and tools to make it to the top (and back to the bottom). It is not that much about performance climbing anymore, but more about planning and technicalities. I love how rich the world of climbing is and how far you can potentially go, if you are motivated enough. Let's see how far we'll dare to go. {end-key}  Footnotes    Look at this  new sexy media page   ↩   "},{ id: 29, directory: "Musings", title: "#36 Keeping up with the fruits and veggies", subtitle: "Update #36", slug: "/posts/36-keeping-up-with-the-fruits-and-veggies", dateCreated: "2023-08-08",  content: " So, it's been a month since the last  weeknote . I should rename these weeknotes into something like updates to feel less guilty anytime I skip writing these short entries.  1   Well, you know what, it's done. Now, let's get started with the actual update.  Learning about fermentation  As you might have seen, I dove head first into the theory of  fermentation . What a fascinating and magical world! I think I need to read a bit more about it to have a more structured understanding of it. But I already feel like I have learned a lot and built a good foundation from just that Wikipedia browsing afternoon. To be continued... Building a bookshelf  Robin's parents came to visit a few weeks ago. We took advantage of Robin's father woodworking skills to build our bookshelf together. I love how it looks and am so happy to finally be able to store our books in a proper bookshelf  2  ! Learning Russian  So far, I managed to learn the Cyrillic alphabet. Robin and I took our first Russian lesson on italki with a Russian dude who speaks French fluently  3  . Our goal would be to have a basic understanding of Russian while traveling across Central Asia next year. Keeping up with all these fruits and veggies  Summer is amazing for fruits and veggies here. We almost don't buy any extra vegetables at the market nowadays, as our garden is providing us with new zucchini, eggplants, tomatoes, green beans (and more) every few days. The only problem is that we tend to have a lot of a few produces — yep congrats, that's how seasonal veggies work, Clara. This really challenges my creativity, as I need to find diverse ways to eat the same vegetables every day. The same is true with fruits, as we have a lot of plums from the garden. We also buy some extra peaches, apricots, and melons from local farmers. The duration of a fruit season is usually quite limited, so it feels a bit like now or never. I love this abundance. Having too much of these fruits and veggies lead me to discover  zucchini soup ,  peach ice tea , zucchini terrine,  tomato and zucchini tarte tatin ,  gaspacho ,  fruit clafoutis , and more.  Otherwise, I'm thinking of redesigning my website (once again). I am getting bored of this minimalist black and white style. And — let's be honest — I just want to redesign it because it's fun. Until next time! {end-key}  Footnotes    And honestly, who cares beside me anyway?  ↩   They hadn't moved from their cardboard boxes since our arrival in February — shame on us.  ↩   I am so impressed with these people who fluently speak multiple languages. I am usually not too bad in the classroom, but cannot remember any of these new language skills long term.  ↩   "},{ id: 30, directory: "Musings", title: "#37 Bà ngoại", subtitle: "Update #37", slug: "/posts/37-ba-ngoai", dateCreated: "2023-09-20",  content: " I went back to Switzerland a few weeks ago. For these trips, I usually stay at my mother's place. This time, I stayed with my grandma ( bà ngoại   1   as I call her). I can't remember the last time I spent so many days with my bà ngoại. It's crazy to think that it might run back to when I was still a toddler, wrecking havoc in her house. Growing up, she was my hero. She had faced many hardships: extreme poverty, war, migration. But anytime she tells me the stories from her past, they are surrounded by an aura of positivity. She says luck has always been on her side. But what I believe is that she was the one to sow these seeds of luck. Growing up, I felt like my bà ngoại was invincible. During that trip in Switzerland, it hit me how fragile and old she had become. As she used to be the one to carry me, I am now helping her to walk around. As she was the one tending to my childish scrapes and wounds, I am now taking her mind off her constant physical pains. She lived her life to the fullest, grateful for all the silver linings. Her body is now failing her, but her mind is still as sharp, her mindset content. I now realize that she was not invincible in body, but in her way of seeing life — fighting hard for her ideals, yet savoring every drop of what she had. {end-key}  Footnotes    Pronounced as /ɓaː ŋwaːj/  ↩   "},{ id: 31, directory: "Musings", title: "#38 Hiking, eating, harvesting", subtitle: "Update #38", slug: "/posts/38-hiking-eating-harvesting", dateCreated: "2023-09-27",  content: " A lot has happened in the last two weeks. To begin with, Robin's otitis evolved into a semi-facial paralysis. That was super scary. Thankfully, we now know for sure that the condition is benign and should pass with time. With that diagnosis in the background, we tried to go on with life and our regular activities. Activity #1: hiking  On Sunday, we went to St-Melany for the first time, driving on some of the most beautiful roads I have seen in Ardèche. St-Melany also has an artist-themed hike that guides you through amazing landscapes. The only drawback is that the entire loop is 5–6 hours long. That hike is therefore the longest I have ever done (yes, I'm weak). I am happy to announce, that although I almost gave up in the middle, I still successfully completed the loop without crying. Plus, a great thing about doing that hike end of September is that we were able to nibble on lots of wild grapes on the way! Activity #2: eating nougat  On Monday, Robin had a doctor appointment in Montélimar, a legit (albeit small) city one hour drive away. So, we thought that it would be nice to spend the day there, doing some city-related activities. The idea was nice. The only problem was that it was a  Monday . And French cities are  dead  on Mondays. In short, all the nice restaurants, coffee shops, museums (nah, kidding), book stores, even libraries we wanted to visit were all closed (sob). But that's  fine , because Montélimar is a wonderful city where everything is under construction. I'll let you imagine how lovely the atmosphere was. [insert construction noises]  Okay , I'm making it sound worse than what it was. In reality, we stumbled upon a cute Vietnamese restaurant — yes, one of the rare restaurants to be open. Bonus points: we also visited two nougat producers and brought back a lot of sweet edible souvenirs. We discovered  black nougat  that is made with mostly honey and as little sugar as possible. It's official, I can now never go back to regular white nougat. Activity #3: harvesting pepper  France planted a Sichuan pepper tree in her garden about ten years ago. This is now the second year the tree gives her enough flowers for a harvest. Yes, flowers and not fruits. When talking about Sichuan pepper, the name is misleading, as it is in fact not closely related to black pepper. Apparently, it's not the seeds that you dry and eat, but the flowers  1   that cover them. To harvest the pepper, cut some branches to bring back home. This part is already tricky as the branches  do not want to be touched . They are super spiky and will not hesitate to hurt you. Once home, separate the flowers from the branches. Wait for flowers to dry. We are now currently in the waiting phase. Let's see what it looks like in a few weeks! {end-key}  Footnotes    Another Wikipedia nugget: Sichuan pepper in Chinese is literally called  flower pepper .  ↩   "},{ id: 32, directory: "Musings", title: "#39 A week in Hérault", subtitle: "Update #39", slug: "/posts/39-a-week-in-herault", dateCreated: "2023-10-10",  content: " I had my 31st birthday last week. It still feels a bit weird being completely past my twenties. Here, I am just still flailing around, trying to figure out life as I go. At least it's exciting. For the occasion, Robin and I took off to the Hérault department and explored it for a week. Robin currently has a slight knee injury, so we were a bit limited in our activities: no climbing, swimming or hiking, but plenty of eating, wine tasting, village hopping, and library browsing. Here are some highlights:  Morning start at the café in Saint-Jean-de-Buèges, warm sweater, hot coffee and croissant Truly stunning church in St-Guilhem-Le-Désert, pond with carps in the inner garden, warm and quiet light Twisting roads lined with plane trees, sun rays piercing through, long shadows White wine at a wine bar terrace in Sète, a plate of fresh oysters from the producer stall next door Sharing wine and a meal in Karima's flat in Montpellier, floor made of old flat stones Second-hand bookstore in Pézenas, old beautiful books about oddly specific topics A winemaker and his wife in Minerve, his hands still full of dirt from the vineyard as we are savoring the wine Ice cream and coffee in the church square of Sauve, a fresh breeze in the late afternoon light   We met several people on the way — always open, and friendly, often creative. From this short trip, I come back with the renewed motivation to make my own soap, knit something from A to Z, paint some more. Good thing that these are all winter-friendly activities. {end-key}"},{ id: 33, directory: "Musings", title: "#4 Glühwein and note-taking", subtitle: "Update #4", slug: "/posts/4-gluehwein-and-note-taking", dateCreated: "2020-11-28",  content: "    What To Consider When Considering Data Vis Rules  by Lisa Charlotte Ros   Life peek: Wholesome feelings after work  Being in semi-lockdown in Berlin feels both fast and slow. Days resemble each other and night falls early. Yet, I feel like I have been doing a lot of new small exciting things lately — this Update being one of them. Week after week, there are always a few highlights worth mentioning. Work wise, this last week was pretty intense for me: lots of Black Friday-related emergencies (ugh, marketing agencies) and the confirmation of my promotion to Senior Designer. After hectic days at work, Robin and I did what was left for us in a city half in lockdown: go for a walk in the cold and drink Glühwein. It seems like every coffee shop, bakery and restaurant has now started making its own mulled wine. Some are fruitier, some are spicier, some are just cheaper. I love them all. I love the contrast between the hot spicy drink, the warm lights and the cold evening weather. It gives me wholesome feelings. Read &#x26; Watched   How to Be Great? Just Be Good, Repeatably (article, 2019)   Like many others, Steph Smith explains why building a  habit of progression  might be your best bet at becoming great. She defines it as a continuous effort, where inputs lead to expected outputs. In short:  If I were to continue this every day for the next year, would I be in a better place?   The Zettelkasten Method: Examples to help you get started. (article, 2020)   Rebecca Williams gives a clear and concise explanation to get started with the Zettelkasten Method — beginner friendly! Her step by step approach with concrete examples really helped me settle on a working note-taking system. Discovered   PaperSpan    Save articles in a few clicks and add highlights for free.  As I got more serious with my note-taking habit, I also needed a place to save what I read. My only requirement was unlimited highlights. Pocket and Instapaper felt a bit overpriced for my simple needs. I decided to opt for PaperSpan which does the job perfectly.  Newslettrs   A newsletter-dedicated inbox out of your email — beta mode. Robin was telling me that I have an early adopter personality. Newslettrs is one of those services I subscribed to less than 30 minutes after discovering them. I had enough of dividing my email inbox between work emails and newsletters — which I think require a completely different mindset. Separating them completely sounded like an ideal plan to me! Overheard   Not only is any sufficiently advanced technology indistinguishable from magic; any sufficiently advanced technologist seems like a magician. — Autotranslucence,  Becoming a magician   {end-key}"},{ id: 34, directory: "Musings", title: "#40 Brr", subtitle: "Update #40", slug: "/posts/40-brr", dateCreated: "2023-10-17",  content: " Winter is coming (brr)  Two days ago, I felt cold again for the first time since last winter. Just a few weeks ago, we were still burning under the sun; but now, summer feels like a long-forgotten memory. You should see me in the morning: wearing multiple layers, a scarf around the neck, warm socks and still feeling  slightly cold . Our house doesn't have any heating system except for a  wood stove in the living room . And come on, it is actually  warm inside  — as in 20°C, we don't need to light a fire yet. (Do we???) I think my body might just need time to adjust — hopefully. We went to a board game festival  A climbing friend (surprise) invited Robin and I to go to a local board game festival  1  . In the last few years, we have started to play a lot more, especially in board game cafés when traveling. Searching for board game cafés is actually on my to-do list whenever we discover a new city. Back to the board game festival, we arrived close to the end of the event but were able to play a few new games.  A two-person game to be played with dice, where the goal is to be the fastest player to find an itinerary between points A and Z. (Ok yeah, it's not clear at all. But I didn't  love  that game, so that's it.)  In the Footsteps of Darwin : A volunteer paired us with a family of three and started explaining the rules to us. Once you thought the list of rules was coming to an end, another one was added on top of it all. It was slightly overwhelming at first, but the game itself was actually  nice . It reminded me of  Azul , the mechanics encouraging you to create your little collection on your side, without being too openly agressive towards the other players. Overall: a relaxing game.  Quoridor : A beautiful game where your goal is to be the first to cross the board. Everyone also has small wooden walls that they can use to add obstacles on the paths of other players.  Twin it! poster : An A1 poster that you can use as a decoration... and mean of torture for your guests. It is filled with different (but very similar) patterns; your goal is to find five twin patterns. Good luck.  2    I got a new phone  ...or more like Robin got me one. He was doing some research for himself, and as I also needed a new phone  3  , we now have twin Samsung Galaxy S22 that we bought second-hand. A few things that convinced us to get that specific phone model is that:  It takes excellent pictures (which we thought would be extra-nice for our coming trip). Since the S22 model, Samsung guarantees 4 years of software updates (which is great). It is affordable in second-hand and not too old (so chances are that I won't need to replace it for longer). ...and other things Robin probably thought about.  Other things I did   I made my  first solid shampoo  a few days ago. I haven't tested it yet, but the process was much easier than I had expected! Once I had the ingredients ready (not that many, not that weird), the full recipe only took 20 minutes and simple kitchen equipment. I browsed for hours on  searchmysite.net , a search engine that only goes through personal websites, the small web, digital gardens and so on. Sites are ranked higher when they have no ads. We watched  A Haunting in Venice , the new Hercule Poirot movie by Kenneth Branagh. It is much darker that the first two movies, which I really liked. {end-key}   Footnotes     Festival du Jeu de Société d'Ardèche , the sixth edition  ↩   I am tempted to recreate something similar for ourselves. It would be a good opportunity to learn to use Inkscape.  ↩   My Samsung Galaxy S7 was not supported by my bank app anymore. Ugh.  ↩   "},{ id: 35, directory: "Musings", title: "#41 Chestnut time", subtitle: "Update #41", slug: "/posts/41-chestnut-time", dateCreated: "2023-10-27",  content: " I'll keep this update short as we have many guests at home at the moment. Robin's parents, brother and aunt are visiting us just at the right time: chestnut season! In the last week:  We gathered and roasted  so many  chestnuts! We discovered that there are many different varieties that all have their specificities. To look into. The  castagnades  — or chestnut festivals — have started in the neighboring villages. We went to the one in Joyeuse, stumbled upon some winemaker friends, ate a lot of chestnut-themed dishes and had fun! We did a lot of end-of-summer tasks in the garden: we strengthened the compost container, removed the dying tomato plants and started to prep some of the terraces for next year. We picked up several baskets of walnuts to dry. We finally started to crack the dried walnuts from last year. We finished watching the second season of  Our Flag Means Death , a hilarious pirate-themed tv series. {end-key} "},{ id: 36, directory: "Musings", title: "#42 The never-ending to-do list", subtitle: "Update #42", slug: "/posts/42-the-never-ending-to-do-list", dateCreated: "2023-11-23",  content: " Life peek  As usual, too many things happened since the  last update . We spent 10 days in Berlin-Brussels-Ghent visiting friends. When we came back, we realized that the  clock is ticking . (Indeed, we will leave for the Big Trip in less than a month. The problem is that we still have tons of unfinished admin tasks to complete before then.) Aaaand... that's what we have been up to for a week now: working through our to-do list. Do you know the 80/20 Pareto rule? The concept about focussing on the 20% most important things leading to 80% of the desired outcome? Unsurprisingly, we are doing the opposite and spending most of our time prepping (and overengineering)  Eau de poisson  — our travel blog. Even though we are trying our hardest at not being efficient at all, we are still managing to make some progress on our to-do list. (Incredible feat, I know.) Between two tasks, we also found time to harvest our olives, harvest some more chestnuts to grill in our stove oven, eat some raclette with friends (yeeess, it's the season!), make a first batch of Sauerkraut. Reading  When in Berlin, I stumbled upon a second-hand copy of one of my favorite books of all time:   The Fifth Season   by N. K. Jemisin. It is the first volume of the  Broken Earth  trilogy, a sci-fi / fantasy adventure with incredible world building, beautiful poetic writing, and an epic storyline. Bonus points: the three volumes won the Hugo Award  three years in a row . For the last few days, I have been rereading the series — and yes, it is still as good. {end-key}"},{ id: 37, directory: "Musings", title: "#43 Fondues and digital storage", subtitle: "Update #43", slug: "/posts/43-fondues-and-digital-storage", dateCreated: "2023-12-08",  content: " We are less than a week away from the Big Trip. Ok, that's not entirely true as we are going to stop in Switzerland for a week first — Christmas, yay!  Anyway , who cares, it feels like everything starts in six days. As we get closer to the deadline, we spend more time with friends, saying goodbyes, eating fondue  1  . The  to-do list  has also shrunk in half. Robin has spent considerable time working on our picture storage and our general digital storage. About that: for the last few years, we had been keeping our pictures on Nextcloud Photos — a great open source alternative to Google Photos. This system has been working so far, but we realized that we didn't want to go on with it. First, it is inevitable that we will hit a storage limit soon enough. Second, our current internet connection in the countryside is really not cloud-friendly (for large files at least). Third,  do we even need to have access to all of our pictures at any time, anywhere???  That third question made up our minds. We are going for a more low-tech approach:   Keep the most recent pictures on our phones — crazy how much storage space there is on those small gadgets.  Every few months, sort the pictures, only keeping a subselection. Upload everything on two separate hard drives (for backups).  It's decided then, we'll leave for our trip with a clean Nextcloud instance,  phones taking good pictures  and a portable hard drive for backups. {end-key}  Footnotes    We managed to eat three fondues in the past week, and that wasn't even planned — surprising, I know.  ↩   "},{ id: 38, directory: "Musings", title: "#44 Family gatherings", subtitle: "Update #44", slug: "/posts/44-family-gatherings", dateCreated: "2023-12-24",  content: " December 24th, early hours, we are finally setting off on our Silk Road journey. We are currently in the train from Zurich to Vienna, and I am struggling to focus on writing this update. Our window keeps offering postcard-worthy vignettes: vertiginous mountains dropping straight into the lake, their sharp peaks cutting into the sky, the quiet dark waters contrasting with the snow-covered mountain tops.  1   We packed the last ten days with warm and fuzzy feelings. We met a few good friends, spent time with family, ate some proper cheese raclette, played plenty of games, said goodbyes, laughed a lot and cried a bit. For our last evening, we had a Christmas dinner with my extended family. My grandparents came to Switzerland with five daughters. These daughters married, had kids — my cousins, who now have partners of their own. All in all, we packed 18 people in my grandma's tiny apartment. And  oh my , I love the chaotic energy of these family gatherings. First off, no one knows the exact starting time of the event. People come whenever they want and bring some food or drinks to share. It can result in an interesting menu. Sushi can rub shoulders with dry meats, Vietnamese salad, rice and curries, noodles and Peking duck, not mentioning the diverse range of desserts. Second, everyone seems to be shouting — or so you would think as an external audience. In fact, people are talking normally, the decibel level is simply  much  higher than for your average Swiss household. Third, we play games with passion. As far as I can remember, my cousins and I have been loudly playing games while the adults are having adult conversations. It is now almost a tradition to bring a game you like and play with those who feel like it. Finally, everything seems to be happening at the same time. One person will still be eating appetizers, as someone else is mixing main dishes, people stand up, change seats,  shout  talk across the table, put something in your plate because-it's-good-you-should-try-it, clear plates, bring desserts, pour drinks, start games. Yesterday gathering was no exception. I went to bed, ears ringing and eyes shining. Today, my eyes feel tired, but my belly, heart (and backpack) are full. {end-key}  Footnotes    To be specific, these breathtaking views can be seen along the  Walensee  between Zurich and Buchs.  ↩   "},{ id: 39, directory: "Musings", title: "#45 Turkish experiences", subtitle: "Update #45", slug: "/posts/45-turkish-experiences", dateCreated: "2024-01-07",  content: " We have now been on the road for about two weeks.  1   As we planned it, we are speeding through the first few destinations: Vienna, Bucharest, Istanbul. These cities and countries are closer to home and easier to access by train for a future trip. As we have a deadline by the end of this month  2  , we are prioritizing the regions that are farther away — where we are less likely coming back to. Needless to say, night trains and eight-hour bus trips are now our best friends. We take breaks here and there for a couple of nights, enough time to cool down, get a sense of a place. At the moment, we are in Türkiye, just leaving Amasya — a small town by the river, tucked in between mountains. Cats are everywhere, people are often welcoming even though hard to communicate with, we drink tea from dawn to dusk, the muezzin meditative chant echoes throughout our days. Robin and I both got sick, from the travels, general tiredness, and are now just getting back on our feet. We are still adapting to this new rhythm, trying to find the balance between exploring, working on our creative projects, and resting. The FOMO is strong, as we go through places we will probably never come back to.  When we were in Ankara, we watched  Love, Deutschmarks and Death (2022)  by Cem Kaya, a documentary about the Turkish immigration to Germany in the sixties. It is about resilience, community, and how music brings it all together. Having lived in Berlin for a few years — right by the Turkish neighborhood, I regret not having seen it earlier. We also watched a couple other Turkish movies (in buses, while on the road). I love how it adds context and depth to the experience here. I am learning new words, new non-verbal cues and social interactions. I will definitely continue doing so whenever possible — preferably during a long bus ride of 8 hours. {end-key}  Footnotes    Btw, we keep a  daily log  on our travel blog (in French).  ↩   It's kind of complicated, but we basically have a plane that we can't miss then.  ↩   "},{ id: 40, directory: "Musings", title: "#46 Three thousand kilometers", subtitle: "Update #46", slug: "/posts/46-three-thousand-kilometers", dateCreated: "2024-01-27",  content: " This is day 35 of our Silk Road journey. Looking at  our itinerary , we have covered approximately half the distance from France to China — hard to believe. We are currently in Tbilisi (Georgia). If I drew a straight line from here to my home in France, it would equal 3270 km!  1   A piece of good news: we were able to receive our Chinese visa in Yerevan (Armenia). Until just a few weeks ago, we were not sure that our China plan would work at all.  About the Chinese visa  A few of the numerous Chinese visa requirements:  The visa can theoretically only be requested in your country of residence. (We obviously don't live in Yerevan.) Once you receive the visa, you only have three months to enter China or your visa expires. (Which allows for a more limited travel time if you ask for the visa early in your journey.)   Thankfully, we stumbled upon the  Caravanistan forum , a.k.a. THE hub for all world travelers. That's where we found out (from other people experiences) that the consulate in Yerevan was a bit looser with the residence regulation. And unexpected bonus, our visas will only expire in six months, so we have  plenty  of time to travel to China. In a nutshell, this means that we can happily slow down our traveling pace. We'll try to stay longer in different locations, get a sense of the place and see if we want to go through faster or slower. The second half of this journey might be quite different from the first. Let's see what it brings. {end-key}  Footnotes    This stretch of traveling on land unfortunately stops here, as we will need to fly into Baku in a week. The land borders of Azerbaijan are all closed since COVID and show no sign of opening anytime soon.  ↩   "},{ id: 41, directory: "Musings", title: "#47 Four months on the road", subtitle: "Update #47", slug: "/posts/47-four-months-on-the-road", dateCreated: "2024-04-30",  content: " We left Switzerland on December 24th, 2023. Exactly four months later, we are completing the first leg of our journey. We have arrived in Hanoi (Vietnam) and we'll settle down here for a month or two. On the way, we realized that the overland itinerary from Europe to Asia is taken by many. We have made a few great friends who traveled alongside us wherever our paths crossed. They made our adventures so much brighter, and I am forever grateful for those encounters. Four months of active traveling, four months of constant moving, no time to get used to a bed, and even less to get used to a city. We took our time getting to Vietnam, spending full weeks in cities that weren't considered touristy — simply to create spaces to breathe, slow down, get accustomed to the changing scenery. However, it still feels like we have been traveling too fast. We had a goal in mind (getting to Southeast Asia by May) and moved forward accordingly. This means we skipped a lot of things and made choices to drastically slow down in areas that might seem odd. Not to  do  anything, mind you ; just to take it all in. We heard of this backpacker, hitchhiking his way across the world, just a few countries behind us. He was supposed to travel for two years, but is already on his way back after six months — burnt out. On the other hand, we met this guy in a hostel, who has been jumping from city to city, from dorm to dorm, for two full years. The duration he spends on average in one location is two weeks ; his max so far has been six weeks. Travel burnout seems to be a thing. Everyone has their own threshold. For us, it feels like  now  would be a good time for a longer break. Settling down in Hanoi — a city we know and love — seems like the perfect choice. We are now back in (or pretty close to) our comfort zone. We are still discovering new things every day, but in a context familiar enough to be able to think further than the day-to-day logistics. During this parenthesis, I'd love to take the time to process the last hectic months, look back on the places and experiences we have been through. Documenting it on  eaudepoisson.com  (for memories) and maybe even creating a small analog zine out of it (for fun!). Of course, there will also be plenty of planning and decision-making for the rest of the journey. But let's not get ahead of ourselves. Until then {end-key}"},{ id: 42, directory: "Musings", title: "#5 Coffee tasting and cozy December", subtitle: "Update #5", slug: "/posts/5-coffee-tasting-and-cozy-december", dateCreated: "2020-12-05",  content: "    VDInk 03/31  by Osmar Axel Rivera Mejia   Life peek: Cozy December  Last Tuesday, something very special happened: December 1st. Robin and I were waiting for it for a long time as we could finally open our Coffee Advent Calendar from  19grams  — a Berlin-based coffee roastery. I am now excited to start every morning with a meditative coffee tasting session. Besides that, this week definitely had some end-of-the-year vibes. I have been listening to a lot of chill classics: some jazz, soul and a bit of blues. It actually feels so much like cozy December that I even started reflecting on the year — here is my  Year in Review 2020 . I like Steve Schlafman's approach of spreading this reflection over two weeks. For once, it doesn't feel like an overwhelming chore. Read &#x26; Watched   Some Kids I Taught and What They Taught Me (book, 2019)   Kate Clanchy loves her students and you feel it. In this memoir, she invites you to meet some of them on a journey about passion, cultural bridges, pain and creativity. A moving and raw account of thirty years of teaching and learning. I am now going straight away to her new book  How to Grow Your Own Poem .  How Technology Grows (article, 2018)   In this long-form article, Dan Wang denounces our fixation on the digital and computing industry when talking about technology. He explains the three forms technology can take — as tools, instructions or process knowledge. Very interesting read on the lure of the digital and what truly drives innovation.  Tasting RARE Russian Food in Karelia! (video, 2019)   I've just discovered this Berlin-based food vlogger on YouTube and I already love his channel. I can't exactly say what I like about it. Maybe it's his long historical and contextual intros, the way he travels — that feels so close to mine — or maybe just his genuine love for food. Discovered   Hypothesis   Was I talking about PaperSpan just last week? Well... Robin convinced me to upgrade to the next level of annotating tools: Hypothesis. Built on open-source technology and interoperable standards, you can highlight and annotate the web completely free of charge. Its development is transparent and collaborative through their public roadmap on GitHub. {end-key}"},{ id: 43, directory: "Musings", title: "#6 Bringing home some old records", subtitle: "Update #6", slug: "/posts/6-bringing-home-some-old-records", dateCreated: "2020-12-12",  content: "    Christmas Music animated illustration  by Maria Scherlies   Life peek: Books and records  This last week, I was super excited to finally receive the   Data Visualization Handbook   that I ordered from a Finnish university. I have just started it, but have already taken a few notes that might enter my Zettelkasten. On a side note, the book itself is gorgeous with a very smart layout. I am really enjoying it so far! There are always heaps of zu verschenken cardboard boxes in the Berlin streets. Some contain treasures, while others are just useless things people leave on the sidewalk when moving out. On Saturday, Robin and I stumbled upon one box full of old vinyl records. They were mostly classical music — which is not our go-to genre — but we still took a few records home. Our plan for the next few weeks: broaden our knowledge of classical music. Read &#x26; Watched   The Polymath Playbook   I had this article on my to-read list for quite some time, but only got to it this week. I read it in parallel with The definition of a polymath by the same author and realized that I tend to be more of a scholar than a polymath — learning as an end goal, instead of applying this new knowledge in real life. One quote that stuck with me:  The best place to start? That thing you secretly want to do.   The Untouchables (movie, 1987)   I have been revisiting classics this last week — or more like first time visiting them. After  The Good, The Bad and The Ugly , I travelled back to the prohibition in Chicago. The photography of  The Untouchables  from De Palma is spectacular — you could definitely frame most shots. Discovered   GitJournal   I currently use Foam and Obsidian for my Zettelkasten. They both work great on desktop, but I was missing a mobile solution to record fleeting thoughts and notes on the fly. GitJournal is an open-source note-taking app that syncs with the Git repository of your choice. Overheard   I'm barely keeping up in most conversations — I think that's my learning style? — Brandon Toner on Twitter  {end-key}"},{ id: 44, directory: "Musings", title: "#7 Living better with less", subtitle: "Update #7", slug: "/posts/7-living-better-with-less", dateCreated: "2020-12-19",  content: "    Unused artwork  by John Mata   Life peek: Fostering Conversations  I am not much of an outdoor person — lots of allergies and just  life ya know . From time to time, I will force myself out of the door for a long walk with Robin. And we invariably talk about thought-provoking topics. I am not sure yet, but I think the secret lies in having a protected space in time for a passive activity — walking in this case. This passive activity doesn't use a lot of mental space. It allows your mind to wander just enough, which fosters these organic conversations. Our morning coffee tasting routine is another example. More often than not, it will conclude on an in-depth discussion about something random. I might just bring up this topic on our next walk. Read &#x26; Watched   Your Lifestyle Has Already Been Designed (article, 2010)   This article shook me. It was written in 2010 by a guy finding a 9–5 job after 9 months backpacking. He talks about how much time and energy your regular job uses, resulting in you paying for convenience, rewards, a bit of excitement, an escape. In other words: the more we work, the more it pushes us to consume. The more we consume, the more we have to work to afford it.  The Nuclear Family Was a Mistake (article, 2020)   Fascinating article about why we envision the ideal family as two parents with 2.5 kids, and how the stable nuclear family isn't viable anymore. Spoiler: it is because our economical and cultural situation is very different. The author explains how we are now going back to new extended family structures with forged families and communities. Antilibrary   Wabi-Sabi (2008)    Wabi-sabi  (侘寂) is a Japanese concept describing the appreciation of things imperfect, impermanent, and incomplete. It is a short meditative read by Leonard Koren that might help me to let go and be a bit less of a perfectionist.  Recommended by  Dense Discovery    The Art of Frugal Hedonism   Written by Australian couple Annie Raser-Rowland and Adam Grubb, this self-help book sounds like a light-hearted (and funny) read about living better with less. High-priority on my to-read list.  Recommended by  Dense Discovery   Discovered   Dark Reader   I didn't discover this browser extension in the last week, but it is just too awesome to not mention it at least once. Context: I do  everything  in dark mode. So whenever a website not supporting dark mode pops up, my eyes hurt. A lot. When enabled, Dark Reader automatically switches those websites to dark mode with your preferred settings. Bonus points:  it's open-source .  Date Night Questions   If you have enough of small-talk leading nowhere, this web-app is made for you. With different decks for light conversations vs intimate talks, you will stumble upon questions like:  About what are you superstitious? Is there anyone you look up to? When is the last time I impressed you?  Overheard   We judge ourselves by our intentions and others by their behaviour. — Stephen M.R. Covey  {end-key}"},{ id: 45, directory: "Musings", title: "#8 Remote Christmas", subtitle: "Update #8", slug: "/posts/8-remote-christmas", dateCreated: "2020-12-26",  content: "    Present yoga illustration  by Olga Semklo   Life peek: Christmas via Skype  Christmas is an important event in my family, even though no one is Christian. It has always been a pretext to have a large family gathering with good food — and when I was younger — presents. This year, we weren't able to meet in person because of the pandemic. Instead, we had a video conference dinner party. In total: four cameras and thirteen people, all eating and talking at the same time. Needless to say, it was a hectic and noisy event. Yet, I felt at home in the middle of this chaos. Remote or not, the atmosphere was comparable to the one of our usual family reunions: heart-warming, frenetic, sometimes boring, but mostly fun. Read &#x26; Watched   The Murderbot Diaries (books, 2017)    The Murderbot Diaries  by Martha Wells have a Deadpool meets The Martian vibes. Created by the Company, Murderbot is a conscious (and self-conscious) machine that managed to hack itself to become free. It could have taken revenge on humans but chose instead to binge-watch hours of TV series. Who said sci-fi had to be serious? Discovered   IndieWeb   The IndieWeb is a movement that Robin introduced me to a few months ago. The concept is to take control of your content and be able to connect with others without using silos — centralized for profit websites. This last week, I finally acted on that knowledge and started to create my IndieWeb friendly website in Next.js. I am pretty excited by this project and will tell you more about it in the next issue 📪  Excalidraw   Excalidraw makes it easy to create amazing-looking sketches in your browser. Draw a diagram or write your love message directly on the board before sharing your unique link with whomever you want. {end-key}"},{ id: 46, directory: "Musings", title: "#9 New website for 2021", subtitle: "Update #9", slug: "/posts/9-new-website-for-2021", dateCreated: "2021-01-02",  content: "    Halo  by Marina Dunbar   Life peek: New website for 2021  I had an incredibly productive last week of the year, as I threw together my new website from scratch  1  . Having a clear and attainable goal really helped me speed learn through the basics of React. I realized that I actually do learn best by  doing . With Robin, we spent New Year's eve doing a number of adult things:  having a fancy meal (homemade tagliatelle topped with  this grated cheese ) with wine (brought back from our cycling trip to the Chianti region) discussing our Year in Review (based on  Schalfman's template ) going to bed at 11.15pm  Read &#x26; Watched   Soul (movie, 2020)   We all agree that 2020 was pretty sad in terms of movie releases. Well, Pixar just saved the year by releasing  Soul , an animated movie about passion, jazz, purpose and what it means to be alive. I just added it to my list of favorite Pixar movies of all time — just next to   Coco  (2017) .  2   Antilibrary  Everything is closed in Berlin at the moment, except for restaurant takeouts, grocery stores and... bookstores. We decided to cycle to Dussman, a giant bookstore with an amazing English section. It couldn't be avoided, my  antilibrary  got quite a few additions.  Piranesi (2020)   As usual, I spent way too much time in the fantasy / sci-fi section. With its mythological, dreamy setup,  Piranesi  really reminded me of   Circe   by Madeline Miller — which I loved. This novel by Susanna Clarke will probably not be in the antilibrary very long as I intend to read it  very  soon.  Children of Time (2015)   I know nothing about the author — Adrian Tchaikovsky, but I know I love smart sci-fi and space odysseys that dive into civilization building. I am adding it here for next time I will be in this kind of mood. Discovered   Tailwind CSS   Okay, this tool is definitely not new on the market. But as I only knew HTML and CSS so far, Tailwind CSS was a little revolution. It made the whole website building process a lot easier and beginner-friendly for me. {end-key}  Footnotes    Or practically in my map of the world, as I used this  Next.js starter .  ↩   I also watched the new DC release:  WW84 , but needed to close my eyes at times because it was just too cheesy to witness.  ↩   "},{ id: 47, directory: "Musings", title: "A list of Berlin things", subtitle: "A non-comprehensive list of things I'll miss from Berlin", slug: "/posts/a-list-of-berlin-things", dateCreated: "2022-12-31",  content: " It's the last day of 2022, 5.40 AM, and I can't seem to sleep anymore. A lot of thoughts are going through my mind. Most of them relate to our  move to France  in less than a month. I won't dive into the list of logistical to-dos that is keeping me awake. I'll go the other way around and take a snapshot in time. A snapshot of our time spent in Berlin and the things I'll miss from here.  Our morning bouldering routine with friends, climbing at 7 AM and going to  La Maison  for pastries The amazing Neapolitan pizza restaurants Getting our weekly groceries from  Herbie Unverpackt  and chatting with Carmela The extensive and delicious Turkish food offer in Neukölln The Middle Eastern breakfasts on Sonnenallee Being welcomed regulars at  La Vineria , our local wine bar Cycling everywhere Sunny lunches and coffee breaks by the Landwehr canal Playing chess or working at  Geschwister Nothaft Café  Reliving the Southeast Asian food stall experience at the Thai Park Cooling down in lakes in the summer Being max. 15 minutes walk from everything we love Weekend trips in Poland to eat pierogi The strong and friendly bouldering culture Beers and picnics at Tempelhofer Feld The carefree pace of life Playing table tennis in the streets on summer nights Always discovering cool new places to eat and drink The diverse mix of people playing pétanque in front of  La Maison  The super spicy Sichuanese restaurants Glühwein at Christmas markets And all the precious friends we've made on the way  I have forgotten many things, for certain. But as my friend Max said:  I am not crying, you are . Anyway. {end-key}"},{ id: 48, directory: "Musings", title: "A walk down the Sonnenallee", subtitle: "Along the Arab street, from Hermannplatz to Sonnenallee station", slug: "/posts/a-walk-down-the-sonnenallee", dateCreated: "2023-11-07",  content: " Robin and I are back for a week in Berlin. Among many things, we went back to North Neukölln — one of our favorite neighborhoods in Berlin. This time around, we slowly walked down the Sonnenallee, or the Arab street as some call it. In total: 38 minutes to walk from Hermannplatz to the Sonnenallee S-Bahn station. As we leave Hermannplatz, we first stop at  Al Faisal Grills and Restaurant , a Syrian joint that is always full. We grab a  manousheh  for two to share on the way. With zaatar and olive oil dripping on our fingers, we get deeper and deeper into the street. Store signs transition from German to Arabic only, Halal butchers appear here and there, people around us are speaking Arabic first and foremost. Our second stop is at  Akroum Snack , a tiny restaurant that you might miss if you are not paying attention. We order a  fatteh  to share, and it is honestly the best fatteh I know of: creamy, generous, topped with thin and super crispy pieces of fried pita. As we are waiting for our order, we get the usual Arabic tea served in these types of restaurants (self-service style):  Take a glass. Pour a few centimeters of black tea from tap 1. Fill the rest of your glass with hot water from tap 2. Add a generous amount of sugar. Mix, taste and adjust ingredients if necessary.  Our order arrives, and we walk to a table to the back of the restaurant. From there, Robin has a good viewpoint on the rest of the restaurant. Thanks to that perfect viewpoint, we learn a new fatteh-eating technique! When eating fatteh, we used to overfill our pita breads with it, always making a mess. Robin has now observed that  some people  take a small piece of pita, shape it like a cone, fill that cone with fatteh and eat it all in one single bite. Tested and immediately adopted. At that point, our bellies already feel quite full (we had eaten breakfast not so long ago). We stop at supermarket stalls on the way, looking at the products displayed: grape leaves in jars, instant noodles with an Arabic branding, hummus sold in cartons, giant bags of dried coral lentils, trays of dried dates, all kinds of BBQ related products (coal, gas, grills). Oh wait, we track back a few steps for a final sweet stop. We have never noticed this pastry shop before:  Malek Al Konafeh . They specialize in  kunafeh , a dessert made with spun pastry, soaked in sweet syrup, layered with cheese and pistachio. We take our sweet and cheesy dessert and savor it at a table right outside the shop. It fills whatever space was left in our bellies. As we eat, we observe whatever is happening on the street. Cars honking, one small red car stopping in the middle of the street, the driver comes out and changes his license plate. We draw some more sweet cheesy strings from our aluminum plate to our mouths. Some drivers get out of their cars and are now discussing in the middle of the street. A big white car comes out of a parking space, immediately replaced by the small red car. The red car driver replaces his license plate with a new one once again — it is put on upside down. Almost bursting, we continue our walk down Sonnenallee: shops selling wedding dresses and hijabs, pastry shops selling giant colorful cakes and baklava, gold jewelry stores, Döner joints, shops displaying shiny home decor items. As we get closer to Sonnenallee station, the street chaos seems to tame down, there are fewer people, more space, less noise. But the true end of the Arab street might be marked by the clothing store specialized in techno rave outfits — mostly skimpy black leather clothes that look like bikinis. We do a final stop as we arrive at the Sonnenallee station. We order a filter coffee and a flat white in a familiar hipster coffee shop. Students and young professionals are sitting at various tables, some are wearing beanies, thick black leather boots, most are working on their laptops. From there, we see and hear the S-Bahn run by. It feels like the last three kilometers could have happened in a dream. We are gently waking up. {end-key}"},{ id: 49, directory: "Musings", title: "Add a crayon effect to your illustrations", subtitle: "An Adobe Illustrator CC tutorial", slug: "/posts/add-a-crayon-effect-to-your-illustrations", dateCreated: "2019-04-30",  content: " I love the cleanliness and flexibility of vector illustration. But sometimes, it just feels…  too perfect . The grit and texture that come with raw sketches usually get lost translation. Today, I'll show you how to easily bring your creations to life with a crayon effect, using Adobe Illustrator only. Let's get started! 🚀   Before > after   Step 1  Open a copy of your original vector illustration in Adobe Illustrator. (Wouldn't want any accident happening, right? 😘) Step 2  Open the  Brushes  panel ( Window > Brushes ) and check  Show Art Brushes  in the panel menu.   Select your favorite brush!   Select the stroke you would like to convert into a crayon line, scroll down the brushes and choose your favorite one. (So many brushes! 🖌)   Repeat for all the vector strokes you'd like to convert into brush strokes.     Step 3  In this example, I'd like to add a brush stroke around the face to define it better. To do so, begin by making a copy of the face shape layer (here called skin).   Copy of the skin layer     Swap the swatches   With the layer selected, convert the shape into a stroke by swapping the Fill and Stroke swatches ( Shift + X ). You now have your outlining stroke! (Yaay 🎉) Feel free to edit the color and line weight to see it better. Step 4 (Optional)  Let's say you don't want your newly created stroke to completely outline the face shape (because that would just look weird). Let's see how we can manage to have a clear hairline at the top with ✂ Use the  Scissors  tool ( C ) to cut two endpoints on your stroke to delete that unwanted upper stroke.   ✂ Select endpoint #1…     …and endpoint #2     See? No more stroke!   Once you have your final stroke, convert it as previously done (Step 2) into a brush stroke. Step 5  Add some details to your illustration to bring it to life. Here, I've added definition to the hair and pink to the cheeks. And I got this flawless self-portrait as a result 💁‍♀️   So flawless.   Step 6  Before celebrating the creation of your masterpiece, convert your entire illustration to an outline ( Object > Path > Outline Stroke ). This will make sure that all the proportions stay the same, even if you scale your illustration up or down. (Yep, I'd rather stay the same size everywhere…)   Outlining makes sure to keep the proportions right.   And ta-da! 🎉   Good job guys 👍  "},{ id: 50, directory: "Musings", title: "Choses que j'ai apprises à Berlin", subtitle: "Un exercise de poésie", slug: "/posts/ce-que-jai-appris-a-berlin", dateCreated: "2023-08-20",  content: " À assumer ma singularité en m'habillant tout de noir Qu'un bon Döner est (1) pas cher, (2) à moins de 5 minutes de marche À traverser la route en moins de temps qu'il n'en faut pour dire  Kartoffelpuff-  À arrêter toute activité en cours pour profiter d'un rayon de soleil passager À boire une bière en marchant, en métro, à vélo Qu'en mars, une journée n'est pas complète si on n'a pas vu la neige, le ciel bleu et la grêle Quand manger des asperges (à tous les repas, tant qu'il y en a) Que seuls trois sports importent : le ping pong, la pétanque et le bloc Qu'en décembre, on fait le plein de Glühwein (faute de soleil) Où passer le meilleur Nouvel An : chez soi, les volets fermés Et bien sûr, à parler allemand : Hallo, ich möchte ein Schokocroissant bitte danke ciao {end-key}"},{ id: 51, directory: "Musings", title: "City kids going cottagecore", subtitle: "Moving to the countryside", slug: "/posts/city-kids-going-cottagecore", dateCreated: "2022-12-15",  content: " So, Robin and I are moving to the countryside in France pretty soon. End of next month actually  1  . The thing is that I am a city kid. Okay, sure, my hometown in Switzerland is really not that big. In all fairness, you could actually say that it's tiny (whooping population of 12,000+ people).  But , it is still a town. It has public transports, libraries, grocery stores, fancy or dirt-cheap restaurants, all the conveniences of every day life. In contrast, we are going to move to the most remote village in the Ardèche region. It has at most five permanent residents, you can only get there by car, and the closest city — if it qualifies as a city at all — is 30 minutes away (by car, of course). In short, I am both excited and anxious about this move. At least, it's not a no-return type of decision for us.  2   Now, I am pretty sure that I imagine and idealize life in the countryside on many levels. But here is still a list of things I am looking forward to start doing (or doing more of)  3  :  Making more things from scratch Fermenting and preserving all kinds of vegetables Cooking bigger quantities of delicious food to store (in a huge freezer, yay) Building and fixing things, from woodworking to house improvements Getting into pottery Learning more about plants and animals (especially the local ones)  Well, we'll see if reality holds up to these projections. {end-key}  Footnotes    Writing it down make it feel so  real , oh my.  ↩   We are lucky enough to have Robin's grandma sheltering us, as her house is big enough, and her heart even bigger.  ↩   It might be that I will never get to doing these, just like New Year resolutions.  ↩   "},{ id: 52, directory: "Musings", title: "High tech", subtitle: "Comic strip #1", slug: "/posts/comic-high-tech", dateCreated: "2021-10-19",  content: "  "},{ id: 53, directory: "Musings", title: "Obstacle", subtitle: "Comic strip #2", slug: "/posts/comic-obstacle", dateCreated: "2021-10-21",  content: "  "},{ id: 54, directory: "Musings", title: "Create a duotone image in 5 minutes", subtitle: "An Adobe Photoshop CC tutorial", slug: "/posts/create-a-duotone-image-in-5-minutes", dateCreated: "2019-05-09",  content: " There is a new kid in town and its name is  duotone . You've certainly seen that refreshing effect used more than once this past year. The duotone effect consists of combining two contrasting colors in your image — one color for the dark tones and one color for the light tones. You can easily use this effect to strengthen a brand visual identity based on a few colors and get a vibrant result. Let's get started! 🚀   Before > after   Step 1  Choose a super cool image and open the  Layers  panel ( Window > Layers ). Click on the  Adjustment Layer  icon — the black and white circle thingy — and select  Gradient Map…    Gradient Map…> In the Layers panel, click on Adjustment Layer > Gradient Map…   Step 2  A gradient preview appears, click on it to open the  Gradient Editor . (The default gradient looks okay, but meh, it would just be waaay too easy 😘)   Click on the gradient preview.   Step 3  Here comes the fun part. Double-click on the left Color Stop to change its color. Do the same for the right Color Stop.  By default, the  left Color Stop  replaces the dark tones of your picture 🕶 The  right Color Stop  replaces the light tones ☀  Feel free to move the stops around to control the color distribution 🎨   And here is your duotone image! 🎉   Yay, you got your beautiful duotone image!   Step 4 (Optional)  Wanna get a bit crazy and go  triotone ? Add additional Color Stops to your gradient by clicking right underneath it. In this example, I've added a lighter blue in the middle of my gradient to give more definition to my picture.   You did it! 💁‍♀️   Triotone power! 💜💙🧡  "},{ id: 55, directory: "Musings", title: "Dans la cabane", subtitle: "Un exercise de poésie", slug: "/posts/dans-la-cabane", dateCreated: "2023-06-16",  content: " Ça fait un moment que j'essaie de faire plus de projets créatifs. Il y a maintenant un an, j'étais tombée sur le Twitter de Kate Clanchy, une poète et enseignante anglaise. Elle a une approche à la poésie très vulgarisante, permettant à ses élèves de tout milieu de s'y mettre. J'ai commencé son livre  How to grow a poem  il y a quelques jours. Chaque chapitre est composé d'exercices très pratiques pour écrire différents poèmes. Je n'ai jamais vraiment fait de poésie, malgré le fait que plusieurs personnes de ma famille en ont beaucoup écrit (en vietnamien). Comme on dit, il n'est jamais trop tard, je me lance.  Dans la cabane  Un enfant encore rempli d'émerveillement Rentre dans sa cabane. En son centre, il met Ses jumelles, Son carnet de nuages, la première page commencée. Dans la cabane, il met les biscuits aux amandes de maman. Il y met son meilleur copain Célestin. Il ajoute son vélo rouge, les petites roues arrières, Puis finalement les enlève, il n'en a plus besoin. Il y met les découvertes qu'il a faites, Les fleurs qui s'endorment quand il pleut, Les fourmis qui marchent à la queue leu leu. À sa cabane, il ajoute le vieux squelette brun de tricératops visité au musée la semaine dernière avec l'école. À côté, le coprolite trouvé avec papy. Il y dépose ses pleurs avant le premier jour d'école, Et tout doucement, le bisou de Marie-Lou. (C'est son amoureuse.) Il fait de la place pour la navette spatiale qu'il construira quand il sera grand, Ses aventures d'explorateur de l'espace, Même celles où il aura au moins trente ans. Il les place toutes dans la cabane. Il y a déjà mis beaucoup de choses dans sa cabane. Mais elle est plus grande qu'elle n'en a l'air, Il pourra y mettre encore plus de choses demain. {end-key}  Texte inspiré du poème  Table  d'Edip Cansever. "},{ id: 56, directory: "Musings", title: "Dans le quartier de Thanh Công il y a le village de Thanh Công (2004)", subtitle: "A tiny note about the short", slug: "/posts/dans-le-quartier-de-thanh-cong-il-y-a-le-village-de-thanh-cong-2004", dateCreated: "2023-04-02",  content: "  Directed by Vàng Anh Phan Thi  This short of 33 minutes brought me right back in the streets of Hanoi. Through her lense, Vàng Anh Phan Thi shows us the reality of living in this small neighborhood. The speakers installation storyline felt almost like a backdrop to the daily lives of Thanh Công inhabitants. Rating: 4/5 {end-key}"},{ id: 57, directory: "Musings", title: "Destination", subtitle: "A stop motion experiment", slug: "/posts/destination", dateCreated: "2022-02-13",  content: " So yep, I finally got the  disease . In the last few days of quarantine, I was feeling 100% recovered and had way too much time on my hands. This is what happened.  .embed-container { position: relative; margin: 2rem 0; padding-top: 56.25%; height: 0; overflow: hidden; max-width: 100%; } .embed-container iframe, .embed-container object, .embed-container embed { position: absolute; top: 0; left: 0; width: 100%; height: 100%; }      I unexpectedly enjoyed the creative process and even  took a few rough notes  for the next  *fingers crossed*  project. {end-key}"},{ id: 58, directory: "Musings", title: "Digital doodling", subtitle: "Trying out Tayasui Sketches", slug: "/posts/digital-doodling", dateCreated: "2021-12-25",  content: "   It's been a long time I had wanted to get a bit more familiar with digital drawing. I discovered the  Tayasui Sketches  software that was just what I needed: available on macOS, free to use, fast to learn, delivering great results out of the box. Above is one of the first doodles I made using that tool! {end-key}"},{ id: 59, directory: "Musings", title: "Down By Law (1986)", subtitle: "A tiny note about the movie", slug: "/posts/down-by-law-1986", dateCreated: "2020-11-26",  content: "  Directed by Jim Jarmusch  Not so much about the prison and crimes, than the encounter of three very different men that become friends. Rating: 4/5 {end-key}"},{ id: 60, directory: "Musings", title: "Fermentation: behind-the-scenes", subtitle: "Biochemistry stuff I barely understand", slug: "/posts/fermentation-behind-the-scenes", dateCreated: "2023-07-27",  content: " In the last few years, I have superficially dabbled with fermenting my own food, making kimchi, Sauerkraut, kombucha or kefir. I have also learned that so many of the things I drink or eat are in fact fermented — think cheese, wine, beer, yogurt, bread. Since Wednesday, I have been diving deep into the intricate world of fermentation from a complete other direction. I stepped outside my comfort zone and had a look at this whole fermentation shenanigan from the point of view of biochemistry. Here are the things I've learned.  (Disclaimer: I am mostly quoting Wikipedia as that's where my fermentation rabbit hole started. Please  get back to me  if I wrote anything wrong-I-don't-really-know-what-I'm-talking-about.)   What is fermentation?  In biochemistry,  fermentation  is the process where living organisms extract energy from carbohydrates (e.g. starch or sugar) in the absence of oxygen. The food industry sets a broader scope, as it is more largely defined as a metabolic process that produces chemical changes in organic substances through the action of enzymes. There are tons of different fermentation processes that can produce a number of things: alcohol (a.k.a. ethanol), acid, carbon dioxide, and others. For instance, yeasts are used in wine and beer-making to produce alcohol, while bacteria are used for yogurts to produce acid. Sometimes, a mix of both yeasts and bacteria can be used in tandem as well. Let's have a look at a few types of fermentation used in the food-industry — because who cares about the other types anyway.  1   Lacto fermentation   Lacto fermentation , also called  lactic acid fermentation , is the process where lactic acid bacteria (LAB) convert sugars into energy, with lactic acid (and sometimes other stuff as well) as by-products. This occurs in an anaerobic environment  2  . Common examples of lacto fermentation are kimchi, Sauerkraut or yogurt. LAB already naturally exists in many vegetables, so you don't need to add external bacteria for the fermentation to occur. The bacteria keep on living in the final product and make them  probiotic , which is good for your gut health. Let's use the  Sauerkraut recipe  as an example. At its simplest, you only need to mix cabbage with water and let it rest at room temperature to kickstart the fermentation process. Notice how the recipe insists on preventing the minced cabbage from coming into contact with the air. The reason is to keep the environment anaerobic, where LAB thrive whereas other *bad bacteria* would die. As the fermentation goes on, all the glucose present in the cabbage is transformed into lactic acid, making Sauerkraut sour. Another example of lacto fermentation is cheese. The basic concept of cheese is to divide milk into solid curds and liquid whey. To do so, we need to either add an acid element to the milk (e.g. vinegar) or mix in LAB that will consume the lactose and transform it into lactic acid.  3   Ethanol fermentation   Ethanol fermentation  or  alcoholic fermentation  is the process where yeasts — a.k.a. mushrooms — convert sugars into energy in an anaerobic environment (same old, same old). However, this time, the by-products are ethanol and carbon dioxide. In the case of food, ethanol makes things — well — alcoholic, whereas carbon dioxide produces bubbles. In short, this type of fermentation is used for wines, beers, whiskeys, rums and tons of other alcoholic beverages. In most cases, the brewers will let the carbon dioxide escape during a primary fermentation. When needed, a second fermentation (sometimes with extra yeast) can be done in the final bottle. As the carbon dioxide doesn't have anywhere to go, it is integrated into the liquid as bubbles — yay! Beers and sparkly wines are perfect examples of this second fermentation. With ethanol fermentation, the yeasts die after a while: either when all the sugars have been converted into ethanol or when the alcohol content has reached 15% alcohol per unit volume — strong enough to kill all types of yeasts. When a beer is unfiltered, you can see the sediment of dead yeasts remaining at the bottom of the bottle. Winemakers also use this yeast property to make fortified wines (e.g. Porto wine). They can stop the fermenting process by adding some extra alcohol in the wine. This usually results in a sweeter and stronger wine. The fermenting process can also be stopped by simply dropping the temperature of the tank.  Fun fact: ethanol fermentation also takes place in some species of fish where it provides energy when oxygen is scarce.  Malolactic fermentation   Malolactic fermentation  is a process in winemaking. It is usually done after the first ethanol fermentation, by introducing LAB into the must. Those bacteria will consume the tart-tasting malic acid, naturally present in grape must, and produce softer-tasting lactic acid and carbon dioxide. The final aromas are said to be less acid, apple-like and more buttery, milky, oily or vanilla-like. Most red wines undergo malolactic fermentation, as well as some fuller white wines like Chardonnay. LAB can be very sensitive to the sulfites (sulfur dioxide) added during the winemaking process. When choosing to not use sulfites, wild LAB native to the grapes can initiate uncontrolled malolactic fermentation that might give an off-taste to some natural wines. Carbonic maceration   Carbonic maceration  is a fermentation technique usually used for specific types of wines (e.g. Beaujolais). It results in wines with low tannins that should be drunk young. It actually falls under the ethanol fermentation family. The difference is that instead of crushing the grapes, most of the fruits are kept whole in a sealed container. The fruits at the bottom are crushed­ — because of  duh  gravity — and undergo conventional ethanol fermentation. The carbon dioxide produced saturates the air, creating an anaerobic environment  4  . The carbon dioxide gas stimulates the fermentation process, which occurs inside of each berry. Carbonic maceration started as a winemaking technique, but people have started using it during coffee preprocessing as well. Before the coffee berries are ready to be roasted, they need to first be fermented, so that the fruit pulp can be removed from the seed. The method is similar: keep the berries intact in a closed vessel to undergo a carbonic maceration. In the case of coffee, the vessel is pumped with extra carbon dioxide to begin with. The resulting coffee profiles are often described as more intense. Aerobic fermentation   Aerobic fermentation  is the process where acetic acid bacteria (AAB) metabolize sugars or ethanol in an aerobic environment  5  . Its by-product is acetic acid. As there is oxygen involved, it is more accurate to call this fermentation an  oxidation  process. This type of fermentation is used for the production of vinegar. If making wine vinegar, the ethanol present in the wine is converted into acetic acid. This is also why we usually use sulfites in the winemaking process. Sulfites prevent oxidation, thus avoiding the wine transforming into vinegar. Special mention: SCOBY  A SCOBY is a  symbiotic culture of bacteria and yeast . It is a culinary fermentation culture (or starter) where we find both yeasts and bacteria living in harmony — isn't it beautiful? Common examples of food / drinks made from a SCOBY are kombucha, kefir, sourdough bread, vinegar, soy sauce or sour beers. A SCOBY is a mix of yeast, LAB, and AAB. The SCOBY, when mixed with water and sugar-rich ingredients, undergo multiple types of fermentation and oxidation. To be completely honest, I haven't fully understood what ferments what and when. But all in all, yeasts consume sugars and produce ethanol. The LAB (anaerobic) consume sugars and produce lactic acid. The AAB (aerobic) consume either sugars or ethanol and produce acetic acid. The end results is often acidic and just a tiny tiny bit alcoholic.  That's it for today's episode on food fermentation! I might update it in the future, as I learn more about this fascinating world! {end-key}  Footnotes    Actually, fermentation also occurs in other super interesting cases. For instance, during intense exercise, Mammalian muscles will undergo fermentation and transform some of those stored sugars into energy and lactic acid as a by-product. I now finally understand why my arms hurt after an intense climbing session!  ↩   An anaerobic environment has no oxygen.  ↩   A lot of cheeses are also made with rennet, a set of enzymes, that will also divide the milk into curds and whey. The bonus effect is that the curds will be less fragile and more gel-like.  ↩   Carbon dioxide is heavier than oxygen. As it saturates the tank, oxygen is pushed out from the tank through a one-way valve.  ↩   They like oxygen.  ↩   "},{ id: 61, directory: "Musings", title: "Free Solo (2018)", subtitle: "A tiny note about the movie", slug: "/posts/free-solo-2018", dateCreated: "2020-11-26",  content: "  Directed by Elizabeth Chai Vasarhelyi and Jimmy Chin  One of the only times I had cold sweat during a film. Rating: 4/5 {end-key}"},{ id: 62, directory: "Musings", title: "Gardening", subtitle: "Another try at knowledge management", slug: "/posts/gardening", dateCreated: "2021-10-24",  content: " For many years, I have struggled with knowledge management — as I still do. I remember enthusiastically filling small cards a few years back when I first stumbled upon the  index card method  — or Zettelkasten. It never took off. A few years later, I discovered that this method had been adapted digitally by multiple different projects. I have a (too) early adopter personality and tried them all —  TiddlyWiki ,  Obsidian ,  Foam . Actually, I even mentioned it in a  weeknote  just last year. I realize that the problem doesn't lie in the tools and features, but being able to fully commit to a process. I tend to over-complicate things and not start as simple as I should. Building that new habit is then much harder, as the barrier to entry seems too high and complex. Currently, I am trying to not focus so much on the tool and its features, but on the note-taking habit itself. I recently created a super simple note-basket where I could record my evolving notes and thoughts on different topics. It exists without a publish button — in the sense that a note will be updated again and again without having a final version. So far, I try to not over-engineer it and not censor myself too much. No back-linking, predefined categories, or other must-have fancy features. I want to — will — build a solid habit first and add on top of it later. Anyway. {end-key}"},{ id: 63, directory: "Musings", title: "I made a font", subtitle: "Or how I created a simple font as a complete typography-noob", slug: "/posts/i-made-a-font", dateCreated: "2023-01-10",  content: " Sample of the  Eau de poisson  font    A bit of context  A year ago, I bought a  Fontself  license with my company learning budget. Since then, I remember starting a project or two, but never completing them. Fast-forward to January 2023, I finally find a reason to have a typography pet project. Robin and I wanted to bring our shared blog  Eau de poisson  back to life. In our usual knowing-how-to-prioritize fashion, we decided to first reformat and redesign the entire website. And for that, we needed a new font. A font that was casual and laid back, but that would also support Latin-extended characters — mostly French, Vietnamese and other major Latin languages. The existing  Gaegu  typeface would have been perfect. However, it was mostly created for Korean and didn't support Latin-Extended characters. So, I decided to create my own typeface, shamelessly taking inspiration from Gaegu, but making sure to include all the characters I needed. For a first font creation, I'll give myself some leeway and say it's okay. The creative process  I don't really know if professionals could use Fontself. I somehow doubt it. Yet, for non-precise amateur work, it made the entire process super easy! 1. Draw your glyphs  You first need to install the Fontself plugin in Illustrator or Photoshop (I have the Illustrator plugin). From there, you can use Fontself's character grid template, or you can also create your own.   You then get into the heart of the process, either hand drawing your glyphs or using the pen tool for pixel perfect work. I drew mine with my graphic tablet, making sure to follow the guides.  1     2. Optimize your work  Once you completed your character set, make sure to duplicate it to optimize it.  Transform all strokes into shapes. ( Object > Expand appearance ) Merge all parts of one character into one group. ( Pathfinder > Unite ) For a better performance, I also simplified all the shapes, while preserving enough details. ( Right-click > Simplify... )  3. Upload your characters on Fontself  You can then upload all your characters into the plugin. You will get an overview like this.   In this first screen, you check that your characters are all on the same baseline, and that the ascender and descender are correctly positioned. You make these edits by dragging the characters and guides directly. 4. Save hours of work with one click  We now get to the cool part:   In only one click, Fontself will go through your entire character set and do the base spacing / kerning work automatically. The letters g, j, or i usually have a smaller left-margin? The plugin will calculate and preset that margin for you. Of course, the result is by far not perfect, but it does save you a lot of time! Once Fontself has done most of your work,... you still have a lot to do. Here comes the manual spacing and kerning part of the process. 5. Spacing and kerning, the manual way   What is the difference between spacing and kerning? When we talk about a font's spacing, we are referring to the generic amount of space between the characters. This sets the font rhythm. We refer to kerning to talk about the adjustment of space between two specific characters.  Fontself gives you control glyphs that define the rhythm of your font. Traditionally, we use  nnnnn  and  ooooo  for lowercase,  HHHHH  and  OOOOO  for uppercase. As you scroll down, you compare the usual  ooooo  rhythm with an extra glyph in the middle. You can then adjust the spacing around these glyphs to fit the rhythm you settled for. Make sure you spend most of your time on the spacing before getting to kerning. The process should then be fairly similar. The only difference being that you adjust a space between two specific glyphs only.      And that's it! It took me a day and a half to finish, even though I made the simplest font I could think of. I am really happy with the result, even though it doesn't produce the most legible body text — but who cares, I only use it for headings. Anyway. {end-key}  Footnotes    Pay extra attention to the top (ascender) and bottom (descender) guides, not going beyond them, as it might ruin the final font line-height.  ↩   "},{ id: 64, directory: "Musings", title: "In Bruges (2008)", subtitle: "A tiny note about the movie", slug: "/posts/in-bruges-2018", dateCreated: "2020-12-19",  content: "  Directed by Martin McDonagh  Super funky characters, really liked the movie! Rating: 3.5/5 {end-key}"},{ id: 65, directory: "Musings", title: "IndieWeb browsing", subtitle: "How I lost myself in the webring", slug: "/posts/indieweb-browsing", dateCreated: "2022-04-16",  content: " This weekend, I spent a lot of time jumping from site to site through the IndieWeb webring —  check my homepage footer to start browsing . I loved stumbling upon these little corners of the internet that people would customize to fit their own personality and preferences. A few of these websites were business-driven, but a lot of them were just quirky personal websites that you would never find through regular SEO. Below is an unordered list of newly discovered websites that I liked for one reason or another. Either because it was just  so different , or because it inspired me in some way.   maya.land   ohhelloana.blog   tommi.space   kandr3s.co   www.zinzy.website   prtksxna.com  {end-key} "},{ id: 66, directory: "Musings", title: "Interactive fiction", subtitle: "So I got into text adventure games", slug: "/posts/interactive-fiction", dateCreated: "2022-04-06",  content: " A bit more than a month ago, I rediscovered the very niche game subgenre of interactive fiction — or IF for the initiated. Interactive Fiction usually refers to text adventure games where you can dictate the actions of your character, similar to the Choose Your Own Adventure gamebooks.   I fell in love with this type of gaming which allows you to dive into completely different worlds through words alone. Some stories will be more narrative-based, while others will contain a lot of puzzles. But really, it's up to you and your tastes! If you would also like to dip your fingers (or full arms) into this gaming genre, here are some links I found useful as an IF beginner:   A Beginner's Guide to Playing Interactive Fiction  — clear and concise overview of IF  Interactive Fiction Database  — all the games are there!  Games for Beginners  — list of beginner-friendly games from an IFDB poll   If you are not convinced yet, here are a few game reviews and synopses that can give you an idea of the diversity of genres.  Anchorhead  (synopsis)   You take a deep breath of salty air as the first raindrops begin to spatter the pavement, and the swollen, slate-colored clouds that blanket the sky mutter ominous portents amongst themselves over the little coastal town of Anchorhead.  Travel to the haunted coastal town of Anchorhead, Massachusetts and uncover the roots of a horrific conspiracy inspired by the works of H. P. Lovecraft. Search through musty archives and tomes of esoteric lore; dodge hostile townsfolk; combat a generation-spanning evil that threatens your family and the entire world.   9:05  (a review)  You're woken up by the telephone, having slept a lot longer than you should have, and there'll be hell to pay if you don't get out of the house in a hurry. A very short game with a devious twist that the IF theorists out there will find interesting. Gains immeasurably on replay. — Carl Muckenhoupt   Photopia  (a review)  Scenes from a handful of ordinary lives alternate with chapters of a child's colorful science-fantasy. Sweet and sad, and complex enough that you may need to go through it twice in order to fully understand how all the fragments fit together. Very story-driven, with menu-based conversations and virtually no puzzle content. [...] — Carl Muckenhoupt  {end-key}"},{ id: 67, directory: "Musings", title: "Just the two of us", subtitle: "A tiny note about life stuff", slug: "/posts/just-the-two-of-us", dateCreated: "2023-05-29",  content: " Robin and I just realized that this is the first time in a month that we are back to being just the two of us. It feels great to spend long stretches of time with friends and create a new sort of routine together. However, the rhythm we are used to is also completely disrupted. We don't have as much time to talk or share the little things we usually do. After some time, it feels right to get back to our own little world. {end-key}"},{ id: 68, directory: "Musings", title: "Last homage to my old website", subtitle: "A toast to all lost websites", slug: "/posts/last-homage-to-my-old-website", dateCreated: "2022-04-03",  content: " I realized that I change my website more often than I go shopping. I can't decide if it's a good or bad thing. Even if he says otherwise, that's because of  Robin . He recently came up with  brut, his own static site generator,  as an answer to existing generators lack of simplicity and stability. For a noob coder like me, trying to keep up with Next.js constant breaking changes was draining. It made my personal website feel like a chore to maintain. So as soon as the alpha version of brut was out, I proceeded to remake my entire website based on it. I took this opportunity to simplify the UI and focus on accessibility. For my blog posts, I also moved away from my customized MDX files towards simpler Markdown files. This was to make sure any future transitions I wanted to make  1   would be as seamless as possible. Even though, my last website was terribly non-accessible, I did like it. So as a tribute to it, before it forever disappears in the that-was-my-website graveyard of the World Wide Web, I toast to it one last time:            Footnotes    As we all know this new website will only last until the next redesign craze comes around.  ↩   "},{ id: 69, directory: "Musings", title: "Peeking into the IndieWeb world", subtitle: "My first baby steps", slug: "/posts/my-first-indieweb-steps", dateCreated: "2022-04-10",  content: " Robin introduced me to the IndieWeb about a year ago. And as it usually happens, I dived head first into it before he even had time to close his tabs. I remember struggling quite a bit with the concepts, as the documentation is a bit messy and the standards are overall loosely applied by the community. All in all, I wasn't so successful in getting any closer to the IndieWeb grail. But what is the IndieWeb to begin with?  Chris Aldrich explains it way better than I could in his  introduction to the IndieWeb . But to keep things short, it is a movement that is built on three main principles:  Ownership of your content — No vendor lock-in as you publish and own your content. Author centric — You can be connected to other services and platforms, but everything comes back to your site. Control of your content and format — You completely control how your website looks and works.  After my recent  website redesign , I decided to have another go at this IndieWeb stuff. And I think I was a bit more successful this time around. IndieWeb baby steps...   ...a guide by someone who didn't understand half of what she did.  If you are completely new to this world, I absolutely recommend reading the  introduction to the IndieWeb  mentioned before. It really helped me get the why the IndieWeb part pinned down. On the side, it is also always good to have a look at the official  IndieWeb Getting Started page . Most documentation is available there. So what about some concrete steps, now?  Domain name — First of all, you will need to own your domain name to set up most of the IndieWeb stuff we'll talk about.  Wiki: how to get your domain name.  Web sign in — Something that is pretty common in IndieWeb circles is to enable login to services with your domain name alone.  Wiki: how to set up web sign in.  Join the IndieWeb webring — Once you have a web sign in setup, you are ready to join the  IndieWeb webring ! Or at least, I would highly recommend browsing through the ring — start at the  footer of my homepage  — to see how all of these standards are applied by real websites. Mark up your content — Use  microformats  to mark up your content and enable a ton of IndieWeb features in the future. I started with an   h-card   (or profile card) on my homepage, as well adding microformats to my blog posts. Enabling webmentions — After months, Rowan Manning's article   Webmentions for your Static Site   finally clarified for me what webmentions were all about. I went for the simplest setup, using  webmention.io  to receive webmentions and  Telegraph  to manually send them.  If you are still confused after reading the many wikis and articles, feel free to have a look at my homepage markup. It definitely helped me to compare the documentation to some real life examples. Another good link to have open on the side at all times is  IndieWebify.me  that will give you a few starting tips and allow you to test some of your setups. Now that everything  seems  to be technically viable, the question that remains is: How do I really want to use these tools? {end-key}"},{ id: 70, directory: "Musings", title: "My quest for true small caps", subtitle: "A web design adventure", slug: "/posts/my-quest-for-true-small-caps", dateCreated: "2021-01-07",  content: "  Note: this article is based on the previous design of my website.   When I launched my website with Next.js, I naturally veered towards a minimal typography-based design. The idea seemed simple to implement, yet elegant. Nothing could go wrong. My first version looked fine: white on dark gray, big bold headings and a more discreet sans serif body. But I wanted to add some  spice . This was going to be my personal website, and it needed to have something special. How about changing links to something else? Maybe small caps? And there began my quest to find true small caps on the web. Challenge #1: Finding typefaces with small caps  It is hard finding an open source typeface that actually supports small caps. This actually made things easier for me, as I was already planning to spend hours trying out different typefaces. I knew I wanted a simple sans serif typeface to complement the quirky Vesper Libre of my headings. Only one sans serif font was open source with small caps support: Source Sans Pro. My choice was made. Challenge #2: Are those really true small caps?  Mmh... are those true small caps or fake ones? Honestly, I wasn't sure. Red flag: the supposedly small caps did look thinner than the regular letters.    Before: fake small caps     After: true small caps  What saved me was this (amazing looking)  blog post on small caps . I didn't go through everything explained there, but I learned that small caps actually have Unicode characters. I could now have  actual true small caps  next to my maybe true small caps and see if they differed. They did.  Shoot . Challenge #3: Setting up true small caps  I dived into the small caps rabbit hole again and realized that small caps — and other special glyphs — were only supported by the OpenType format (OTF). That explained a lot as I was using the WOFF and WOFF2 formats so far. I was hesitant to switch to an OTF font as the WOFF format is made for the web and loads  much  faster. I settled on an in-between decision by keeping both of them. My main font would remain in WOFF / WOFF2, while I would load the small caps from the OTF file. I basically treated my main font and small caps as two distinct font families. For reference, here is the  @font-face  code I used for my small caps characters:  @font-face {   font-family: SourceSansProSmallCaps;   src: url(/fonts/SourceSansPro-Regular.otf) format(opentype);   font-weight: 400;   font-style: normal;   font-display: swap; /* Avoids blank space while loading  */   unicode-range: U+000-2AF; /* Download small caps ranges only */ }    Let's dive into the details:   font-display: swap  means that the block time is short for an infinite swap period. Your computer might take a tiny moment to load your small caps. This line of code avoids having a blank space where your small caps should be during that time.  unicode-range: U+000-2AF  loads the IPA Extensions Unicode range only — that's where the small caps are.  And that's how I managed to set up true small caps on my website. Anyway. {end-key}"},{ id: 71, directory: "Musings", title: "Now archive #1", subtitle: "A snapshot in time", slug: "/posts/now-archive-1", dateCreated: "2021-08-21",  content: " Now in Berlin for more than two years with my partner  Robin . Summer here is great: I am bouldering a lot, drinking orange wine on sunny terraces, trying not to kill the kids screaming in the backyard (not mine — neither the kids nor the backyard),  baking bread  and making half-baked resolutions. But I spend most of my free time typing and clicking away on my laptop, working full-time as a graphic designer. {end-key}"},{ id: 72, directory: "Musings", title: "Now archive #2", subtitle: "A snapshot in time", slug: "/posts/now-archive-2", dateCreated: "2021-10-24",  content: " Now in Berlin for more than two years with my partner  Robin . I am actively working on not being blown away by the wind and rain, and getting used to waking up in the dark. I boulder two or three times a week with friends — usually at 7am before going for pastries together. I am currently nurturing my very new  digital garden . For once, I am trying to  not over-complicate the process  to make the habit easier to build. As usual, I am reading sci-fi books. I finally started the  Dune series  after watching the recent movie by Villeneuve.  I needed to know what happened next!  I planted an avocado seed a few months back and finally gave it a name (Sigmund). It is now 55 cm tall and not dying — yet. Otherwise, I spend most of my free time typing and clicking away on my laptop, working full-time as a visual designer. {end-key}"},{ id: 73, directory: "Musings", title: "Now archive #3", subtitle: "A snapshot in time", slug: "/posts/now-archive-3", dateCreated: "2022-02-20",  content: " I have now been in Berlin for more than three years — kinda crazy. My routine mostly consists of bouldering a few times a week, meeting with friends for morning pastries and working as a visual designer. A few current side-projects:  Nurturing my planner shop   Life Under Control   Building my website from scratch — yes,  again  — with  brut  {end-key} "},{ id: 74, directory: "Musings", title: "Now archive #4", subtitle: "A snapshot in time", slug: "/posts/now-archive-4", dateCreated: "2022-06-25",  content: " I am based in Berlin since 3+ years and working as a UX designer. Outside of work, you'll find me bouldering and sharing morning pastries with friends. A few side-projects:  Nurturing my planner shop   Life Under Control  , aiming for one new listing every month Reading  100 Things Every Designer Needs to Know About People  by Susan M. Weinschenk and updating  my UX learnings notes  accordingly Going through a  wine tasting project  with Robin consisting of trying 73 different wines. Current progress: 2/73 Participating in my first D&#x26;D campaign. My character is a dragonborn who is on a personal cooking quest. {end-key} "},{ id: 75, directory: "Musings", title: "Now archive #5", subtitle: "A snapshot in time", slug: "/posts/now-archive-5", dateCreated: "2022-11-04",  content: " I am based in Berlin since 3+ years and working as a UX designer. (And if everything goes to plan, soon moving to the countryside in France). Outside of work, you'll find me bouldering  1   religiously and sharing morning pastries with friends in any kind of weather. A few side-projects:  Going through a  wine tasting project  with Robin consisting of trying 73 different wines. Learning about UX and updating  my UX living notes  accordingly Nurturing my planner shop   Life Under Control   Participating in my first D&#x26;D campaign. My character is a dragonborn who is on a personal cooking quest. Working on my own Figma wireframing library — because why not. {end-key}   Footnotes    My home gym is BoulderKlub :)  ↩   "},{ id: 76, directory: "Musings", title: "Now archive #6", subtitle: "A snapshot in time", slug: "/posts/now-archive-6", dateCreated: "2023-02-06",  content: " Still trying to find my marks in this new countryside life. I spend my free time:  Tending to the wood stove fire to heat the main living room Exploring the house Unpacking our boxes Planning and cooking meals Visiting neighbors and Robin's grandma for coffee (or wine) Reading graphic novels Thinking about post ideas for  Eau de poisson  {end-key} "},{ id: 77, directory: "Musings", title: "Now archive #7", subtitle: "A snapshot in time", slug: "/posts/now-archive-7", dateCreated: "2023-05-17",  content: " I am just starting my  early retirement parenthesis  — a.k.a, I am unemployed for the first time in four years. I do have a few productive plans for these next few months, but to be honest, I mostly want to slow down a bit and enjoy life differently. Currently, I spend most of my days:  reading sci-fi novels tending to the garden blogging here and  there  climbing outdoors and at my local gym working on a private Pocket alternative with Robin exploring new recipes doing small creative projects learning Typescript drinking wine with France {end-key} "},{ id: 78, directory: "Musings", title: "Now archive #8", subtitle: "A snapshot in time", slug: "/posts/now-archive-8", dateCreated: "2023-07-12",  content: " I am currently  slow living  in Ardèche and spending my days following my heart desires. Top of my mind at the moment:  Climbing, climbing, climbing (and learning new cool outdoor techniques) Working on a private Pocket alternative with Robin Learning about plants Reading Tending to my website Gardening and preserving food Planning our big 2024 trip {end-key} "},{ id: 79, directory: "Musings", title: "Now archive #9", subtitle: "A snapshot in time", slug: "/posts/now-archive-9", dateCreated: "2023-11-17",  content: " I am probably eating roasted chestnuts (with salted butter) and sitting by the fire to not freeze to death. It just hit Robin and I, that our year-long trip will start in about a month. So, at the moment, we are pretty much spending our days:  planning the last logistics bits (or more like chunks) of The Big Trip (trying to) complete all the French admin that we have been dragging for months making sure all our creative digital projects are in a good state (as they'll probably go dormant for a few months) climbing whenever the weather is clear enough prepping our old travel blog  Eau de poisson  to be usable on the go drinking lots of the wine bottles we have been hoarding and that are not meant to age being half anxious, half excited, and half cold (wait, that doesn't sound right) {end-key} "},{ id: 80, directory: "Musings", title: "Seven Psychopaths (2012)", subtitle: "A tiny note about the movie", slug: "/posts/seven-psychopaths-2012", dateCreated: "2020-11-26",  content: "  Directed by Martin McDonagh  Completely crazy and over the top characters, dialogues are amazing. Loved how they mixed the screenplay and reality. Rating: 3.5/5 {end-key}"},{ id: 81, directory: "Musings", title: "Sicilian notes", subtitle: "A tiny note about our trip in Sicily", slug: "/posts/sicilian-notes", dateCreated: "2023-06-02",  content: " A list of things noticed during our 2-week trip in Sicily:  People saying bombastic Traffic rules seem to simply be suggestions (especially speed limits) Savory or sweet, everything is fried  Arancina  or  arancino , that is the question  1    Pani câ meusa  (or spleen sandwich) is surprisingly good Breakfast (when there is one) should be sweet Bread is yellow, not sure why  Cassata  might be one of the rare layered cake I like Recycling is going strong (or theoretically at least with all these different bins) So many stray cats A few stray dogs A carb with carb with carb sandwich  2   Clear turquoise water that are postcard-worthy The landscape looks similar to some parts of the Ardèche region in France (but as a drier version most of the time) Some people wearing winter jackets even though it's more than 25 °C A mix of overpopulated cities and lands only inhabited by cows and sheeps Olive trees Orange trees Lemon trees Fresh lemons that you can practically eat like apples Empty streets from 1–3pm, except for the occasional tourist A very strong pride in being Sicilian No one cares about the Italian national day Churches 24 hours vending machine clusters everywhere Lots of precise hand gestures Much more tourism than what I would have expected Very cool pebble beaches (sand is overrated) Toilets that flush with a manual tap system Terrible napkins that feel like paper Natural reserves where you can swim just like at a regular beach  3   A very poor public transportation system — the car is king An obsession with the Moor's heads  4   as: ceramic vases, book holders, tablecloth decoration, bed sheet pattern, pillow cases, magnets, mugs, ... Many archeological monuments (a.k.a. lots of protected old rocks that are expensive to see) Ficuses, my new favorite type of tree!  5   {end-key}   Footnotes    There is apparently a strong debate between West ( arancina  team) and East ( arancino  team) Sicily on how to call these delicious rice balls.  ↩   Thinking of the  panelle e crocchè  sandwich that is made with bread, chickpea fritters and fried potato balls  ↩   This feels weird to me, aren't natural reserves usually supposed to be protected from any kind of human impact?  ↩   The Moor's heads are symbolic of Sicily. The story is about an impossible love between a Moor young woman and a man. Spoiler: She cuts his head and makes a vase with it when she learns that he is married with kids. #romantic #truelove  ↩   The coolest ficus tree is in the Palermo botanical garden.  ↩   "},{ id: 82, directory: "Musings", title: "Sick days cheatsheet", subtitle: "Or how to help my tired cloudy mind", slug: "/posts/sick-days-cheatsheet", dateCreated: "2022-11-30",  content: " I am not sure what is happening to me, but since about a year, I get sick much more severely than before. A common cold can lead to exhaustion, strong headaches and nausea. As it is happening more and more often, here is a list of the things I personally do to help me feel better (in no particular order). For the immune system   Sleep, sleep, sleep Take some Vitamin C (e.g. oranges and broccoli) Drink herbal teas with honey and ginger Eat some rice porridge with tumeric  1   To test: defrost and drink an  immunity shot   For a sore throat   Gargle salt water a few times a day Eat a teaspoon of honey Drink a few drops of propolis  For a congested nose   Take a hot shower Drink a lot of water throughout the day Do a steam bath  2   Rinse nostrils with a neti pot a few times a day  3    For nausea   Controlled breathing  4   Eat bland foods, avoid spicy or fatty foods To test: wrist acupressure to  point Neiguan (P-6)  To test: candied ginger or fennel tea  To prevent sinusitis  I have a sensitive nose that easily transforms an inoffensive cold into full-fledged sinusitis. Here are ways to prevent it:  Rinse nostrils with a neti pot a few times a day  3   Drink lots of water to keep the mucus thin Take a hot shower and do a steam bath  2   Avoid dry environments Sleep with your head elevated Blow your nose gently, one nostril at a time. Forceful blowing can irritate the nasal passages and propel bacteria-laden mucus back up into your sinuses. {end-key}   Footnotes    Always make sure to store some porridge in the freezer!  ↩   Essential oils: 1 drop lavender, 1 drop peppermint  ↩   ↩ 2    Preboil 3/4 cup water and let it cool down, mix in 1 tsp salt and 1/2 tsp baking soda. You can keep this solution for a week.  ↩   ↩ 2    Inhale through nose to a count of three, hold breath to a count of three, then exhale to a count of three  ↩   "},{ id: 83, directory: "Musings", title: "Split (2016)", subtitle: "A tiny note about the movie", slug: "/posts/split-2016", dateCreated: "2020-12-21",  content: "  Directed by M. Night Shyamalan  Really nice tension building all throughout the movie, amazing acting. Rating: 4/5 {end-key}"},{ id: 84, directory: "Musings", title: "Styling animated rainbow links", subtitle: "With CSS only", slug: "/posts/styling-animated-rainbow-links", dateCreated: "2022-04-15",  content: " This new website layout is quite minimalist and mostly plays on typography styles and hierarchy. I wanted to add some  spice  to such a simple base — so what better way than doing it rainbow-style on hyperlinks?  1   The general idea is that my hyperlinks are underlined by default  2   as they are the same color as the body text. When hovered of focused, they are animated with CSS keyframes that seamlessly transition through a list of colors. To differentiate the  focus  from  hover  state, I removed the underline and added an  outline  around the link instead  3  . As you noticed, I am not a very talented web dev blogger, so I'll just drop my CSS below.  a {   color: white;   text-underline-position: under; }  a:hover {   animation: color-change 5s infinite; }  a:focus {   animation: color-change 5s infinite;   text-decoration: none;   outline-offset: 0.1em;   outline: 0.1em solid; }  @keyframes color-change {   5% {     color: hsl(0, 100%, 30%);   }   15% {     color: hsl(41, 100%, 33%);   }   30% {     color: hsl(140, 100%, 27%);   }   50% {     color: hsl(230, 100%, 50%);   }   70% {     color: hsl(261, 100%, 55%);   }   90% {     color: hsl(320, 100%, 31%);   } }    {end-key}  Footnotes    Please don't answer to that, this is a rhetorical question.  ↩   Accessibility note: If your links are not underlined, you have to make sure  they contrast enough with your body text .  ↩   Just tab through the links to test it out!  ↩   "},{ id: 85, directory: "Musings", title: "The Good, the Bad and the Ugly (1966)", subtitle: "A tiny note about the movie", slug: "/posts/the-good-the-bad-and-the-ugly-1966", dateCreated: "2020-12-05",  content: "  Directed by Sergio Leone  It might not have helped that we watched the 3 hours version, but I was falling asleep by the end — it was very late. Amazing silences, staging and soundtrack otherwise. Rating: 3.5/5 {end-key}"},{ id: 86, directory: "Musings", title: "The Guernsey Literary and Potato Peel Pie Society", subtitle: "A review of the audiobook", slug: "/posts/the-guernsey-literary-and-potato-peel-pie-society", dateCreated: "2022-04-09",  content: "  January 1946: London is emerging from the shadow of the Second World War, and writer Juliet Ashton is looking for her next book subject. Who could imagine that she would find it in a letter from a man she's never met, a native of the island of Guernsey, who has come across her name written inside a book by Charles Lamb... As Juliet and her new correspondent exchange letters, Juliet is drawn into the world of this man and his friends — and what a wonderfully eccentric world it is. The Guernsey Literary and Potato Peel Pie Society — born as a spur-of-the-moment alibi when its members were discovered breaking curfew by the Germans occupying their island — boasts a charming, funny, deeply human cast of characters, from pig farmers to phrenologists, literature lovers all. Juliet begins a remarkable correspondence with the society's members, learning about their island, their taste in books, and the impact the recent German occupation has had on their lives.  I got quite sick this last week and wasn't able to do much except for lying in bed. It happens that this exciting activity pairs pretty well with audiobooks. So I set on searching for an available audiobook on  Libby   1   and stumbled on  The Guernsey Literary and Potato Peel Pie Society  by Mary Ann Shaffer and Annie Barrows. I had already read the book a few years ago and remembered loving it. As I was practically in a dying state, it seemed like it would have just the kind of wholesome vibe I needed. Plus, I was curious about how an audio version of this epistolary novel would fare. So I closed my eyes and pressed on  play  (or more likely the other way around). And what an adventure! I wholeheartedly recommend listening to the  Penguin Random House edition of the audiobook . The voice actors are able to give life to the many characters in such a natural way, it is hard to believe that it is only a five-people cast. There is something about this novel and these characters that just made me want to start over, as soon as I had finished it. Listening to it feels like sitting in a cozy armchair with a cat on your laps, drinking hot chocolate out of your favorite heavy ceramic mug with jazz music in the background.  2   In short,  The Guernsey Literary and Potato Peel Pie Society  made me feel like I was part of a quirky, loving and caring community. And isn't that what everyone is looking for in the end? {end-key}  Footnotes    Libby is a great library app, if you have the chance to try it!  ↩   Is this example too  me -specific? This must be a universal feeling, no?  ↩   "},{ id: 87, directory: "Musings", title: "The Host (2006)", subtitle: "A tiny note about the movie", slug: "/posts/the-host-2006", dateCreated: "2020-12-12",  content: "  Directed by Bong Joon-ho  I had high-ish expectations and was finally a bit underwhelmed. Too over the top and kitsch for my tastes. Overall entertaining movie. Rating: 2.5/5 {end-key}"},{ id: 88, directory: "Musings", title: "The Untouchables (1987)", subtitle: "A tiny note about the movie", slug: "/posts/the-untouchables-1987", dateCreated: "2020-12-11",  content: "  Directed by Brian De Palma  Astonishing photography. Very conservative in its depiction of women and morals, but was it represented like that on purpose? Hard to tell in retrospect. Rating: 4/5 {end-key}"},{ id: 89, directory: "Musings", title: "Three years blogging", subtitle: "And apparently still going", slug: "/posts/three-years-blogging", dateCreated: "2023-11-26",  content: " As I was browsing (a.k.a. overengineering) my website again, I realized that last Friday marked the three-year anniversary of my blogging adventure. On November 24, 2020  1  , I wrote my first update:  #3 Friends and digital gardening . I remember that I decided to start this website during the COVID-19 lockdown. At that time, the concept of  Blogging on The Internet  sounded very daunting to me, but I felt like I wanted to do something in public, something that kept me accountable. To build courage, and mostly to get started, I wrote two weekly updates that were just a few lines long. I didn't publish them. On the third week, I finally felt brave enough and published my first real blog post. It has now been three years, and I am proud of having kept this routine for so long. My individual posts are not that good, I am definitely not the best writer out there. Yet, I feel like I was able to create a space that is really mine. Here, I feel safe to experiment, code, document, have fun, overthink, share truths-because-its-on-the-internet, doodle, learn, create, write posts that don't fini  2   {end-key}  Footnotes    For the eagle eyes out there: there are a few blog posts preceding that one, but I had published those on Medium, before deciding on bringing them back to my own website later on.  ↩   Nah, just kidding. I won't be a meanie and leave you hanging. Long story short: happy anniversary to this website, yay!  ↩   "},{ id: 90, directory: "Musings", title: "UI/UX Learnings", subtitle: "Living notes", slug: "/posts/ui-ux-learnings", dateCreated: "2022-07-13",  content: " I started using Notion a bit more to keep living notes — posts that I want to access and update easily through time. Here are the ones I am keeping on my UI/UX learning journey.   UI Pattern Library   UX Theory  {end-key} "},{ id: 91, directory: "Musings", title: "Victoria (2015)", subtitle: "A tiny note about the movie", slug: "/posts/victoria-2015", dateCreated: "2020-11-26",  content: "  Directed by Sebastian Schipper  Tensed film, irritating main character. The amateur camera style literally gave me a headache. Rating: 3.5/5 {end-key}"},{ id: 92, directory: "Musings", title: "Waking up in Berlin", subtitle: "A last note", slug: "/posts/waking-up-in-berlin", dateCreated: "2023-01-29",  content: " 6.20 AM The last days were spent sorting four years of stuff, painting walls, and deep cleaning our flat. Between chores, we said goodbye to places and friends. These few morning hours (or minutes) awake will probably be the last we spend in Berlin as residents. So many memories, so many feelings — what a grand adventure it was! Next time I write will be from France. See you there! {end-key}"},{ id: 93, directory: "Musings", title: "Wonder Woman 1984 (2020)", subtitle: "A tiny note about the movie", slug: "/posts/wonder-woman-1984-2020", dateCreated: "2020-12-26",  content: "  Directed by Patty Jenkins  Needed to close my eyes from time to time because it was just too cheesy to witness. Entertaining anyway. Rating: 2.5/5 {end-key}"},{ id: 94, directory: "Musings", title: "Words I'm growing suspicious of", subtitle: "Ranting (in a list format)", slug: "/posts/words-i-m-growing-suspicious-of", dateCreated: "2023-12-12",  content: " The other day, I stumbled upon the website of company XYZ. And once again, I got frustrated at their use of superlatives. So, I started a list of words. Some words were included because they are too big, overused, overselling. Others, because they are often selling a behavior that I think is just a bit... off. All in all, it's just a list of terms that I grew suspicious of for various reasons.  streamlining seamless AI personal branding mission pioneering SEO engagement / impressions smart [insert object name] terms and conditions productivity empowering self development environmental friendly zero waste carbon neutral / offsets sustainable growth  (This list is still WIP.) Anyway. {end-key}"},{ id: 95, directory: "Learning", title: "Stars or planets?", slug: "/learning/astro-1-stars-or-planets", dateCreated: "2023-11-15",  content: " Here are three quick ways to differentiate a planet from a star. 1. How bright is it?  Sometimes, the object is just too bright to be a star. Venus is the brightest planet. It usually appears close to the morning sun or sunset. 2. Does it scintillate?  Let's be clear, neither planets or stars really twinkle. The difference in scintillation will be explained by how big the objects are seen from Earth. Stars appear as dots, whereas planets appear as discs. (Because stars are much farther than planets, they also appear smaller.) The light coming from a star will be refracted in many directions when going through our turbulent atmosphere. It will make it look like it is scintillating. On the contrary, the light coming from a planet and going through our atmosphere will seem more stable as its extra surface will allow it to compensate the variations. 3. Where is it positioned?  All planets follow the  ecliptic plane  as all planets rotate around the sun. As Earth is also rotating around the sun, it will seem like all planets (and the moon) are following the imaginary trajectory traced by the sun."},{ id: 96, directory: "Learning", title: "Colloids: From mayonnaise to hairspray", slug: "/learning/cooking-1-colloids-from-mayonnaise-to-hairspray", dateCreated: "2023-09-30",  content: " A colloid is a mixture where dispersed insoluble particles is suspended throughout another substance. This two-phase mixture has a dispersed phase (the suspended particles) and a continuous phase (the medium of suspension). An emulsion is a common colloid, where two or more liquids that are normally unmixable or unblendable (e.g. oil and water) are merged into a stabilized and homogeneous mixture. Mayonnaise is an emulsion of oil dispersed into the water (a.k.a. the egg). Other colloid examples:  liquid dispersed in gas (liquid aerosol): fog, hairspray solid dispersed in gas (solid aerosol): smoke, ice cloud gas dispersed in liquid (foam): whipped cream liquid dispersed in liquid (emulsion): milk, mayonnaise, butter, hand cream, latex solid dispersed in liquid (sol): pigmented ink, sediment gas dispersed in solid (solid foam): aerogel, styrofoam liquid dispersed in solid (gel): agar, gelatin solid dispersed in solid (solid sol): cranberry glass "},{ id: 97, directory: "Learning", title: "Food pigments: Keep green food green", slug: "/learning/cooking-2-food-pigments-keep-green-food-green", dateCreated: "2023-09-30",  content: " A group of molecules called chlorophyll is at the basis of photosynthesis. It enables plants to absorb light to power the plant. It also has a side effect: it gives the plant its green color. Chlorophyll molecules aren't very stable molecules. They can break down or change configuration quite easily, which can change their color. Acidity, heat, and damage can also impact chlorophyll molecules, and thus the color of your green food. Blanching brightens your veggies  The chlorophyll is present  inside  the plant, in its cells. When blanching, the gases trapped between the tissue's cell escape, the cell walls collapse and the chlorophyll becomes more visible. Heat also plays a second role: it destroys a plant enzyme called chlorophyllase, which is responsible for breaking down the chlorophyll pigment into a brown color. However, if vegetables are cooked too long, their bright green transforms into a dull olive gray. The prolonged heating causes the cells to release acids. These push out the magnesium atom in the center of the chlorophyll molecule, causing it to turn dull green. When cooking, we should then avoid adding acids (e.g. lemon, vinegar) to blanching water. On the opposite, adding a pinch of baking soda can help retain some of the bright green color. Alkalines will indeed counteract the plant acids dulling effect. Note: This is also the reason why lactic fermentation of green veggies will make their color duller."},{ id: 98, directory: "Learning", title: "Food pigments: Purple, blue or red blueberries", slug: "/learning/cooking-3-food-pigments-purple-blue-or-red-blueberries", dateCreated: "2023-09-30",  content: " Anthocyanins give their purple color to fruits and vegetables such as blueberries, pomegrenates, grapes or purple beans. When cooking (heating), the anthocyanin pigment leaches out. Heat also destroys the pigment: as the purple color disappears, the green chlorophyll becomes more visible. When the pH is high enough, as when making blueberry pie with lemon juice, the anthocyanins stabilyze into a red hue. When the pH is more alkaline, the hue will tend to be on the blue side. Therefore, a dash of lemon juice or a pinch of baking soda can greatly alter the pigment of your dish!"},{ id: 99, directory: "Learning", title: "Food pigments: Red beets", slug: "/learning/cooking-4-food-pigments-red-beets", dateCreated: "2023-09-30",  content: " Betalain is the red pigment of your beet, amaranth, rhubarb or Swiss chard. These pigments can be preserved with some vitamin C (e.g. lemon juice), which will protect them from enzymes that cause them to turn brown."},{ id: 100, directory: "Learning", title: "How to cut open a pomegranate", slug: "/learning/cooking-5-how-to-cut-open-a-pomegranate", dateCreated: "2023-10-15",  content: "  Cut through the hat of a pomegranate (at about a quarter from its tip), making sure not to cut through the seeds. Empty the had from its seeds and set aside in a bowl. Get back to the fruit body and cut through its skin vertically, just like when cutting an apple into quarters. The fruit should open up when you gently press on it. Remove the membranes that are easily detachable. Turn the fruit on its head (above the bowl) and hit on its back until all the seeds have fallen off. "},{ id: 101, directory: "Learning", title: "Data viz: Data visualization handbook", slug: "/learning/data-viz-1-data-visualization-handbook", dateCreated: "2022-05-05",  content: " It's been a while  Data visualization handbook  by Juuso Koponen and Jonatan Hildén has been waiting on my book shelf. I started to go through it about a week ago and here are a few quotes that stayed with me. Notes before we start:  The page numbers I use are from the first edition (2020). Some quotes might only be understandable within context. "},{ id: 102, directory: "Learning", title: "Data viz: Introduction", slug: "/learning/data-viz-2-introduction", dateCreated: "2022-05-05",  content: " Text v. data visualization   When should data be visualized, and when is text enough? In short, if something can be expressed just as clearly or even more clearly in words, visualization is unnecessary — sometimes even couterproductive. (p. 29)  Golden rule   [There is] one rule that an information designer should always follow: choose the clearest presentation method available. Disregard any other rules and guidelines we present in this book, when they are in conflict with this golden rule of information design. (p. 43)  Micro/macro reading   At its best, a visualization gives an overview of the topic quickly, but also rewards the reader who spends time exploring it in depth. (p. 43) "},{ id: 103, directory: "Learning", title: "Data viz: General principles of visualization design", slug: "/learning/data-viz-3-general-principles-of-visualization-design", dateCreated: "2022-05-05",  content: " Focal points guide the eyes   With the use of focal points, the designer can introduce linear features into an open narrative structure. If it is important that specific elements be read before others, these elements should be highlighted visually so that the reader's gaze is drawn to them first. Highlighting key details is also justified by the fact that it makes it easy for the reader to understand which of the features in the graphic are the most essential. (p. 90)  Note: For instance, a big central image can act as a focal point. Integrate text and graphics   In an eye-tracking study conducted at Lund University, the subjects were presented with newspaper spreads that had identical content, with the text and graphics laid out in different ways. [...] The study also looked at an alternative,  integrated  layout, in which the graphic elements were placed withing the text column, so that the teext and graphics that are related in content were located close to each other. When a page was laid out like this, readers read both the body text and the related graphics. This solution places less cognitive strain on the readers, because they do not have to form mnemonic links between the text and the separate graphics. (p. 90)  Nominal v. ordinal scales   A nominal scale divides items into classes, the only relevant characteristic of which is that they are separate. They have no particular order or distance on the scale. Examples of such classes include nationality, profession, membership in a certain group, and so on. [...] Values on a nominal scale cannot be mathematically compared. [...] An ordinal scale differs from the nominal scale in that the classes have a rank order: the first to tenth income decile, the Likert scale (strongly agree, agree, etc.), a ranking in comparison or competition, a military rank, and so on. It is possible to determine the median value for items on an ordinal scale, and to calculate the difference between two rankings, but because the scale does not say anything about the  magnitude  of differences between classes, most other calculations are not meaningful. (p. 94) "},{ id: 104, directory: "Learning", title: "Géologie: Les familles de roches", slug: "/learning/geologie-1-les-familles-de-roches", dateCreated: "2023-12-09",  content: " Les roches peuvent être classées en trois grandes familles : les roches sédimentaires, magmatiques et métamorphiques. Donner un nom à une roche, c'est reconstituer les conditions de sa formation. Roches sédimentaires  Les roches  sédimentaires  se forment à température et pression ambiantes. Dire qu'une roche est sédimentaire signifie qu'elle s'est formée à la surface de la Terre, parfois à l'air libre (ex : dunes de sable), mais le plus souvent sous l'eau (vase dans un lac, une mer), et qu'elle s'est solidifiée par la suite. Elles proviennent de l'accumulation de sédiments (éléments solides) qui se déposent le plus souvent en couches ou lits superposés, appelés strates. On rassemble sous le nom de diagenèse l'ensemble des processus par lesquels les dépôts issus de l'érosion sont transformés en roches sédimentaires. Roches magmatiques  Les roches  magmatiques  (températures très élevées) se divisent en deux groupes :  les roches volcaniques, issues de la solidification d'un magma qui a atteint la surface de la Terre (par l'intermédiaire d'une éruption volcanique). les roches plutoniques (Pluton était le dieu du monde souterrain) lorsque le magma qui ne pouvait monter jusqu'à la surface, a cristallisé en profondeur.  Reconnaître quelque part une roche  volcanique  signifie qu'il y a eu une éruption volcanique autrefois à cet l'endroit ; les magmas issus de ce volcan se sont solidifiés en refroidissant rapidement. Dire qu'une roche est  plutonique  signifie qu'elle provient d'un magma qui a cristallisé en profondeur, ou si l'on préfère, que l'endroit où se trouve cette roche était situé lors de sa formation à 10 ou 20 km sous terre. Roches métamorphiques  Les roches  métamorphiques  sont des roches formées à une forte profondeur, là où pression et température sont élevées. Dire qu'une roche est métamorphique signifie qu'en profondeur (entre 10 et 50 km en général), une roche (de n'importe quel type) a été transformée (sans fondre) par des fortes pressions et températures."},{ id: 105, directory: "Learning", title: "Géologie : Les roches sédimentaires", slug: "/learning/geologie-2-les-roches-sedimentaires", dateCreated: "2023-12-09",  content: " Déterminer une roche sédimentaire  Les roches sédimentaires constituent une famille d’une grande diversité d’aspect. Les plus courantes sont les grès, les calcaires ou les charbons. Elles peuvent être meubles (sable) ou consolidées (grès). Sur le terrain, elles sont (pour la plupart) disposées en couches horizontales qui ont pu être par la suite plissées ou fracturées. Au niveau de l’échantillon les couches sont rarement visibles. Il faut donc faire appel à d’autres critères de détermination ; ce sont dans l’ordre d’importance, la présence de fossiles (morceaux de coquilles…), de minéraux particuliers comme la calcite (effervescence à l’acide) ou le sel, de grains (souvent du quartz) ou de morceaux de roches pris dans un ciment. La couleur des roches sédimentaires varie beaucoup, du gris au rouge et du blanc au jaune ; même si la couleur des roches n’est jamais un élément neutre, elle n’est pas un critère de détermination. Formation des roches sédimentaires  Les roches sédimentaires se forment à la surface de la Terre ou dans l’eau. Les roches sédimentaires proviennent de la destruction d’autres roches ; elles incluent souvent des traces de vie. Les morceaux (par exemple des galets) de ces autres roches ou les éléments chimiques qui les composaient sont transportés par les rivières, les glaciers ou le vent plus ou moins loin de leur origine jusqu’à leur milieu de dépôt. D’ordinaire après leur dépôt, les roches sédimentaires ont été consolidées (calcaire, grès) par un processus que l’on appelle diagenèse et qui consiste en un tassement suivi d’un durcissement. Cependant des roches meubles (sables) ou même liquides (pétrole) sont des roches sédimentaires."},{ id: 106, directory: "Learning", title: "Grande-voie: Le matériel", slug: "/learning/grande-voie-1-le-materiel", dateCreated: "2023-10-01",  content: " En juillet dernier, Robin et moi avons fait  notre première grande-voie . Cet après-midi, on va renouveler l'aventure avec un copain. On en profite pour se rafraîchir la mémoire niveau manips et équipement. Équipement nécessaire :  Une corde simple (ou doublée pour trois personnes) Une longe avec son mousqueton-à-vis (par personne) Assez de dégaines pour la voie (+ une dégaine pour le relai) Une sangle et son mousqueton (par personne) Un mousqueton primaire (par personne) Un reverso avec deux mousquetons (par personne) "},{ id: 107, directory: "Learning", title: "Grande-voie: Les manips au relai", slug: "/learning/grande-voie-2-les-manips-au-relai", dateCreated: "2023-10-01",  content: " Se sécuriser soi-même   En arrivant au relai, insérer une dégaine dans la boucle ou plaquette qui va dans la direction de la voie (souvent la plus haute). Faire en sorte que le mousqueton inséré soit retourné pour que l'ouverture soit vers l'extérieur. Se vacher temporairement dans la dégaine (mousqueton supérieur). Annoncer VACHÉE à son second. Insérer un mousqueton-à-vis dans la seconde plaquette (ouverture vers l'extérieur). Faire passer la sangle dans ce mousqueton-à-vis — puis le visser — et dans la première dégaine (mousqueton supérieur). Tendre la sangle dans la direction d'où il me sera confortable d'assurer le second. Faire en sorte que la couture de la sangle soit proche du mousqueton avant de faire un noeud de huit à son bout. Idéalement, l'angle de la sangle ne devrait pas dépasser 45 degrés. Accrocher un mousqueton-à-vis (le primaire) au noeud de huit de cette sangle. Ajouter une longe molle au système en faisant un noeud de huit assez loin sur notre propre corde et le glisser dans la dégaine (mousqueton inférieur). Transférer la vache principale de la dégaine au primaire, réajuster sa longueur si nécessaire. À ce stade, bien vérifier que tous les mousquetons ont été vissés correctement.  Assurer son second   Ravaler la corde pour pouvoir assurer son second (prendre sec). Si le relai est plat et confortable, ravaler la corde en faisant un tas bien rangé à ses pieds. Si le relai est moins confortable, ravaler la corde en faisant des boucles sur notre pontet en le faisant passer une fois à gauche et une fois à droite en alternance. Commencer par des petites boucles qui s'agrandissent au fur et à mesure. Cela sert à éviter que la corde ne s'emmêle lorsque je la retournerai pour la passer à mon·a second. Accrocher le reverso sur le primaire à l'aide d'un mousqueton-à-vis. Faire passer la corde du grimpeur dans le reverso et la tenir en place grâce à un mousqueton-à-vis supplémentaire. Bien visser tous les mousquetons. Vérifier que le système d'assurage bloque correctement en tirant sur la corde du grimpeur. Lorsque je suis prête à assurer, annoncer DÉPART SUR LA [COULEUR] et assurer la montée en moulinette de mon·a second. À son arrivée, mon second, se vache sur le primaire et ajoute une vache molle sur la dégaine (mousqueton inférieur). Retirer le système d'assurage en défaisant le reverso et la corde lui passant en travers.  Transférer le système d'assurage   Reprendre les dégaines si nécessaire. Retourner la corde de mon pontet au pontet de mon second (petites boucles sur le dessus). Défaire ma vache molle. Le reverso étant toujours sur le primaire, faire passer ma corde dans le reverso et la tenir en place grâce à un mousqueton-à-vis supplémentaire. Prendre ma corde passant entre mon pontet et le reverso et la faire passer par la dégaine. C'est le point de renvoi qui me permet de tomber sur la dégaine et non sur mon second. Vérifier que tous les mousquetons soient bien visser. Demander à mon second de me prendre sec, retirer ma longe primaire et commencer à grimper. "},{ id: 108, directory: "Learning", title: "Making solid shampoo", slug: "/learning/making-solid-shampoo", dateCreated: "2023-10-16",  content: " Yesterday, I tried making solid shampoo for the first time. Here is the rough recipe I followed. Ingredients:  50 g SCI (Sodium Cocoyl Isethionate) 15 g bhringaraj powder 17 g chaulmoogra oil 15 g rosmarin hydrolat Optional: 24 drops of essential oils (tea tree, lavender, lemon)  The list of ingredients looks more dauting than it really is, as most of these ingredients can be swapped to other alternatives (e.g. choose your own powder, oil, water-based solution, and essential oils). For instance, a super simple recipe (not tested though) would be something like SCI + a powder + coconut oil + water. Note about the SCI  Apparently, using solid shampoo only can be irritating for your scalp (because of the SCI). It is usually recommended to alternate with a liquid shampoo — which doesn't need SCI to keep it solid. SCI is also  not very ecologically friendly to produce . I will probably look into making some liquid shampoo as well in the future. Observations and potential improvements   So easy to make! Basically: mix everything in a  bain-marie  and knead your mixture by hand with the essential oils. Too many essential oil drops; I would bring it down to max 20 drops. Would mixing only the SCI with the oil first help to melt it completely? Using a solid spoon could help to break it down as well.   Update 20/10/23: I tested the shampoo and it works great — my hair feels very clean. I wonder if it might be a bit too basic though. Would mixing in some citric acid make it smoother to use? "},{ id: 109, directory: "Learning", title: "Middle dot", slug: "/learning/middle-dot", dateCreated: "2023-10-21",  content: " Since I got back to writing in French in  Eau de poisson , I have started to use the middle dot more and more (for inclusive language). I finally found a simple shortcut that works with my French (US) keyboard on Ubuntu:  Compose key  +  .  +  - "},{ id: 110, directory: "Learning", title: "Nuages: Les bases", slug: "/learning/nuages-les-bases", dateCreated: "2023-11-27",  content: " Qu'est-ce qu'un nuage ?  Un nuage est en météorologie une masse visible constituée initialement d'une grande quantité de gouttelettes d’eau (parfois de cristaux de glace associés à des aérosols chimiques ou des minéraux) en suspension dans l’atmosphère au-dessus de la surface d'une planète. Comment se forme un nuage ?  Prenez un volume d'air saturé de vapeur d'eau (invisible). Reroidissez cet air (a.k.a. la pression baisse), une partie de la vapeur d'eau présente dans l'air se condense, on voit apparaître un nuage. Un nuage peut se former de différentes façons :  Par convection : L'air saturé prend de l'altitude par convection. La pression diminue et il y a de donc plus de condensation et de nuages. Par soulèvement orographique : Un nuage peut aussi se former lorsque l'air saturé rencontre un relief (qui monte), suit sa courbe et se refroidit donc. Par contact avec une terre froide : Par exemple, lorsque l'air chaud et humide de la mer rencontre une côte froide. L'air se refroidit et se condence. Par soulèvement frontal : Un front d'air chaud rencontre un front d'air froid. Il se soulève et peut former des nuages.  La forme des nuages  La forme d'un nuage est surtout déterminée par la stabilité de l'air qui l'entoure.  Air ambiant stable : Une masse d'air, humide et plus chaude que l'air extérieur stable, va se soulever. Elle va monter et baisser en température jusqu'au moment où sa température va égaliser celle de l'air ambiant. Pas très haut en altitude, le nuage va donc s'arrêter de monter, s'affaisser, s'allonger pour former un nuage  stratiforme . Air ambiant instable : Une masse d'air, humide et plus chaude que l'air extérieur instable (forts mouvements d'air verticaux), va également se soulever. À la différence du premier cas, l'air ambiant instable va descendre en température plus rapidement en montant en altitude. La masse d'air a donc plus de chemin à parcourir pour atteindre un point d'équilibre des températures. Le nuage formé est bourgeonnant et  cumuliforme .  Les nuages stratiformes auront donc tendance à se former plus proches du sol, jusqu'à environ 7 km d'altitude. Les nuages cumuliformes peuvent quant à eux s'étendre verticalement jusqu'à environ 20 km d'altitude. Au bout du compte, la différence entre la formation de différents types de nuages ​​est due aux températures de condensation. Certains nuages ​​se forment à des températures plus élevées et d'autres à des températures plus basses. Plus la température de la formation est basse, plus le nuage sera épais. Certains types de nuages ​​produisent également des précipitations, d'autres non. Si la température est trop basse, le nuage qui se forme sera constitué de cristaux de glace."},{ id: 111, directory: "Learning", title: "Russian: Duolingoing", slug: "/learning/russian-1-duolingoing", dateCreated: "2023-09-29",  content: " Sooo I've been Duolingoing Russian for approximately a month now. Compared to other Duolingo languages, the Russian course is really bad. For me, the main issue is that we are missing some grammar notes or simple explanations for each lesson. Well, the internet heard my plea and provided. The  Duonotes wiki offers context  for the Russian Duolingo course! I am going to review that wiki and copy-paste the parts I find interesting or would like to remember."},{ id: 112, directory: "Learning", title: "Russian: To be or not to be", slug: "/learning/russian-2-to-be-or-not-to-be", dateCreated: "2023-09-29",  content: " Russian does not have articles, nor does it normally use the verb to be in the present tense. An em-dash is used instead of the verb to be between the two nouns. Example: МОККА — КОФЕ ~ a mocha is coffee"},{ id: 113, directory: "Learning", title: "Russian: Cases, the basics", slug: "/learning/russian-3-cases-the-basics", dateCreated: "2023-09-29",  content: " Russian has 6 cases: Nominative (subject), Accusative (direct object), Genitive (of something), Prepositional (place or topic), Dative (recipient, indirect object), Instrumental (means of action). Some nouns of foreign origin are indeclinable, i.e. all their forms are the same. This includes words like метро, Дженни or кафе."},{ id: 114, directory: "Learning", title: "Russian: To have", slug: "/learning/russian-4-to-have", dateCreated: "2023-09-29",  content: " У  A  есть  B  ~ By  A  there is a  B . which is used to translate  A  has a  B . The owner (A) is in the Genitive case, while the subject (B) is in the Nominative case. Omit есть if the existence of the object is obvious or not the point. This is very typical for describing traits or a number of objects. This also applies when expressing temporary states and illnesses (She has a migraine)."},{ id: 115, directory: "Learning", title: "Russian: You, you, you", slug: "/learning/russian-5-you-you-you", dateCreated: "2023-09-29",  content: " Russian makes a distinction between ты, singular you, and вы, plural you (you all). The latter also doubles for polite you, with verbs also in plural.  Use ты with friends and your family members. Use вы with adult strangers, your teachers and in other formal interactions (at the store, the doctor's, the airport etc.). People use вы with those who are much older. Nobody is polite toward kids. "},{ id: 116, directory: "Learning", title: "Russian: About names", slug: "/learning/russian-6-about-names", dateCreated: "2023-09-29",  content: " Patronyms  Russians have three names; their first name and their surname — just like you have — and a patronymic (отчество), which is based on their father's name. Male patronymic a end in -евич or -ович, while female patronymics end in -евна or -овна. A very common polite pattern is to use a person's first name and a patronymic: Иван Иванович, вы заняты? ~ Ivan Ivanovich (= son of Ivan), are you busy? Nicknames  Russian has a casual diminutive form for many common names. They are often used with some degree of affection: Иван→Ва́ня, Мария→Маша, Александр(Александра)→Са́ша, Евгений(Евгения)→Женя, Елена→Лена, Алексей→Лёша, Пётр→Петя Greeting phrases  Здравствуйте/Привет! ~ Hello! (polite/informal) Меня зовут Клаус. А тебя/Вас? ~ My name is Klaus. And you? (polite/informal)  [Mnemonic] зовут as z'oeufs vous, an egg with a name tag  Как вас/тебя зовут? ~ What is your name? (polite/informal) Очень приятно. ~ Very nice. as in Nice to meet you. Мне тоже. ~ Me too. as in Nice to meet you too. До свидания/Пока! ~ Goodbye! (polite/informal)"},{ id: 117, directory: "Learning", title: "Russian: Frequent words", slug: "/learning/russian-7-frequent-words", dateCreated: "2023-10-02",  content: " Not being super convinced by the Russian Duolingo course, I started through this Anki deck:  The Ultimate Guide to Russian – The Most Used 5000 Words . This is my first time using Anki, so let's see how it fares (or more like how I fare at it). I'll post some of the dumb mnemonics I think of on the way."},{ id: 118, directory: "Learning", title: "Russian: Hello", slug: "/learning/russian-8-hello", dateCreated: "2023-10-02",  content: " Доброе утро! ~ Good morning!  [Mnemonic] утро as in autre. The start of another day.  Добрый день! ~ Good afternoon!  [Mnemonic] день as in day. The day is already well advanced (= afternoon).  Добрый вечер! ~ Good evening!  [Mnemonic] вечер starts with the ve of evening, чер as in cher. An expensive and luxurious evening night. "},{ id: 119, directory: "Learning", title: "Stop motion: First steps", slug: "/learning/stop-motion-first-steps", dateCreated: "2022-02-13",  content: " Here are the learnings I have from my first  stop motion project :  Pay more attention to the little emotional gestures and details. Focus less on going from A to B and more on the little repetitive movements that make the object look alive. Be careful about not bumping the camera. Use tripod whenever possible. Make sure my body doesn't come in the light way and cause unexpected shadows. Think through the logistics for the complex moves first.  E.g. How am I going to make this jar gradually fall? It will avoid extra work remaking scenes.  Make sure the camera focus is correct to avoid bluriness. For one viewpoint, simulate the whole sequence first to make sure the scene layout is right and everything is where it should be. "},{ id: 120, directory: "Media", title: "12 Angry Men", slug: "/media/12-angry-men", dateCreated: "2021-01-01",  content: "  The defense and the prosecution have rested and the jury is filing into the jury room to decide if a young Spanish-American is guilty or innocent of murdering his father. What begins as an open and shut case soon becomes a mini-drama of each of the jurors' prejudices and preconceptions about the trial, the accused, and each other. Source: TMDB "},{ id: 121, directory: "Media", title: "1984", slug: "/media/1984", dateCreated: "2021-01-01",  content: ""},{ id: 122, directory: "Media", title: "A Gentleman in Moscow", slug: "/media/a-gentleman-in-moscow", dateCreated: "2022-05-01",  content: " Weirdly wholesome, loved listening to the audiobook!"},{ id: 123, directory: "Media", title: "Adieu Bohème", slug: "/media/adieu-boheme", dateCreated: "2022-01-01",  content: " One-shot short movie, full of life and feelings.  Once the show is over at Bastille Opera's 6th basement, Professor Turrel and his team open their shop. The service offers the opportunity for couples to say goodbye beautifully: recycling sets, performers and musicians. But tonight, we follow Thibaut's story: a man who couldn't say goodbye to his love. Source: TMDB "},{ id: 124, directory: "Media", title: "Aftersun", slug: "/media/aftersun", dateCreated: "2022-12-01",  content: "  Sophie reflects on the shared joy and private melancholy of a holiday she took with her father twenty years earlier. Memories real and imagined fill the gaps between miniDV footage as she tries to reconcile the father she knew with the man she didn't. Source: TMDB "},{ id: 125, directory: "Media", title: "Ali and Nino", slug: "/media/ali-and-nino", dateCreated: "2024-02-20",  content: "  Ali and Nino  is a classic for both Georgia and Azerbaijan for a reason. I wasn't expecting to like it so much. Beautiful writing that gets you from the start."},{ id: 126, directory: "Media", title: "Amélie", slug: "/media/amelie", dateCreated: "2021-01-01",  content: "  At a tiny Parisian café, the adorable yet painfully shy Amélie accidentally discovers a gift for helping others. Soon Amelie is spending her days as a matchmaker, guardian angel, and all-around do-gooder. But when she bumps into a handsome stranger, will she find the courage to become the star of her very own love story? Source: TMDB "},{ id: 127, directory: "Media", title: "And Then There Were None", slug: "/media/and-then-there-were-none", dateCreated: "2021-01-01",  content: ""},{ id: 128, directory: "Media", title: "Basara series", slug: "/media/basara", dateCreated: "2021-01-01",  content: " This might be my favorite manga of all times."},{ id: 129, directory: "Media", title: "Broken Earth trilogy", slug: "/media/broken-earth-trilogy", dateCreated: "2021-01-01",  content: " Wowowow, this might be my favorite fantasy series ever!  A SEASON OF ENDINGS HAS BEGUN. IT STARTS WITH THE GREAT RED RIFT across the heart of the world's sole continent, spewing ash that blots out the sun. IT STARTS WITH DEATH, with a murdered son and a missing daughter. IT STARTS WITH BETRAYAL and long-dormant wounds rising up to fester. This is the Stillness, a land familiar with catastrophe, where the power of the earth is wielded as a weapon. And where there is no mercy. Source: Open Library "},{ id: 130, directory: "Media", title: "Can't We Talk About Something More Pleasant?", slug: "/media/cant-we-talk-about-something-more-pleasant", dateCreated: "2023-07-31",  content: " One of the books that really made me think about the difficulties surrounding aging. Deeply personal and touching account from Roz Chast, who is able to make us laugh and cry in the face of death."},{ id: 131, directory: "Media", title: "Captain Fantastic", slug: "/media/captain-fantastic", dateCreated: "2021-01-01",  content: "  A father living in the forests of the Pacific Northwest with his six young kids tries to assimilate back into society. Source: TMDB "},{ id: 132, directory: "Media", title: "Chroniques birmanes", slug: "/media/chroniques-birmanes", dateCreated: "2021-01-01",  content: ""},{ id: 133, directory: "Media", title: "Circe", slug: "/media/circe", dateCreated: "2021-01-01",  content: ""},{ id: 134, directory: "Media", title: "Coraline", slug: "/media/coraline", dateCreated: "2021-01-01",  content: "  A young girl discovers an idealized parallel universe behind a secret door in her new home, unaware that it contains a sinister secret. Source: TMDB "},{ id: 135, directory: "Media", title: "Django Unchained", slug: "/media/django-unchained", dateCreated: "2021-01-01",  content: "  With the help of a German bounty hunter, a freed slave sets out to rescue his wife from a brutal Mississippi plantation owner. Source: TMDB "},{ id: 136, directory: "Media", title: "Do the Right Thing", slug: "/media/do-the-right-thing", dateCreated: "2021-01-01",  content: "  Salvatore Sal Fragione is the Italian owner of a pizzeria in Brooklyn. A neighborhood local, Buggin' Out, becomes upset when he sees that the pizzeria's Wall of Fame exhibits only Italian actors. Buggin' Out believes a pizzeria in a black neighborhood should showcase black actors, but Sal disagrees. The wall becomes a symbol of racism and hate to Buggin' Out and to other people in the neighborhood, and tensions rise. Source: TMDB "},{ id: 137, directory: "Media", title: "Dune", slug: "/media/dune", dateCreated: "2021-12-01",  content: "  Paul Atreides, a brilliant and gifted young man born into a great destiny beyond his understanding, must travel to the most dangerous planet in the universe to ensure the future of his family and his people. As malevolent forces explode into conflict over the planet's exclusive supply of the most precious resource in existence-a commodity capable of unlocking humanity's greatest potential-only those who can conquer their fear will survive. Source: TMDB "},{ id: 138, directory: "Media", title: "Everything Everywhere All At Once", slug: "/media/everything-everywhere-all-at-once", dateCreated: "2022-12-01",  content: "  An aging Chinese immigrant is swept up in an insane adventure, where she alone can save what's important to her by connecting with the lives she could have led in other universes. Source: TMDB "},{ id: 139, directory: "Media", title: "Ex Machina", slug: "/media/ex-machina", dateCreated: "2021-01-01",  content: "  Caleb, a coder at the world's largest internet company, wins a competition to spend a week at a private mountain retreat belonging to Nathan, the reclusive CEO of the company. But when Caleb arrives at the remote location he finds that he will have to participate in a strange and fascinating experiment in which he must interact with the world's first true artificial intelligence, housed in the body of a beautiful robot girl. Source: TMDB "},{ id: 140, directory: "Media", title: "Fight Club", slug: "/media/fight-club", dateCreated: "2021-01-01",  content: "  A ticking-time-bomb insomniac and a slippery soap salesman channel primal male aggression into a shocking new form of therapy. Their concept catches on, with underground fight clubs forming in every town, until an eccentric gets in the way and ignites an out-of-control spiral toward oblivion. Source: TMDB "},{ id: 141, directory: "Media", title: "Freakonomics", slug: "/media/freakonomics", dateCreated: "2021-01-01",  content: ""},{ id: 142, directory: "Media", title: "Free Solo", slug: "/media/free-solo", dateCreated: "2021-01-01",  content: " Cold sweats.  Follow Alex Honnold as he attempts to become the first person to ever free solo climb Yosemite's 3,000 foot high El Capitan wall. With no ropes or safety gear, this would arguably be the greatest feat in rock climbing history. Source: TMDB "},{ id: 143, directory: "Media", title: "Get Out", slug: "/media/get-out", dateCreated: "2021-01-01",  content: "  Chris and his girlfriend Rose go upstate to visit her parents for the weekend. At first, Chris reads the family's overly accommodating behavior as nervous attempts to deal with their daughter's interracial relationship, but as the weekend progresses, a series of increasingly disturbing discoveries lead him to a truth that he never could have imagined. Source: TMDB "},{ id: 144, directory: "Media", title: "Gone Girl", slug: "/media/gone-girl", dateCreated: "2021-01-01",  content: ""},{ id: 145, directory: "Media", title: "Guillermo del Toro's Pinocchio", slug: "/media/guillermo-del-toros-pinocchio", dateCreated: "2023-01-01",  content: ""},{ id: 146, directory: "Media", title: "Half of a Yellow Sun", slug: "/media/half-of-a-yellow-sun", dateCreated: "2021-01-01",  content: "  In 1960s Nigeria, Ugwu, a boy from a poor village, goes to work for Odenigbo, a radical university professor. Soon they are joined by Olanna, a young woman who has abandoned a life of privilege to live with her charismatic lover. Into their world comes Richard, an English writer, who has fallen for Olanna’s sharp-tongued sister Kainene. But when the shocking horror of civil war engulfs the nation, their loves and loyalties are severely tested, while their lives pull apart and collide once again in ways none of them could have imagined Source: Open Library "},{ id: 147, directory: "Media", title: "Harry Potter series", slug: "/media/harry-potter-series", dateCreated: "2021-01-01",  content: " Childhood classics.  Rescued from the outrageous neglect of his aunt and uncle, a young boy with a great destiny proves his worth while attending Hogwarts School for Wizards and Witches Source: Open Library "},{ id: 148, directory: "Media", title: "Her", slug: "/media/her", dateCreated: "2021-01-01",  content: "  In the not so distant future, Theodore, a lonely writer, purchases a newly developed operating system designed to meet the user's every need. To Theodore's surprise, a romantic relationship develops between him and his operating system. This unconventional love story blends science fiction and romance in a sweet tale that explores the nature of love and the ways that technology isolates and connects us all. Source: TMDB "},{ id: 149, directory: "Media", title: "His Dark Materials series", slug: "/media/his-dark-materials-series", dateCreated: "2021-01-01",  content: ""},{ id: 150, directory: "Media", title: "Hocus Pocus", slug: "/media/hocus-pocus", dateCreated: "2021-01-01",  content: " Childhood classics, not sure how well it has aged though.  After 300 years of slumber, three sister witches are accidentally resurrected in Salem on Halloween night, and it is up to three kids and their newfound feline friend to put an end to the witches' reign of terror once and for all. Source: TMDB "},{ id: 151, directory: "Media", title: "Howl's Moving Castle", slug: "/media/howls-moving-castle", dateCreated: "2023-06-29",  content: " Incredibly poetic, Miyazaki at his best.  Sophie, a young milliner is turned into an elderly woman by a witch who enters her shop and curses her. She encounters a wizard named Howl and gets caught up in his resistance to fighting for the king. Source: TMDB "},{ id: 152, directory: "Media", title: "Hunt for the Wilderpeople", slug: "/media/hunt-for-the-wilderpeople", dateCreated: "2021-01-01",  content: "  Ricky is a defiant young city kid who finds himself on the run with his cantankerous foster uncle in the wild New Zealand bush. A national manhunt ensues, and the two are forced to put aside their differences and work together to survive. Source: TMDB "},{ id: 153, directory: "Media", title: "In the Mood for Love", slug: "/media/in-the-mood-for-love", dateCreated: "2023-09-20",  content: " A beautiful love story unfolded in snapshots. It is quiet, subtle, heartbreaking in its restraint. The photography is lush and evocative, each scene a piece of art."},{ id: 154, directory: "Media", title: "Inception", slug: "/media/inception", dateCreated: "2021-01-01",  content: "  Cobb, a skilled thief who commits corporate espionage by infiltrating the subconscious of his targets is offered a chance to regain his old life as payment for a task considered to be impossible: inception, the implantation of another person's idea into a target's subconscious. Source: TMDB "},{ id: 155, directory: "Media", title: "Interstellar", slug: "/media/interstellar", dateCreated: "2021-01-01",  content: "  The adventures of a group of explorers who make use of a newly discovered wormhole to surpass the limitations on human space travel and conquer the vast distances involved in an interstellar voyage. Source: TMDB "},{ id: 156, directory: "Media", title: "Joker", slug: "/media/joker", dateCreated: "2021-01-01",  content: "  During the 1980s, a failed stand-up comedian is driven insane and turns to a life of crime and chaos in Gotham City while becoming an infamous psychopathic crime figure. Source: TMDB "},{ id: 157, directory: "Media", title: "Jumanji", slug: "/media/jumanji", dateCreated: "2021-01-01",  content: "  Four teenagers in detention discover an old video game console with a game they’ve never heard of. When they decide to play, they are immediately sucked into the jungle world of Jumanji in the bodies of their avatars. They’ll have to complete the adventure of their lives filled with fun, thrills and danger or be stuck in the game forever! Source: TMDB "},{ id: 158, directory: "Media", title: "Kill Bill", slug: "/media/kill-bill", dateCreated: "2021-01-01",  content: "  An assassin is shot and almost killed by her ruthless employer, Bill, and other members of their assassination circle – but she lives to plot her vengeance. Source: TMDB "},{ id: 159, directory: "Media", title: "Kirikou and the Sorceress", slug: "/media/kirikou-and-the-sorceress", dateCreated: "2021-01-01",  content: " Childhood classics.  Drawn from elements of West African folk tales, it depicts how a newborn boy, Kirikou, saves his village from the evil witch Karaba. Source: TMDB "},{ id: 160, directory: "Media", title: "Kubo and the Two Strings", slug: "/media/kubo-and-the-two-strings", dateCreated: "2021-01-01",  content: "  Kubo mesmerizes the people in his village with his magical gift for spinning wild tales with origami. When he accidentally summons an evil spirit seeking vengeance, Kubo is forced to go on a quest to solve the mystery of his fallen samurai father and his mystical weaponry, as well as discover his own magical powers. Source: TMDB "},{ id: 161, directory: "Media", title: "La Horde de Contrevent", slug: "/media/la-horde-de-contrevent", dateCreated: "2021-01-01",  content: ""},{ id: 162, directory: "Media", title: "La La Land", slug: "/media/la-la-land", dateCreated: "2021-01-01",  content: "  Mia, an aspiring actress, serves lattes to movie stars in between auditions and Sebastian, a jazz musician, scrapes by playing cocktail party gigs in dingy bars, but as success mounts they are faced with decisions that begin to fray the fragile fabric of their love affair, and the dreams they worked so hard to maintain in each other threaten to rip them apart. Source: TMDB "},{ id: 163, directory: "Media", title: "La Quête d'Ewilan series", slug: "/media/la-quete-dewilan-series", dateCreated: "2021-01-01",  content: ""},{ id: 164, directory: "Media", title: "Le Grand Méchant Renard", slug: "/media/le-grand-mechant-renard", dateCreated: "2021-01-01",  content: "  Face à un lapin idiot, un cochon jardinier, un chien paresseux et une poule caractérielle, un renard chétif tente de trouver sa place en tant que grand prédateur. Devant l'absence d'efficacité de ses méthodes, il développe une nouvelle stratégie. Sa solution : voler des oeufs, élever les poussins, les effrayer et les croquer. Mais le plan tourne au vinaigre lorsque le renard se découvre un instinct maternel... "},{ id: 165, directory: "Media", title: "L'échappée belle", slug: "/media/lechappee-belle", dateCreated: "2021-01-01",  content: ""},{ id: 166, directory: "Media", title: "Life of Pi", slug: "/media/life-of-pi-book", dateCreated: "2021-01-01",  content: ""},{ id: 167, directory: "Media", title: "Life of Pi", slug: "/media/life-of-pi", dateCreated: "2021-01-01",  content: "  The story of an Indian boy named Pi, a zookeeper's son who finds himself in the company of a hyena, zebra, orangutan, and a Bengal tiger after a shipwreck sets them adrift in the Pacific Ocean. Source: TMDB "},{ id: 168, directory: "Media", title: "Loki", slug: "/media/loki", dateCreated: "2021-12-01",  content: " I am pretty sure I liked it, just because I love Loki's character.  After stealing the Tesseract during the events of “Avengers: Endgame,” an alternate version of Loki is brought to the mysterious Time Variance Authority, a bureaucratic organization that exists outside of time and space and monitors the timeline. They give Loki a choice: face being erased from existence due to being a “time variant” or help fix the timeline and stop a greater threat. Source: TMDB "},{ id: 169, directory: "Media", title: "Lord of the Flies", slug: "/media/lord-of-the-flies", dateCreated: "2021-01-01",  content: ""},{ id: 170, directory: "Media", title: "Maniac", slug: "/media/maniac", dateCreated: "2021-01-01",  content: "  Two strangers are drawn to a mysterious pharmaceutical trial that will, they're assured, with no complications or side-effects whatsoever, solve all of their problems, permanently. Things do not go as planned. Source: TMDB "},{ id: 171, directory: "Media", title: "Mary Poppins", slug: "/media/mary-poppins", dateCreated: "2021-01-01",  content: " I just loved this movie as a child.  Mr Banks is looking for a nanny for his two mischievous children and comes across Mary Poppins, an angelic nanny. She not only brings a change in their lives but also spreads happiness. Source: TMDB "},{ id: 172, directory: "Media", title: "Memento", slug: "/media/memento", dateCreated: "2021-01-01",  content: "  Leonard Shelby is tracking down the man who raped and murdered his wife. The difficulty of locating his wife's killer, however, is compounded by the fact that he suffers from a rare, untreatable form of short-term memory loss. Although he can recall details of life before his accident, Leonard cannot remember what happened fifteen minutes ago, where he's going, or why. Source: TMDB "},{ id: 173, directory: "Media", title: "Memoirs of a Geisha", slug: "/media/memoirs-of-a-geisha", dateCreated: "2021-01-01",  content: ""},{ id: 174, directory: "Media", title: "Métaphysique des Tubes", slug: "/media/metaphysique-des-tubes", dateCreated: "2021-01-01",  content: ""},{ id: 175, directory: "Media", title: "Midsommar", slug: "/media/midsommar", dateCreated: "2021-01-01",  content: " This is the kind of horror movie I'm into.  Several friends travel to Sweden to study as anthropologists a summer festival that is held every ninety years in the remote hometown of one of them. What begins as a dream vacation in a place where the sun never sets, gradually turns into a dark nightmare as the mysterious inhabitants invite them to participate in their disturbing festive activities. Source: TMDB "},{ id: 176, directory: "Media", title: "Moonrise Kingdom", slug: "/media/moonrise-kingdom", dateCreated: "2021-01-01",  content: " My favorite Wes Anderson movie.  Set on an island off the coast of New England in the summer of 1965, Moonrise Kingdom tells the story of two twelve-year-olds who fall in love, make a secret pact, and run away together into the wilderness. As various authorities try to hunt them down, a violent storm is brewing off-shore – and the peaceful island community is turned upside down in more ways than anyone can handle. Source: TMDB "},{ id: 177, directory: "Media", title: "Mushi-Shi", slug: "/media/mushi-shi", dateCreated: "2021-01-01",  content: "  Ginko, a Mushi master, travels from place to place researching the Mushi and helping people who are suffering because of it. Source: TMDB  Comfort anime series, relaxing, feel-good with a bit of magic."},{ id: 178, directory: "Media", title: "Nage libre", slug: "/media/nage-libre", dateCreated: "2021-01-01",  content: "  Josi, Marsha et monsieur Nale, trois potes saumons, remontent la rivière en quête de saumonettes à féconder. Mais très vite, leur soif de liberté les pousse à l'aventure : l'eau serait-elle plus bleue ailleurs ? Un road movie aquatique plein d'humour, de rebondissements et de questions existentielles. "},{ id: 179, directory: "Media", title: "Never Let Me Go", slug: "/media/never-let-me-go", dateCreated: "2021-01-01",  content: ""},{ id: 180, directory: "Media", title: "Orphan black", slug: "/media/orphan-black", dateCreated: "2021-01-01",  content: "  A streetwise hustler is pulled into a compelling conspiracy after witnessing the suicide of a girl who looks just like her. Source: TMDB "},{ id: 181, directory: "Media", title: "Our flag means death", slug: "/media/our-flag-means-death", dateCreated: "2022-09-01",  content: "  After trading in the seemingly charmed life of a gentleman for one of a swashbuckling buccaneer, Stede Bonnet becomes captain of the pirate ship Revenge. Struggling to earn the respect of his potentially mutinous crew, Stede’s fortunes change after a fateful run-in with the infamous Captain Blackbeard. Stede and crew attempt to get their ship together and survive life on the high seas. Source: TMDB  Will I ever hate anything related to Taika Waititi? Mmh... not sure, no.  Our Flag Means Death  has just the right balance of funny and sad, light and dark. All the characters are amazing and the cast improv acting is just on point."},{ id: 182, directory: "Media", title: "Parasite", slug: "/media/parasite", dateCreated: "2021-01-01",  content: "  All unemployed, Ki-taek's family takes peculiar interest in the wealthy and glamorous Parks for their livelihood until they get entangled in an unexpected incident. Source: TMDB "},{ id: 183, directory: "Media", title: "Reservation Dogs", slug: "/media/reservation-dogs", dateCreated: "2022-12-01",  content: "  This comedy follows the exploits of four Indigenous teenagers in rural Oklahoma who steal, rob and save in order to get to the exotic, mysterious and faraway land of California. Source: TMDB "},{ id: 184, directory: "Media", title: "Schindler's List", slug: "/media/schindlers-list", dateCreated: "2021-01-01",  content: "  The true story of how businessman Oskar Schindler saved over a thousand Jewish lives from the Nazis while they worked as slaves in his factory during World War II. Source: TMDB "},{ id: 185, directory: "Media", title: "Sherlock", slug: "/media/sherlock", dateCreated: "2021-01-01",  content: " The TV series that got us all in love with Cumberbatch.  A modern update finds the famous sleuth and his doctor partner solving crime in 21st century London. Source: TMDB "},{ id: 186, directory: "Media", title: "Shutter Island", slug: "/media/shutter-island", dateCreated: "2021-01-01",  content: "  World War II soldier-turned-U.S. Marshal Teddy Daniels investigates the disappearance of a patient from a hospital for the criminally insane, but his efforts are compromised by troubling visions and a mysterious doctor. Source: TMDB "},{ id: 187, directory: "Media", title: "Soul", slug: "/media/soul", dateCreated: "2021-01-01",  content: " Once again, Pixar manages to steal my heart.  Joe Gardner is a middle school teacher with a love for jazz music. After a successful audition at the Half Note Club, he suddenly gets into an accident that separates his soul from his body and is transported to the You Seminar, a center in which souls develop and gain passions before being transported to a newborn child. Joe must enlist help from the other souls-in-training, like 22, a soul who has spent eons in the You Seminar, in order to get back to Earth. Source: TMDB "},{ id: 188, directory: "Media", title: "Spotlight", slug: "/media/spotlight", dateCreated: "2021-01-01",  content: "  The true story of how the Boston Globe uncovered the massive scandal of child molestation and cover-up within the local Catholic Archdiocese, shaking the entire Catholic Church to its core. Source: TMDB "},{ id: 189, directory: "Media", title: "Still Life", slug: "/media/still-life", dateCreated: "2021-01-01",  content: ""},{ id: 190, directory: "Media", title: "Stories of Your Life and Others", slug: "/media/stories-of-your-life-and-others", dateCreated: "2021-01-01",  content: ""},{ id: 191, directory: "Media", title: "Stranger Things", slug: "/media/stranger-things", dateCreated: "2022-01-01",  content: "  When a young boy vanishes, a small town uncovers a mystery involving secret experiments, terrifying supernatural forces, and one strange little girl. Source: TMDB "},{ id: 192, directory: "Media", title: "Subaru series", slug: "/media/subaru-series", dateCreated: "2021-01-01",  content: " This might be the manga that got me started on the manga path.  Chaque jour, à la fin des classes, Subaru prend le chemin de l'hôpital, où elle danse pour son frère, atteint d'une tumeur cancéreuse au cerveau. Lorsque ce dernier décède, la jeune femme traumatisée entreprend de devenir danseuse professionnelle, cet art étant pour elle le seul moyen d'extérioriser ses sentiments. Or, des cabarets aux grandes scènes, la route est longue ... Une intrigue bien ficelée mettant en scène des protagonistes au caractère bien trempé. Si quelques crises de larmes auraient pu être supprimées, le courage, la persévérance et le travail qu'effectue la jeune femme sur elle-même sont rendus avec une belle intensité, à la fois par des dialogues bien sentis et un gracieux coup de crayon faisant se mouvoir les corps avec souplesse. Source: Open Library "},{ id: 193, directory: "Media", title: "The Artist", slug: "/media/the-artist", dateCreated: "2021-01-01",  content: "  Hollywood, 1927: As silent movie star George Valentin wonders if the arrival of talking pictures will cause him to fade into oblivion, he sparks with Peppy Miller, a young dancer set for a big break. Source: TMDB "},{ id: 194, directory: "Media", title: "The Beach", slug: "/media/the-beach", dateCreated: "2021-01-01",  content: ""},{ id: 195, directory: "Media", title: "The Godfather", slug: "/media/the-godfather", dateCreated: "2023-01-01",  content: ""},{ id: 196, directory: "Media", title: "The Great Escape", slug: "/media/the-great-escape", dateCreated: "2021-01-01",  content: "  The Nazis, exasperated at the number of escapes from their prison camps by a relatively small number of Allied prisoners, relocate them to a high-security 'escape-proof' camp to sit out the remainder of the war. Undaunted, the prisoners plan one of the most ambitious escape attempts of World War II. Based on a true story. Source: TMDB "},{ id: 197, directory: "Media", title: "The Great Gatsby", slug: "/media/the-great-gatsby", dateCreated: "2021-01-01",  content: ""},{ id: 198, directory: "Media", title: "The Guernsey Literary and Potato Peel Pie Society", slug: "/media/the-guernsey-literary-and-potato-peel-pie-society", dateCreated: "2021-01-01",  content: " The wholesome read I go back to whenever I need good vibes.  A glorious, gorgeous, completely irresistible novel, Guernsey will win hearts everywhere. An affecting, emotional epistolary novel recalling the charm and warmth of 84 Charing Cross Road and the gentle wit of Alexander McCall Smith, Guernsey tells the story of a small group of neighbors on the island of Guernsey off the coast of England. A pig farmer, an elegant lady, a potion maker, a rag and bone man, a footman posing as a lord who survive the Nazi occupation in the Second World War by gathering over humble potato peel pie to talk of Shakespeare, Chaucer, and the great novels. With the war over, Juliet, an eager writer from London, uncovers just how dramatic the lives of these islanders really were during the war, and through them, discovers the drama in her own life. Guernsey is a real gem of a novel. It's that rare thing: a small story, simply told, that contains within it enormous depth and heart. Completely accessible and immediately captivating, the characters are nothing short of enchanting. Gloriously honest, moving, sweet and funny, a sheer delight from beginning to end, this is truly a book to fall in love with. Source: Open Library "},{ id: 199, directory: "Media", title: "The Handmaiden", slug: "/media/the-handmaiden", dateCreated: "2021-01-01",  content: "  In 1930s Korea, a swindler and a young woman pose as a Japanese count and a handmaiden to seduce a Japanese heiress and steal her fortune. Source: TMDB "},{ id: 200, directory: "Media", title: "The Handmaid's Tale", slug: "/media/the-handmaids-tale", dateCreated: "2021-01-01",  content: ""},{ id: 201, directory: "Media", title: "The Martian", slug: "/media/the-martian", dateCreated: "2021-01-01",  content: ""},{ id: 202, directory: "Media", title: "The Matrix", slug: "/media/the-matrix", dateCreated: "2021-01-01",  content: "  Set in the 22nd century, The Matrix tells the story of a computer hacker who joins a group of underground insurgents fighting the vast and powerful computers who now rule the earth. Source: TMDB "},{ id: 203, directory: "Media", title: "The Murderbot Diaries", slug: "/media/the-murderbot-diaries", dateCreated: "2021-01-01",  content: " So many books in this series that perfectly mixes sci-fi and comedy. Loved them all.  The series is about an artificial construct designed as a Security Unit, which manages to override its governor unit, thus enabling it to develop independence. It calls itself Murderbot, and likes to watch unrealistic soap operas. As it spends more time with some caring humans, it starts developing feelings that it does not care for. Source: Open Library "},{ id: 204, directory: "Media", title: "The Night Circus", slug: "/media/the-night-circus", dateCreated: "2021-01-01",  content: ""},{ id: 205, directory: "Media", title: "The Omnivore's Dilemma", slug: "/media/the-omnivores-dilemma", dateCreated: "2021-01-01",  content: ""},{ id: 206, directory: "Media", title: "The Others", slug: "/media/the-others", dateCreated: "2021-01-01",  content: "  Grace is a religious woman who lives in an old house kept dark because her two children, Anne and Nicholas, have a rare sensitivity to light. When the family begins to suspect the house is haunted, Grace fights to protect her children at any cost in the face of strange events and disturbing visions. Source: TMDB "},{ id: 207, directory: "Media", title: "The Pianist", slug: "/media/the-pianist", dateCreated: "2021-01-01",  content: "  The true story of pianist Władysław Szpilman's experiences in Warsaw during the Nazi occupation. When the Jews of the city find themselves forced into a ghetto, Szpilman finds work playing in a café; and when his family is deported in 1942, he stays behind, works for a while as a laborer, and eventually goes into hiding in the ruins of the war-torn city. Source: TMDB "},{ id: 208, directory: "Media", title: "The Power of the Dog", slug: "/media/the-power-of-the-dog", dateCreated: "2021-12-01",  content: "  A domineering but charismatic rancher wages a war of intimidation on his brother's new wife and her teen son, until long-hidden secrets come to light. Source: TMDB "},{ id: 209, directory: "Media", title: "The Prestige", slug: "/media/the-prestige", dateCreated: "2021-01-01",  content: "  A mysterious story of two magicians whose intense rivalry leads them on a life-long battle for supremacy -- full of obsession, deceit and jealousy with dangerous and deadly consequences. Source: TMDB "},{ id: 210, directory: "Media", title: "The Queen's Gambit", slug: "/media/the-queens-gambit", dateCreated: "2023-01-01",  content: "  In a Kentucky orphanage in the 1950s, a young girl discovers an astonishing talent for chess while struggling with addiction. Source: TMDB "},{ id: 211, directory: "Media", title: "The Shawshank Redemption", slug: "/media/the-shawshank-redemption", dateCreated: "2021-01-01",  content: "  Framed in the 1940s for the double murder of his wife and her lover, upstanding banker Andy Dufresne begins a new life at the Shawshank prison, where he puts his accounting skills to work for an amoral warden. During his long stretch in prison, Dufresne comes to be admired by the other inmates -- including an older prisoner named Red -- for his integrity and unquenchable sense of hope. Source: TMDB "},{ id: 212, directory: "Media", title: "The Thirteenth Tale", slug: "/media/the-thirteenth-tale", dateCreated: "2021-01-01",  content: ""},{ id: 213, directory: "Media", title: "The Three-Body Problem trilogy", slug: "/media/the-three-body-problem", dateCreated: "2021-01-01",  content: "  Cixin Liu's trilogy-opening novel about first contact with aliens and the clandestine struggle with them over Earth's future, and its scientific progress in particular. Source: Open Library "},{ id: 214, directory: "Media", title: "The Truman Show", slug: "/media/the-truman-show", dateCreated: "2021-01-01",  content: "  Truman Burbank is the star of The Truman Show, a 24-hour-a-day reality TV show that broadcasts every aspect of his life without his knowledge. His entire life has been an unending soap opera for consumption by the rest of the world. And everyone he knows, including his wife and his best friend, is really an actor, paid to be part of his life. Source: TMDB "},{ id: 215, directory: "Media", title: "The Wonderful Story of Henry Sugar", slug: "/media/the-wonderful-story-of-henry-sugar", dateCreated: "2023-10-05",  content: " Wes Anderson directed four shorts based on stories by Roald Dahl. They are poetic, dark, magnificient and should be watched in one sitting.  The Wonderful Story of Henry Sugar The Swan The Rat Catcher Poison "},{ id: 216, directory: "Media", title: "Three Billboards Outside Ebbing, Missouri", slug: "/media/three-billboards-outside-ebbing-missouri", dateCreated: "2021-01-01",  content: "  After seven months have passed without a culprit in her daughter's murder case, Mildred Hayes makes a bold move, painting three signs leading into her town with a controversial message directed at Bill Willoughby, the town's revered chief of police. When his second-in-command Officer Jason Dixon, an immature mother's boy with a penchant for violence, gets involved, the battle between Mildred and Ebbing's law enforcement is only exacerbated. Source: TMDB "},{ id: 217, directory: "Media", title: "Triangle of Sadness", slug: "/media/triangle-of-sadness", dateCreated: "2022-12-01",  content: "  A celebrity model couple are invited on a luxury cruise for the uber-rich, helmed by an unhinged, alcoholic captain. What first appears Instagrammable ends catastrophically, leaving the survivors stranded on a desert island in a struggle of hierarchy. Source: TMDB "},{ id: 218, directory: "Media", title: "WALL·E", slug: "/media/wall-e", dateCreated: "2021-01-01",  content: "  WALL·E is the last robot left on an Earth that has been overrun with garbage and all humans have fled to outer space. For 700 years he has continued to try and clean up the mess, but has developed some rather interesting human-like qualities. When a ship arrives with a sleek new type of robot, WALL·E thinks he's finally found a friend and stows away on the ship when it leaves. Source: TMDB "},{ id: 219, directory: "Media", title: "WandaVision", slug: "/media/wandavision", dateCreated: "2021-12-01",  content: " One of the best series of the Marvel universe.  Wanda Maximoff and Vision — two super-powered beings living idealized suburban lives — begin to suspect that everything is not as it seems. Source: TMDB "},{ id: 220, directory: "Media", title: "What We Do in the Shadows", slug: "/media/what-we-do-in-the-shadows", dateCreated: "2021-01-01",  content: " My first Waititi movie! He just knows how to make me laugh.  Vampire housemates try to cope with the complexities of modern life and show a newly turned hipster some of the perks of being undead. Source: TMDB "},{ id: 221, directory: "Media", title: "Where the mountain meets the moon", slug: "/media/where-the-mountain-meets-the-moon", dateCreated: "2024-03-13",  content: "  In the valley of Fruitless Mountain, a young girl named Minli spends her days working hard in the fields and her nights listening to her father spin fantastic tales about the Jade Dragon and the Old Man of the Moon. Minli's mother, tired of their poor life, chides him for filling her head with nonsense. But Minli believes these enchanting stories and embarks on an extraordinary journey to find the Old Man of the Moon and ask him how her family can change their fortune. She encounters an assorted cast of characters and magical creatures along the way, including a dragon who accompanies her on her quest.  A story inspired by Chinese legends with a One Thousand and One Nights structure. I loved this short read!"},{ id: 222, directory: "Media", title: "Wizard of Oz", slug: "/media/wizard-of-oz", dateCreated: "2021-01-01",  content: " Childhood classics.  Young Dorothy finds herself in a magical world where she makes friends with a lion, a scarecrow and a tin man as they make their way along the yellow brick road to talk with the Wizard and ask for the things they miss most in their lives. The Wicked Witch of the West is the only thing that could stop them. Source: TMDB "},{ id: 223, directory: "Media", title: "World of Tomorrow", slug: "/media/world-of-tomorrow", dateCreated: "2021-01-01",  content: "  A little girl is taken on a mind-bending tour of her distant future. Source: TMDB15 "},{ id: 224, directory: "Media", title: "Years and Years", slug: "/media/years-and-years", dateCreated: "2021-01-01",  content: "  As Britain is rocked by unstable political, economic and technological advances, members of the Lyons family converge on one crucial night in 2019. Over the next 15 years, the twists and turns of their everyday lives are explored as we find out if this ordinary family could change the world. Source: TMDB "},{ id: 225, directory: "Media", title: "Yotsuba&! series", slug: "/media/yotsuba-series", dateCreated: "2021-01-01",  content: " Wholesome and hilarious.  Yotsuba&#x26;! is centered on Yotsuba Koiwai, a five-year-old girl who is energetic, cheerful, curious, odd, and quirky — so much so that even her own father calls her unusual. She is also initially ignorant about many things that a child her age would be expected to know — among them doorbells, escalators, air conditioners, and even playground swings. This naïveté is the premise of humorous stories whereby which she learns about — and frequently misunderstands — everyday things. Source: Wikipedia "},{ id: 226, directory: "Recipes", title: "Asparagus and baked potatoes", slug: "/recipes/asparagus-and-baked-potatoes", dateCreated: "2022-06-18", content: " 5 pommes de terre, 1 botte d'asperges blanches, 1 poignée de noix ou amandes, Miel, Gros sel, Flocons de piment"},{ id: 227, directory: "Recipes", title: "Banana bread", slug: "/recipes/banana-bread", dateCreated: "2021-02-05", content: "This is a recipe that Niamh recommended to me. And as with most dessert recipes, she was spot on. Definitely one of the best banana breads I have eaten. It has a very strong banana taste, is moist in the inside, but crispy on the outside. 115 g butter, cut into chunks, 160 g brown sugar, 5 medium-large mashed bananas, 2 large eggs, 1 tsp salt, 1.5 tsp ground cinnamon, 0.5 tsp ground nutmeg, 260 g flour, 1 tsp baking soda (Natron), >100 sourdough starter (or 1 tsp baking powder)"},{ id: 228, directory: "Recipes", title: "Banana ice cream", slug: "/recipes/banana-ice-cream", dateCreated: "2023-06-30", content: " 5 bananes bien mûres, Un peu de crème, de lait ou de yaourt"},{ id: 229, directory: "Recipes", title: "Basic millet", slug: "/recipes/basic-millet", dateCreated: "2022-10-06", content: " 200 g de millet (environ une tasse), 500 g d'eau"},{ id: 230, directory: "Recipes", title: "Beetroot carpaccio", slug: "/recipes/beetroot-carpaccio", dateCreated: "2023-03-09", content: " 1 beetroot, Olive oil, Flaked salt"},{ id: 231, directory: "Recipes", title: "Bell pepper spread", slug: "/recipes/bell-pepper-spread", dateCreated: "2022-10-31", content: " 1 bell pepper sliced in 8 pieces, 1 large handful of fresh herbs (e.g. cilantro, parsley, etc.), Olive oil, Salt, Lemon or vinegar"},{ id: 232, directory: "Recipes", title: "Brazilian crispy smashed potatoes", slug: "/recipes/brazilian-crispy-smashed-potatoes", dateCreated: "2022-10-06", content: " 20 small potatoes, 2 tbsp olive oil, 1.5 tsp salt"},{ id: 233, directory: "Recipes", title: "Brownies", slug: "/recipes/brownies", dateCreated: "2021-03-24", content: " 110 g butter melted and cooled, 200 g sugar, 2 large eggs, 25 g all-purpose flour, 60 g cocoa powder, 1 pinch salt, 1 pinch baking powder"},{ id: 234, directory: "Recipes", title: "Bulgarian national salad (shopska)", slug: "/recipes/bulgarian-national-salad", dateCreated: "2022-11-12", content: "I was introduced to this super simple salad recipe when traveling in Eastern Europe. A similar recipe exists in multiple of those countries, and always stars three main ingredients: cucumbers, tomatoes and feta. From there, everything else is optional. 1 cucumber, diced, 3 tomatoes, diced, 1 feta cheese, diced"},{ id: 235, directory: "Recipes", title: "Chinese chili oil", slug: "/recipes/chinese-chili-oil", dateCreated: "2022-04-29", content: " 2 cups of neutral oil, 1 cup or less of chili flakes"},{ id: 236, directory: "Recipes", title: "Clafoutis", slug: "/recipes/clafoutis", dateCreated: "2023-07-18", content: "Recette d'un vieux bouquin de cuisine de France. 500 g de fruits avec leurs noyaux (cerises, mirabelles, prunes, baies), 1 cuillère à soupe bombée de farine, 65 g de sucre, 1 pincée de sel, 120 g (12 cl) de lait, 2 oeufs"},{ id: 237, directory: "Recipes", title: "Cleaning spray for everything", slug: "/recipes/cleaning-spray-for-everything", dateCreated: "2023-02-13", content: " Bouteille à sprayer de 500 ml, 1,5 c-à-c. bicarbonate de soude, Eau chaude, 2 c-à-c. vinaigre blanc, 15–20 gouttes d'huiles essentielles (e.g. citron, eucalyptus, pin sylvestre)"},{ id: 238, directory: "Recipes", title: "Cold spicy noodles", slug: "/recipes/cold-spicy-noodles", dateCreated: "2023-08-25", content: " Wheat noodles, 4–6 tbsp oil, 2–3 tbsp gochugaru, 2 tbsp sesame seeds, 2 garlic cloves, cut small, 2 tsp honey, 2–3 tbsp soy sauce, 2 tbsp white vinegar, 3–4 tbsp sesame or peanut paste, 2+ tbsp water, Cucumber, fresh out of the fridge and julienned"},{ id: 239, directory: "Recipes", title: "Couscous", slug: "/recipes/couscous", dateCreated: "2023-10-29", content: "Recette de couscous de Souad, préparée ensemble pour la première fois à Joux en octobre 2023. Il fait légèrement frais dehors, on allume le poêle et on mange tous et toutes du même plat. Autour de la table : France, Souad, Greg, Karima, Thibaud, Robin et moi. Huile d'olive, 1 kg de semoule de blé, 600 g d'eau froide, salée, 2–3 oignons, Persil, Pois-chiches, trempés la veille, 2 boîtes de tomates, 1 petit pot de concentré de tomate, 1 petit verre de vin rouge, Ras el hanout, Sel et poivre, Légumes cuisson lente : carottes, navets, aubergine (coupés en gros morceaux), Légumes cuisson rapide : courgettes, poivrons (coupés en gros morceaux), Épaule d'agneau coupée en gros morceaux, Cuisses de poulet, Merguez"},{ id: 240, directory: "Recipes", title: "Crispy oven fries", slug: "/recipes/crispy-oven-fries", dateCreated: "2022-04-07", content: " 4 large dry and mealy Russet Potatoes, 1 tbsp vinegar, 4 tbsp peanut oil, Coarse salt"},{ id: 241, directory: "Recipes", title: "Croûtons", slug: "/recipes/croutons", dateCreated: "2023-04-20", content: " Old bread, 2 tbsp olive oil, Salt, 1 garlic clove, pressed"},{ id: 242, directory: "Recipes", title: "Dumpling soup with egg whites", slug: "/recipes/dumpling-soup-with-egg-whites", dateCreated: "2022-11-08", content: " 3 egg whites"},{ id: 243, directory: "Recipes", title: "Eggplant tarte tatin", slug: "/recipes/eggplant-tarte-tatin", dateCreated: "2023-07-29", content: "I love tarte tatin. This recipe is Asian-inspired and just hits the right spots for me. 2 eggplants, 3+ tbsp vegetable oil, 4 tbsp rice vinegar, 6 tbsp soy sauce, 4 tbsp caster sugar, 3 tbsp water, 5 cm piece fresh ginger, peeled and finely grated, 2 green chilies, finely sliced (or flaked chili), 2 garlic cloves, peeled and finely sliced, 1 pie dough, Juice of 1 lime"},{ id: 244, directory: "Recipes", title: "Eggplants with a tahini-miso sauce", slug: "/recipes/eggplants-with-tahini-miso", dateCreated: "2023-09-27", content: " 2 medium eggplants, 1.5 tbsp chili powder, 1.5 tbsp smoked paprika, 1.5 tsp sugar, 1 tsp salt, 1/4 cup olive oil"},{ id: 245, directory: "Recipes", title: "Endive pie", slug: "/recipes/endive-pie", dateCreated: "2022-11-12", content: " Dough for one pie, 3–4 endives, 1 apple, thinly sliced, 1 tsp sugar, 20 cl. milk, 1 egg"},{ id: 246, directory: "Recipes", title: "Fennel salad", slug: "/recipes/fennel-salad", dateCreated: "2022-06-06", content: " 1–2 fennel bulbs, thinly sliced, 1 tbsp mustard, 1 tbsp olive oil, 2 tbsp balsamico vinegar (or replace half with apple cider vinegar)"},{ id: 247, directory: "Recipes", title: "Fig pie", slug: "/recipes/fig-pie", dateCreated: "2023-09-01", content: " Figues, 100 g d'amande en poudre (ou noix, noisette), 40 g de sucre, 40 g de beurre, 1 oeuf"},{ id: 248, directory: "Recipes", title: "Flan", slug: "/recipes/flan", dateCreated: "2023-08-08", content: "L'une des premières recettes que j'ai appris à faire petite. 0.5 l de lait, 50 g de sucre, 3 oeufs"},{ id: 249, directory: "Recipes", title: "Freeze spinach leaves", slug: "/recipes/freeze-spinach-leaves", dateCreated: "2023-06-14", content: "It is currently full spring-summer here in Ardèche. We have a lot of amaranth plants growing as weeds among our garden vegetables. Instead of discarding them, the leaves are great to use as an alternative to spinach. Here is a good way to store them for winter. Lots of spinach leaves (or leaves from other similar plants), Ice cubes"},{ id: 250, directory: "Recipes", title: "Fresh cheese", slug: "/recipes/fresh-cheese", dateCreated: "2019-01-27", content: " 1 l milk, 1.5 tbsp white vinegar"},{ id: 251, directory: "Recipes", title: "Gaspacho", slug: "/recipes/gaspacho", dateCreated: "2023-07-30", content: " 6 tomates mûres, 2 poivrons, 2 oignons, 1 concombre, 1 gousse d'ail, 6 c.à.s de vinaigre, 1/2 verre d'huile d'olive, piment en poudre, poivre et sel"},{ id: 252, directory: "Recipes", title: "German Sauerkraut", slug: "/recipes/german-sauerkraut", dateCreated: "2022-03-06", content: " 4 tsp salt, 1 tsp fennel seeds, 1 tsp black pepper"},{ id: 253, directory: "Recipes", title: "German sweet red cabbage", slug: "/recipes/german-sweet-red-cabbage", dateCreated: "2022-10-06", content: " 1 small head of red cabbage, 1 cup apple cider vinegar, 1.5 cups water, 1/3 cup sugar, 8 cloves"},{ id: 254, directory: "Recipes", title: "German Tyrolean Knödel", slug: "/recipes/german-tyrolean-knoedel", dateCreated: "2022-12-04", content: " 200 g dry bread cubes (white part only), 2 eggs, 180 g milk (or water), 180 g spicy mountain cheese, 150 g butter, Salt, pepper, nutmeg"},{ id: 255, directory: "Recipes", title: "Greek tzatziki", slug: "/recipes/greek-tzatziki", dateCreated: "2023-04-27", content: " 1 cucumber, 4 to 5 garlic cloves, minced, 1 tsp white vinegar, Olive oil, 2 cups plain yogurt, strained (or thick greek yogurt)"},{ id: 256, directory: "Recipes", title: "Green tomato chutney", slug: "/recipes/green-tomato-chutney", dateCreated: "2022-10-06", content: " 1 kg green tomatoes, 1 kg red onions, 150 g raisins (or similar dried fruits), 2 tsp salt, 500 g sugar, 1 liter malt, mix A (3/4 cider vinegar, 1/4 balsamic vinegar) or mix B (50:50 rice and balsamic vinegar)"},{ id: 257, directory: "Recipes", title: "Hollandaise sauce", slug: "/recipes/hollandaise-sauce", dateCreated: "2023-02-04", content: " 1 jaune d'oeuf, 1 cuillère à café d'eau froide, 65 g de beurre (ou compléter avec de l'huile d'olive), Jus de citron (ou vinaigre de cidre), Poivre, Sel"},{ id: 258, directory: "Recipes", title: "Hummus", slug: "/recipes/hummus", dateCreated: "2023-04-27", content: " 1 cup dry chickpeas, 1–2 garlic cloves, 1/3 cup tahini, 1 lemon, juiced, Salt, Olive oil"},{ id: 259, directory: "Recipes", title: "Immunity shots", slug: "/recipes/immunity-shots", dateCreated: "2023-11-17", content: "Prepare these frozen immunity shots early winter to be ready for the cold season. 10 cm fresh ginger, 5 cm fresh tumeric, 2 lemons, roughly peeled, 2 oranges, roughly peeled, Black pepper (anti-inflammatory), Honey"},{ id: 260, directory: "Recipes", title: "Indian dal tadka", slug: "/recipes/indian-dal-tadka", dateCreated: "2020-12-24", content: " 1–2 cups of rinsed lentils (mix different types of lentils for a more diverse texture), 0.5 tsp turmeric powder, 1.5 tsp salt, 3.5 cups water, 4 large garlic cloves, crushed, 1 inch ginger, crushed, 1 green chili, crushed, 1.5 tbsp ghee (or oil), 1 tsp cumin seeds, 1 tbsp coriander seeds, crushed, 2 cloves, 1 medium red onion, chopped, 2 medium tomatoes chopped (or tomato puree / paste), 0.5 tsp coriander powder, ¼ tsp garam masala, ¼ tsp red chili powder, 2 tbsp cilantro, chopped, 1 tsp dried fenugreek leaves, crushed"},{ id: 261, directory: "Recipes", title: "Indian ghee", slug: "/recipes/indian-ghee", dateCreated: "2026-01-01", content: "Love the nutty taste and flavor! Bonus points as it can be made out of stale butter. 1 kg beurre à température ambiante"},{ id: 262, directory: "Recipes", title: "Indian pav bhaji", slug: "/recipes/indian-pav-bhaji", dateCreated: "2020-10-10", content: " 4 tbsp coriander seeds, 2 tbsp cumin seeds, 5 dried red chilies (adjust to taste), 8 cloves, 1 tbsp fennel seeds, 1/2 tsp black cardamon seeds (or 24 seeds), 5&nbsp;cm cinnamon stick, 1.5 tsp turmeric powder, 2 tsp ground black pepper, 1/2 tsp ground ginger, 2 tsp salt"},{ id: 263, directory: "Recipes", title: "Italian arancini", slug: "/recipes/italian-arancini", dateCreated: "2022-12-08", content: " Pre-cooked risotto, ideally from the day before"},{ id: 264, directory: "Recipes", title: "Italian bolognese", slug: "/recipes/italian-bolognese", dateCreated: "2023-01-01", content: "This is Valentina's family recipe. Her great-grandma used to make her own tomato paste from scratch. Her grandma reviewed the recipe with some store-bought tomato paste, which works just fine. The secret is chopping the vegetables into super tiny pieces and using tomato paste instead of tomato sauce. 250 g ground beef, 250 g ground pork, 3–4 celery stalks, finely chopped, 3–4 carrots, finely chopped, 1 big onion, finely chopped, 80–100 ml of white wine, 80 g butter + 2–3 spoons olive oil (or 100 g butter with no oil or just olive oil), Tomato paste, ideally the one with added celery and carrots"},{ id: 265, directory: "Recipes", title: "Italian cacio e pepe", slug: "/recipes/italian-cacio-e-pepe", dateCreated: "2021-02-25", content: "I love the one-pan hack of this super creamy cacio e pepe! Spaghetti, Water, Salt, Pecorino (or parmesan), Black pepper, freshly ground"},{ id: 266, directory: "Recipes", title: "Italian carbonara", slug: "/recipes/italian-carbonara", dateCreated: "2022-11-12", content: "Matteo and Valentina introduced Robin and I to an authentic carbonara recipe. It is apparently only one of many subtle variations, as they argued over many details during the cooking process. Just loved learning it with a glass of red wine! 1 pack of tonnarelli, spaghetti, mezze maniche or rigatoni, 200 g guanciale or rindless bacon, 1 cup grated Pecorino Romano (or Parmigiano Reggiano), 3 egg yolks + 1–2 entire eggs, Salt and black pepper"},{ id: 267, directory: "Recipes", title: "Italian gnocchi", slug: "/recipes/italian-gnocchi", dateCreated: "2019-05-24", content: " 500 g pomme de terre, 250 g farine, 1 jaune d'œuf, Sel, 1 pincée muscade, 1 noisette de beurre"},{ id: 268, directory: "Recipes", title: "Italian lasagna", slug: "/recipes/italian-lasagna", dateCreated: "2023-02-04", content: " 1 paquet de lasagnes sèches, 0.5 l de sauce bolognaise, 0.5 l de sauce béchamel, 1 oignon, émincé, 2 gousses d'ail, Poivre, Sel"},{ id: 269, directory: "Recipes", title: "Italian pesto alla trapanese", slug: "/recipes/italian-pesto-alla-trapanese", dateCreated: "2023-06-10", content: "I discovered this pasta sauce recipe, as I was traveling with friends through Sicily in 2023. It is fresh, tomato-ey, and just perfect for summer. 3 medium cloves garlic, A big handful of almonds, blanched and peeled, A bouquet of basil, leaves only, 100 g Pecorino cheese (plus more for serving), 500 g plum tomatoes (meaty ones), seeded, Olive oil, Busiate pasta (for the traditional Trapanese recipe, but any pasta will do), Salt"},{ id: 270, directory: "Recipes", title: "Italian ravioli", slug: "/recipes/italian-ravioli", dateCreated: "2022-10-11", content: " 220 g flour, 2 medium eggs, egg white-water mixture (only water works too)"},{ id: 271, directory: "Recipes", title: "Italian ribollita", slug: "/recipes/italian-ribollita", dateCreated: "2020-06-24", content: "Robin and I discovered ribollita during our tandem bike trip in the Chianti region. We actually rarely use the traditional ingredients, but we love the concept of using up any kinds of vegetables and old bread you have on hand to create a filling soup. ½ lb. dried cannellini beans (1 ⅓ cups), ⅓ cup side-virgin olive oil, ½ onion, 1 carrot, 1 celery stalk, 1 garlic clove, 2-3 thyme sprigs (leaves only), 1 tsp salt, 1 potato, medium sized, ½ lb. lacinato kale (roughly a bunch), ½ head Savoy cabbage (about ¾ lb.), ⅓ lb. Swiss chard, 1 tbsp tomato paste (or ⅓ cup of canned diced tomatoes or 1 fresh tomato), ½ lb. whole wheat, rustic, stale bread, To taste, freshly ground black pepper"},{ id: 272, directory: "Recipes", title: "Italian taralli", slug: "/recipes/italian-taralli", dateCreated: "2022-12-11", content: "I finally made my own taralli after trying them multiple times in Italian bars with Valentina, Matteo and Robin. 550 g plain flour, 10 g (or 2 tsp) salt, to taste, 120 g side virgin olive oil, 200 g dry white wine"},{ id: 273, directory: "Recipes", title: "Italian tiramisu", slug: "/recipes/italian-tiramisu", dateCreated: "2022-11-20", content: "Tiramisu must be one of my favorite desserts. Beware the day I'll own an espresso machine, I'll be making tiramisu all the time. 20 biscuits à la cuillère, 75 g sucre, 250 g mascarpone, 2–3 oeufs, 50 cl café fort refroidi"},{ id: 274, directory: "Recipes", title: "Italian veggie schiaccata", slug: "/recipes/italian-veggie-schiaccata", dateCreated: "2023-09-26", content: "Great to eat as a side or even as a main dish with some accompanying sauce. Some veggies (e.g. 2 carrots, 2 zucchini), grated, 1 potato, grated, 200 g flour, Salt, Olive oil"},{ id: 275, directory: "Recipes", title: "Jam", slug: "/recipes/jam", dateCreated: "2023-07-31", content: "I love Robin's family jam recipe. The result is not too sweet, syrupy with big pieces of fruits left. 5 kg fruit (e.g. apricots, prunes, cherries, etc.), 2.5 kg brown sugar"},{ id: 276, directory: "Recipes", title: "Japanese pickled ginger", slug: "/recipes/japanese-pickled-ginger", dateCreated: "2022-10-06", content: " 1 kg fresh, young ginger roots, 2 tsp salt, 3 cups rice vinegar, 2 cups sugar"},{ id: 277, directory: "Recipes", title: "Korean kimchi stew (kimchijigae)", slug: "/recipes/korean-kimchi-stew", dateCreated: "2019-01-27", content: " 1 pound kimchi, cut into bite size pieces, 1/4 cup kimchi brine, 1/2 pound pork shoulder (or pork belly), 1/2 package tofu, slices of 1&nbsp;cm thick, 3 green onions, 1 medium onion, sliced, 1 tsp salt, 2 tsp sugar, 2 tsp gochugaru (Korean hot pepper flakes), 1 tbsp gochujang (Korean hot pepper paste), 1 tsp toasted sesame oil, 2 cups anchovy stock"},{ id: 278, directory: "Recipes", title: "Korean kimchi", slug: "/recipes/korean-kimchi", dateCreated: "2022-02-10", content: "I love fermenting stuff. Kimchi is one of the first fermenting recipes I learned. It is the perfect addition to any dish you want to brighten or spice a bit. 700 g cabbage, chopped, 3 tbsp salt, water, ginger, minced, garlic, crushed, onion, minced, leek, chopped, green onion, diagonally sliced, carrot, julienned"},{ id: 279, directory: "Recipes", title: "Korean kimchijeon", slug: "/recipes/korean-kimchijeon", dateCreated: "2023-04-05", content: " 200 g kimchi, 2 tbsp kimchi brine, 3 scallions, chopped, 0.5 tsp sugar, 60 g flour, 100 g water (start with less, add if needed)"},{ id: 280, directory: "Recipes", title: "Korean side-spinach (sigeumchi namul)", slug: "/recipes/korean-side-spinach", dateCreated: "2020-11-15", content: " 8 ounces spinach, cleaned and washed, 1 garlic clove, minced, 1 green onion, chopped, 1.5 tsp soy sauce, 1.5 tsp toasted sesame oil, 2 tsp toasted sesame seeds"},{ id: 281, directory: "Recipes", title: "Korean steamed eggs (gyeran-jjim)", slug: "/recipes/korean-steamed-eggs", dateCreated: "2022-01-08", content: " 4 large eggs, 3/4 cup (180 g) water, 1 tsp salt"},{ id: 282, directory: "Recipes", title: "Lacto fermented veggies", slug: "/recipes/lacto-fermentated-veggies", dateCreated: "2022-08-15", content: " Veggies (cut into pieces), Water, Salt"},{ id: 283, directory: "Recipes", title: "Leek pie", slug: "/recipes/leek-pie", dateCreated: "2022-11-30", content: " 1 pâte brisée, 2 poireaux, 200 g crème fraîche ou yaourt, 2–4 oeufs, Poivre, sel, muscade"},{ id: 284, directory: "Recipes", title: "Lemon bars", slug: "/recipes/lemon-bars", dateCreated: "2021-01-21", content: " 250 g flour, 100 g sugar, 230 g butter (diced)"},{ id: 285, directory: "Recipes", title: "Lemon pie", slug: "/recipes/lemon-pie", dateCreated: "2022-03-14", content: " Pâte brisée précuite, 4 gros citrons bio (pour obtenir 150 ml de jus), 3 oeufs, 115 g de sucre, 30 g de beurre doux, 1 cuillère à soupe de maïzena"},{ id: 286, directory: "Recipes", title: "Lemon risotto", slug: "/recipes/lemon-risotto", dateCreated: "2022-03-04", content: " 2 shallots, 1 stick of celery, 60 g unsalted butter, 1 tbsp olive oil, 300 g risotto rice, 1 l vegetable stock, zest and juice of half a lemon, 2 small sprigs of fresh rosemary (finely chopped), 1 large egg yolk, 4 tbsp grated Parmesan cheese (plus more to sprinkle), 60 ml double cream, salt (to taste), pepper (preferably white)"},{ id: 287, directory: "Recipes", title: "Lentil galette", slug: "/recipes/lentil-galette", dateCreated: "2023-10-12", content: " 1 cup red lentils, 1 cup water (+ more for soaking lentils), 1 tbsp oil, Salt"},{ id: 288, directory: "Recipes", title: "Miso butter", slug: "/recipes/miso-butter", dateCreated: "2023-10-12", content: " 6 tbsp butter, cut into pieces (at room temperature), 3 tbsp miso"},{ id: 289, directory: "Recipes", title: "Orange loaf cake", slug: "/recipes/orange-loaf-cake", dateCreated: "2022-02-10", content: " 190 g all-purpose flour, 1 tsp baking powder (or Natron if we use some side sourdough starter), 0.5 tsp salt, 190 g sugar, 170 g butter, at room temperature, 2 large eggs, 5+ tbsp orange juice (or juice of half an orange), 1 orange zest, 250 ml sour cream"},{ id: 290, directory: "Recipes", title: "Overnight oats", slug: "/recipes/overnight-oats", dateCreated: "2022-11-22", content: " Oats, Oat milk or cow milk, Honey (to taste)"},{ id: 291, directory: "Recipes", title: "Pad thai", slug: "/recipes/pad-thai", dateCreated: "2023-09-22", content: " 40 g tamarind pulp, 1/2 cup boiling water, 3 tbsp sugar, 3 tbsp fish sauce, 2 tbsp dark soy sauce, Pepper (to taste)"},{ id: 292, directory: "Recipes", title: "Pain perdu", slug: "/recipes/pain-perdu", dateCreated: "2022-11-09", content: " 25 cl lait, 50 g sucre, 3 oeufs, 6 épaisses tranches de pain"},{ id: 293, directory: "Recipes", title: "Pasta dough", slug: "/recipes/pasta-dough", dateCreated: "2022-10-11", content: " 330 g all-purpose flour or ideally semolina (adjust until right consistency), >3 medium eggs"},{ id: 294, directory: "Recipes", title: "Pasta with bean sauce", slug: "/recipes/pasta-with-bean-sauce", dateCreated: "2020-06-24", content: "A pasta sauce made of beans only? Yep, it is apparently possible and surprisingly good. 0.5 cup side-virgin olive oil, 1 cup walnuts, finely chopped, 0.5 tsp crushed red pepper flakes, 5 garlic cloves — 1 whole, 4 thinly sliced, 1 sprigs rosemary, 1 can white beans, 1.5 cup large tubular pasta, salt, 1 lemon, 2 tbsp finely chopped parsley, ground black pepper"},{ id: 295, directory: "Recipes", title: "Pâte brisée", slug: "/recipes/pate-brisee", dateCreated: "2023-06-30", content: " 210 g blé, 115 g beurre, 0.5 c-à-c sel, 60 ml eau"},{ id: 296, directory: "Recipes", title: "Peach ice tea", slug: "/recipes/peach-ice-tea", dateCreated: "2023-08-08", content: " 3+ peaches, 2 tbsp black tea leaves, 1 l water"},{ id: 297, directory: "Recipes", title: "Persian rice pilaf (sabzi polo)", slug: "/recipes/persian-rice-pilaf", dateCreated: "2022-11-09", content: "One of my go-to recipes for when I have a lot of parsley I want to finish. 2 cups basmati rice, 8.25 cups water (to cook the rice), 5 cups fresh herbs (dill, parsley, cilantro), thinly chopped, 12 cloves garlic, coarsely chopped, 3 tbsp salt, 4 tbsp vegetable oil"},{ id: 298, directory: "Recipes", title: "Pie dough", slug: "/recipes/pie-dough", dateCreated: "2021-01-21", content: " 250 g farine T55 ou autre, 50 g huile, 80 g eau tiède, 1 pincée de sel"},{ id: 299, directory: "Recipes", title: "Polish vegetarian cabbage rolls", slug: "/recipes/polish-vegetarian-cabbage-rolls", dateCreated: "2021-01-19", content: " 1 large cabbage, 2 cups brown rice, cooked, 1 lb veggie crumbles, 0.5 cup onion, minced, 1 garlic clove, minced, 1 tsp ground ginger, 1 tbsp sweet paprika, 1 tbsp parsley, chopped, 1 tsp black pepper, 2 eggs, 0.5 cup vegetable broth, 2 tbsp olive oil, 1 can diced tomatoes, 1 can tomato soup"},{ id: 300, directory: "Recipes", title: "Potato gratin", slug: "/recipes/potato-gratin", dateCreated: "2022-10-31", content: " Potatoes, Milk, Flour, Butter, Cheese"},{ id: 301, directory: "Recipes", title: "Potato soup", slug: "/recipes/potato-soup", dateCreated: "2023-03-01", content: " 8 pommes de terre de taille moyenne, 5 carottes de taille moyenne, 40 g de chorizo (ou autre saucisse séchée), 4 poignées de feuilles d'épinards, 1 oignon, 1 gousse d'ail pressée, Bouillon, Huile, Sel et poivre"},{ id: 302, directory: "Recipes", title: "Puff pastry", slug: "/recipes/puff-pastry", dateCreated: "2022-12-15", content: "This puff pastry recipe doesn't result in as many layers as a store-bought dough, but it does the job with minimal effort. 250 g plain flour, 280 g butter, very cold, 0.5 tsp salt, 120 ml of ice-cold water"},{ id: 303, directory: "Recipes", title: "Rice pilaf", slug: "/recipes/rice-pilaf", dateCreated: "2023-02-03", content: " Long grain rice, 1 onion, sliced, 2 garlic cloves, sliced, Vegetables, diced, 0.5 tsp coriander, 1 tsp paprika, 1 tsp Aleppo pepper, 1 tsp ground turmeric"},{ id: 304, directory: "Recipes", title: "Roasted vegetables", slug: "/recipes/roasted-vegetables", dateCreated: "2022-11-22", content: " vegetables, Olive oil"},{ id: 305, directory: "Recipes", title: "Shortbread", slug: "/recipes/shortbread", dateCreated: "2021-02-14", content: " 240 g butter at room temperature, 120 g sugar, 360 g flour, 1 pod vanilla, seeds scraped out, 0.5 tsp flaky sea salt"},{ id: 306, directory: "Recipes", title: "Slovenian Sauerkraut stew", slug: "/recipes/slovenian-sauerkraut-stew", dateCreated: "2022-04-03", content: " 1 tbsp paprika, 1 tsp chili powder, 500 g sauerkraut, 4 potatoes, peeled and chopped bite-size, 2 carrots, peeled and chopped bite-size, 3 cloves garlic, sliced, 1 onion, diced, 2–3 pickled cucumbers, sliced, 1 stock cube, oil, pepper and salt"},{ id: 307, directory: "Recipes", title: "Solid deodorant", slug: "/recipes/solid-deodorant", dateCreated: "2021-04-11", content: "This is the DIY deodorant recipe I have been using for many years now. Simple and effective. 60 g huile de coco bio, 48 g bicarbonate de soude, 30 g fécule de maïs, 24 gouttes huiles essentielles (palmarosa, tea tree, lavender)"},{ id: 308, directory: "Recipes", title: "Solid shampoo", slug: "/recipes/solid-shampoo", dateCreated: "2023-10-15", content: " 50 g de SCI (Sodium Cocoyl Isethionate), 15 g  de poudre de plante adaptée à vos cheveux (ex : poudre de bhringaraj), 17 g de phase huileuse (huiles végétales ou beurres végétaux, ex : huile de coco), 15 g de phase aqueuse (hydrolat, infusion, macérat aqueux ou eau, ex : hydrolat de romarin à verbénone), 2 g d'acide citrique"},{ id: 309, directory: "Recipes", title: "Sourdough bagels", slug: "/recipes/sourdough-bagels", dateCreated: "2021-01-22", content: " 125 g sourdough starter, 225 g water, 40 g honey (or sugar), 10 g fine salt, 500 g flour, 1 tbsp sugar (for boiling water)"},{ id: 310, directory: "Recipes", title: "Sourdough burger buns", slug: "/recipes/sourdough-burger-buns", dateCreated: "2021-01-22", content: " 200 g levain, 300 g farine T45, 200 g farine semi-complète T110, 8 g sel, 16 g sucre, 200 g lait à température ambiante (ou eau), 60 g beurre mou, 1 oeuf"},{ id: 311, directory: "Recipes", title: "Sourdough naan", slug: "/recipes/sourdough-naan", dateCreated: "2023-09-18", content: " 150 g water, 80 g yogurt, 75 g sourdough starter, 350 g flour, 0.5 tsp salt, 1.5 tbs sugar, 1 tbsp olive oil"},{ id: 312, directory: "Recipes", title: "Stuffed vine leaves", slug: "/recipes/stuffed-vine-leaves", dateCreated: "2023-07-31", content: "France's recipe, perfect for quick out-of-the-freezer starters. 1 tasse d'huile d'olive, 2 c.à.s. de pignons de pin, 2 oignons émincés, 1/2 tasse de riz blanc, 1/4 tasse de raisins secs, sel et poivre, un peu de menthe, 1/2 tasse d'eau chaude, 30 feuilles de vigne fraîches, 1 citron"},{ id: 313, directory: "Recipes", title: "Sushi / gimbap", slug: "/recipes/sushi-gimbap", dateCreated: "2024-05-12", content: "On fait une pause de deux mois à Hanoi lors de notre voyage de l'Europe à l'Asie. On y loue un appart pas cher avec une toute petite cuisine à peine fonctionnelle. Que peut-on préparer avec un minimum de 'cuisine' nécessaire et avec des ingrédients disponibles ? Je me lance dans des rouleaux inspirés des sushis japonais et gimbaps coréens. 3 bols de riz, 3 c-à-s vinaigre de riz, 1.5 c-à-s sucre, Sel"},{ id: 314, directory: "Recipes", title: "Tarte tatin with summer veggies", slug: "/recipes/tarte-tatin-with-summer-veggies", dateCreated: "2023-07-24", content: " Pie dough, 2 tomatoes, 1 zucchini, thinly sliced, 1 onion, thinly sliced"},{ id: 315, directory: "Recipes", title: "Tarte tatin", slug: "/recipes/tarte-tatin", dateCreated: "2023-09-13", content: " 1 pâte à tarte (feuilletée de préférence), 90 g de sucre, 1 c-à-s de miel, 1 c-à-s de vinaigre de cidre ou citron, 30 g de beurre, 1 grosse pincée de sel, 1 kg de pommes, coupées en quartiers, 1 sachet de sucre vanillé, Cannelle"},{ id: 316, directory: "Recipes", title: "Thai red curry paste (prig gang ped)", slug: "/recipes/thai-red-curry-paste", dateCreated: "2022-07-10", content: " 1 tbsp cilantro roots (or stems), 1 tbsp coriander seeds, 0.5 tbsp cumin seeds, 3 dried whole chilis, 1.5 tbsp galangal roots, 35 g garlic, 1 stalk lemongrass, 0.5 tsp peppercorns, 1 tabs. salt, 80 g shallots, 1 tbsp keffir lime zest (or lime zest)"},{ id: 317, directory: "Recipes", title: "Turnips with honey and zaatar", slug: "/recipes/turnip-with-honey-and-zaatar", dateCreated: "2023-03-06", content: "France's recipe: great as a side, can be eaten hot or cold. 1 turnip, Honey, Zaatar"},{ id: 318, directory: "Recipes", title: "Vegetarian gravy (upgraded version)", slug: "/recipes/vegetarian-gravy-upgraded", dateCreated: "2022-12-06", content: " 1 tbsp oil, 3 cloves garlic, chopped, 1 onion, chopped, 1–2 small pre-boiled potatoes, 7 g dried mushrooms, 1/2 cup red wine, 1/4 tsp cayenne pepper, 1/4 tsp cumin, 1 tsp smoked paprika, 1.5 tbsp soy sauce, 1 cup vegetable broth, Shiitake mushrooms, to taste, 1 pinch of salt"},{ id: 319, directory: "Recipes", title: "Vegetarian gravy", slug: "/recipes/vegetarian-gravy", dateCreated: "2022-02-10", content: " 8 tbsp oil, 1 onion, chopped, 3 cloves garlic, minced, 4 tbsp all-purpose flour, 2 tbsp light soy sauce, 1 cup vegetable broth, salt, black pepper, ground"},{ id: 320, directory: "Recipes", title: "Veggie loaf", slug: "/recipes/veggie-loaf", dateCreated: "2022-11-20", content: "Recette de Souad, toute simple et goûtue. 110 g farine blanche, 110 g farine d'épeautre complète, 2 oeufs, 150 g lait, 40 g huile végétale, 0.5 cc bicarbonate, 1 cc vinaigre blanc, Sel et poivre"},{ id: 321, directory: "Recipes", title: "Vietnamese cháo gà", slug: "/recipes/vietnamese-chao-ga", dateCreated: "2022-11-30", content: "Reminescent of childhood memories. Chao was my mother go-to recipe whenever someone in the family would get sick. 1 yellow onion, 1 whole chicken, 1 tbsp salt, 1.5 tbsp sugar, 250 g uncooked jasmine rice, 5&nbsp;cm knob fresh ginger, peeled and finely chopped, 1.5 tbsp fish sauce, 2 tsp freshly ground black pepper"},{ id: 322, directory: "Recipes", title: "Vietnamese đồ chua", slug: "/recipes/vietnamese-do-chua", dateCreated: "2023-02-25", content: " Carrots, peeled and julienned, 2 cups water, 1 cup vinegar (can be a mix of rice, white or apple vinegar), 0.5 cup sugar"},{ id: 323, directory: "Recipes", title: "Watermelon salad", slug: "/recipes/watermelon-salad", dateCreated: "2023-09-14", content: "A refreshing sweet and savory salad for hot summer days. Watermelon, Feta, Fresh mint"},{ id: 324, directory: "Recipes", title: "Whole roasted cauliflower", slug: "/recipes/whole-roasted-cauliflower", dateCreated: "2022-10-06", content: " 2 small heads cauliflower, 2 btsp olive oil, 1.25 tsp flake salt"},{ id: 325, directory: "Recipes", title: "Zucchini soup", slug: "/recipes/zucchini-soup", dateCreated: "2023-07-06", content: " Huile d'olive, 1 kg courgette, 2 oignons, 2 verres de lait et / ou eau, Sel et poivre"},];
+export default [{ id: 0, directory: "Musings", title: "10 beginners mistakes to avoid when designing for print", subtitle: "I made them too and will probably make them again", slug: "/posts/10-beginners-mistakes-to-avoid-when-designing-for-print", dateCreated: "2019-05-27",  content: ""},{ id: 1, directory: "Musings", title: "#10 On obsessions", subtitle: "Update #10", slug: "/posts/10-on-obsessions", dateCreated: "2021-01-10",  content: ""},{ id: 2, directory: "Musings", title: "#11 Habits in the oven", subtitle: "Update #11", slug: "/posts/11-habits-in-the-oven", dateCreated: "2021-01-17",  content: ""},{ id: 3, directory: "Musings", title: "#12 Culinary experiments", subtitle: "Update #12", slug: "/posts/12-culinary-experiments", dateCreated: "2021-01-23",  content: ""},{ id: 4, directory: "Musings", title: "#13 Holiday at home", subtitle: "Update #13", slug: "/posts/13-holiday-at-home", dateCreated: "2021-01-30",  content: ""},{ id: 5, directory: "Musings", title: "#14 Books, books, books", subtitle: "Update #14", slug: "/posts/14-books-books-books", dateCreated: "2021-02-07",  content: ""},{ id: 6, directory: "Musings", title: "#15 Walk on the frozen canal", subtitle: "Update #15", slug: "/posts/15-walk-on-the-frozen-canal", dateCreated: "2021-02-14",  content: ""},{ id: 7, directory: "Musings", title: "#16 Sunbathing on planet Earth", subtitle: "Update #16", slug: "/posts/16-sunbathing-on-planet-earth", dateCreated: "2021-02-22",  content: ""},{ id: 8, directory: "Musings", title: "#17 Back on tracks", subtitle: "Update #17", slug: "/posts/17-back-on-tracks", dateCreated: "2022-11-18",  content: ""},{ id: 9, directory: "Musings", title: "#18 Friendsgiving and D&D", subtitle: "Update #18", slug: "/posts/18-friendsgiving-and-dnd", dateCreated: "2022-11-28",  content: ""},{ id: 10, directory: "Musings", title: "#19 Sprained ankle and interactive fiction", subtitle: "Update #19", slug: "/posts/19-sprained-ankle-and-interactive-fiction", dateCreated: "2022-12-05",  content: ""},{ id: 11, directory: "Musings", title: "#20 Getting ready for Christmas", subtitle: "Update #20", slug: "/posts/20-getting-ready-for-christmas", dateCreated: "2022-12-16",  content: ""},{ id: 12, directory: "Musings", title: "2022 doodling", subtitle: "Playing around with my graphics tablet", slug: "/posts/2022-doodling", dateCreated: "2022-12-29",  content: ""},{ id: 13, directory: "Musings", title: "#21 Berlin post-Christmas", subtitle: "Update #21", slug: "/posts/21-berlin-post-christmas", dateCreated: "2022-12-27",  content: ""},{ id: 14, directory: "Musings", title: "#22 Gently sliding into 2023", subtitle: "Update #22", slug: "/posts/22-gently-sliding-into-2023", dateCreated: "2023-01-02",  content: ""},{ id: 15, directory: "Musings", title: "#23 Galette des rois and IndieWeb nuggets", subtitle: "Update #23", slug: "/posts/23-galette-des-rois-and-indieweb-nuggets", dateCreated: "2023-01-09",  content: ""},{ id: 16, directory: "Musings", title: "#24 Playing with games and typography", subtitle: "Update #24", slug: "/posts/24-playing-with-games-and-typography", dateCreated: "2023-01-19",  content: ""},{ id: 17, directory: "Musings", title: "#25 Unpacking and wood stove fires", subtitle: "Update #25", slug: "/posts/25-unpacking-and-wood-stove-fires", dateCreated: "2023-02-04",  content: ""},{ id: 18, directory: "Musings", title: "#26 Taking the time", subtitle: "Update #26", slug: "/posts/26-taking-the-time", dateCreated: "2023-02-26",  content: ""},{ id: 19, directory: "Musings", title: "#27 (Real) rock climbing and more", subtitle: "Update #27", slug: "/posts/27-real-rock-climbing-and-more", dateCreated: "2023-03-06",  content: ""},{ id: 20, directory: "Musings", title: "#28 On learning", subtitle: "Update #28", slug: "/posts/28-on-learning", dateCreated: "2023-03-24",  content: ""},{ id: 21, directory: "Musings", title: "#29 Time for a break", subtitle: "Update #29", slug: "/posts/29-time-for-a-break", dateCreated: "2023-04-04",  content: ""},{ id: 22, directory: "Musings", title: "#3 Friends and digital gardening", subtitle: "Update #3", slug: "/posts/3-friends-and-digital-gardening", dateCreated: "2020-11-24",  content: ""},{ id: 23, directory: "Musings", title: "#30 Slow start", subtitle: "Update #30", slug: "/posts/30-slow-start", dateCreated: "2023-04-20",  content: ""},{ id: 24, directory: "Musings", title: "#31 Friends, sci-fi and Coppola", subtitle: "Update #31", slug: "/posts/31-friends-sci-fi-and-coppola", dateCreated: "2023-05-14",  content: ""},{ id: 25, directory: "Musings", title: "#32 On the train to Sicily", subtitle: "Update #32", slug: "/posts/32-on-the-train-to-sicily", dateCreated: "2023-05-21",  content: ""},{ id: 26, directory: "Musings", title: "#33 Post-Sicily", subtitle: "Update #33", slug: "/posts/33-post-sicily", dateCreated: "2023-06-05",  content: ""},{ id: 27, directory: "Musings", title: "#34 Mushroom hunting and playing with words", subtitle: "Update #34", slug: "/posts/34-mushroom-hunting-and-playing-with-words", dateCreated: "2023-06-19",  content: ""},{ id: 28, directory: "Musings", title: "#35 From one project to the next", subtitle: "Update #35", slug: "/posts/35-from-one-project-to-the-next", dateCreated: "2023-07-12",  content: ""},{ id: 29, directory: "Musings", title: "#36 Keeping up with the fruits and veggies", subtitle: "Update #36", slug: "/posts/36-keeping-up-with-the-fruits-and-veggies", dateCreated: "2023-08-08",  content: ""},{ id: 30, directory: "Musings", title: "#37 Bà ngoại", subtitle: "Update #37", slug: "/posts/37-ba-ngoai", dateCreated: "2023-09-20",  content: ""},{ id: 31, directory: "Musings", title: "#38 Hiking, eating, harvesting", subtitle: "Update #38", slug: "/posts/38-hiking-eating-harvesting", dateCreated: "2023-09-27",  content: ""},{ id: 32, directory: "Musings", title: "#39 A week in Hérault", subtitle: "Update #39", slug: "/posts/39-a-week-in-herault", dateCreated: "2023-10-10",  content: ""},{ id: 33, directory: "Musings", title: "#4 Glühwein and note-taking", subtitle: "Update #4", slug: "/posts/4-gluehwein-and-note-taking", dateCreated: "2020-11-28",  content: ""},{ id: 34, directory: "Musings", title: "#40 Brr", subtitle: "Update #40", slug: "/posts/40-brr", dateCreated: "2023-10-17",  content: ""},{ id: 35, directory: "Musings", title: "#41 Chestnut time", subtitle: "Update #41", slug: "/posts/41-chestnut-time", dateCreated: "2023-10-27",  content: ""},{ id: 36, directory: "Musings", title: "#42 The never-ending to-do list", subtitle: "Update #42", slug: "/posts/42-the-never-ending-to-do-list", dateCreated: "2023-11-23",  content: ""},{ id: 37, directory: "Musings", title: "#43 Fondues and digital storage", subtitle: "Update #43", slug: "/posts/43-fondues-and-digital-storage", dateCreated: "2023-12-08",  content: ""},{ id: 38, directory: "Musings", title: "#44 Family gatherings", subtitle: "Update #44", slug: "/posts/44-family-gatherings", dateCreated: "2023-12-24",  content: ""},{ id: 39, directory: "Musings", title: "#45 Turkish experiences", subtitle: "Update #45", slug: "/posts/45-turkish-experiences", dateCreated: "2024-01-07",  content: ""},{ id: 40, directory: "Musings", title: "#46 Three thousand kilometers", subtitle: "Update #46", slug: "/posts/46-three-thousand-kilometers", dateCreated: "2024-01-27",  content: ""},{ id: 41, directory: "Musings", title: "#47 Four months on the road", subtitle: "Update #47", slug: "/posts/47-four-months-on-the-road", dateCreated: "2024-04-30",  content: ""},{ id: 42, directory: "Musings", title: "#5 Coffee tasting and cozy December", subtitle: "Update #5", slug: "/posts/5-coffee-tasting-and-cozy-december", dateCreated: "2020-12-05",  content: ""},{ id: 43, directory: "Musings", title: "#6 Bringing home some old records", subtitle: "Update #6", slug: "/posts/6-bringing-home-some-old-records", dateCreated: "2020-12-12",  content: ""},{ id: 44, directory: "Musings", title: "#7 Living better with less", subtitle: "Update #7", slug: "/posts/7-living-better-with-less", dateCreated: "2020-12-19",  content: ""},{ id: 45, directory: "Musings", title: "#8 Remote Christmas", subtitle: "Update #8", slug: "/posts/8-remote-christmas", dateCreated: "2020-12-26",  content: ""},{ id: 46, directory: "Musings", title: "#9 New website for 2021", subtitle: "Update #9", slug: "/posts/9-new-website-for-2021", dateCreated: "2021-01-02",  content: ""},{ id: 47, directory: "Musings", title: "A list of Berlin things", subtitle: "A non-comprehensive list of things I'll miss from Berlin", slug: "/posts/a-list-of-berlin-things", dateCreated: "2022-12-31",  content: ""},{ id: 48, directory: "Musings", title: "A walk down the Sonnenallee", subtitle: "Along the Arab street, from Hermannplatz to Sonnenallee station", slug: "/posts/a-walk-down-the-sonnenallee", dateCreated: "2023-11-07",  content: ""},{ id: 49, directory: "Musings", title: "Add a crayon effect to your illustrations", subtitle: "An Adobe Illustrator CC tutorial", slug: "/posts/add-a-crayon-effect-to-your-illustrations", dateCreated: "2019-04-30",  content: ""},{ id: 50, directory: "Musings", title: "Choses que j'ai apprises à Berlin", subtitle: "Un exercise de poésie", slug: "/posts/ce-que-jai-appris-a-berlin", dateCreated: "2023-08-20",  content: ""},{ id: 51, directory: "Musings", title: "City kids going cottagecore", subtitle: "Moving to the countryside", slug: "/posts/city-kids-going-cottagecore", dateCreated: "2022-12-15",  content: ""},{ id: 52, directory: "Musings", title: "High tech", subtitle: "Comic strip #1", slug: "/posts/comic-high-tech", dateCreated: "2021-10-19",  content: ""},{ id: 53, directory: "Musings", title: "Obstacle", subtitle: "Comic strip #2", slug: "/posts/comic-obstacle", dateCreated: "2021-10-21",  content: ""},{ id: 54, directory: "Musings", title: "Create a duotone image in 5 minutes", subtitle: "An Adobe Photoshop CC tutorial", slug: "/posts/create-a-duotone-image-in-5-minutes", dateCreated: "2019-05-09",  content: ""},{ id: 55, directory: "Musings", title: "Dans la cabane", subtitle: "Un exercise de poésie", slug: "/posts/dans-la-cabane", dateCreated: "2023-06-16",  content: ""},{ id: 56, directory: "Musings", title: "Dans le quartier de Thanh Công il y a le village de Thanh Công (2004)", subtitle: "A tiny note about the short", slug: "/posts/dans-le-quartier-de-thanh-cong-il-y-a-le-village-de-thanh-cong-2004", dateCreated: "2023-04-02",  content: ""},{ id: 57, directory: "Musings", title: "Destination", subtitle: "A stop motion experiment", slug: "/posts/destination", dateCreated: "2022-02-13",  content: ""},{ id: 58, directory: "Musings", title: "Digital doodling", subtitle: "Trying out Tayasui Sketches", slug: "/posts/digital-doodling", dateCreated: "2021-12-25",  content: ""},{ id: 59, directory: "Musings", title: "Down By Law (1986)", subtitle: "A tiny note about the movie", slug: "/posts/down-by-law-1986", dateCreated: "2020-11-26",  content: ""},{ id: 60, directory: "Musings", title: "Fermentation: behind-the-scenes", subtitle: "Biochemistry stuff I barely understand", slug: "/posts/fermentation-behind-the-scenes", dateCreated: "2023-07-27",  content: ""},{ id: 61, directory: "Musings", title: "Free Solo (2018)", subtitle: "A tiny note about the movie", slug: "/posts/free-solo-2018", dateCreated: "2020-11-26",  content: ""},{ id: 62, directory: "Musings", title: "Gardening", subtitle: "Another try at knowledge management", slug: "/posts/gardening", dateCreated: "2021-10-24",  content: ""},{ id: 63, directory: "Musings", title: "I made a font", subtitle: "Or how I created a simple font as a complete typography-noob", slug: "/posts/i-made-a-font", dateCreated: "2023-01-10",  content: ""},{ id: 64, directory: "Musings", title: "In Bruges (2008)", subtitle: "A tiny note about the movie", slug: "/posts/in-bruges-2018", dateCreated: "2020-12-19",  content: ""},{ id: 65, directory: "Musings", title: "IndieWeb browsing", subtitle: "How I lost myself in the webring", slug: "/posts/indieweb-browsing", dateCreated: "2022-04-16",  content: ""},{ id: 66, directory: "Musings", title: "Interactive fiction", subtitle: "So I got into text adventure games", slug: "/posts/interactive-fiction", dateCreated: "2022-04-06",  content: ""},{ id: 67, directory: "Musings", title: "Just the two of us", subtitle: "A tiny note about life stuff", slug: "/posts/just-the-two-of-us", dateCreated: "2023-05-29",  content: ""},{ id: 68, directory: "Musings", title: "Last homage to my old website", subtitle: "A toast to all lost websites", slug: "/posts/last-homage-to-my-old-website", dateCreated: "2022-04-03",  content: ""},{ id: 69, directory: "Musings", title: "Peeking into the IndieWeb world", subtitle: "My first baby steps", slug: "/posts/my-first-indieweb-steps", dateCreated: "2022-04-10",  content: ""},{ id: 70, directory: "Musings", title: "My quest for true small caps", subtitle: "A web design adventure", slug: "/posts/my-quest-for-true-small-caps", dateCreated: "2021-01-07",  content: ""},{ id: 71, directory: "Musings", title: "Now archive #1", subtitle: "A snapshot in time", slug: "/posts/now-archive-1", dateCreated: "2021-08-21",  content: ""},{ id: 72, directory: "Musings", title: "Now archive #2", subtitle: "A snapshot in time", slug: "/posts/now-archive-2", dateCreated: "2021-10-24",  content: ""},{ id: 73, directory: "Musings", title: "Now archive #3", subtitle: "A snapshot in time", slug: "/posts/now-archive-3", dateCreated: "2022-02-20",  content: ""},{ id: 74, directory: "Musings", title: "Now archive #4", subtitle: "A snapshot in time", slug: "/posts/now-archive-4", dateCreated: "2022-06-25",  content: ""},{ id: 75, directory: "Musings", title: "Now archive #5", subtitle: "A snapshot in time", slug: "/posts/now-archive-5", dateCreated: "2022-11-04",  content: ""},{ id: 76, directory: "Musings", title: "Now archive #6", subtitle: "A snapshot in time", slug: "/posts/now-archive-6", dateCreated: "2023-02-06",  content: ""},{ id: 77, directory: "Musings", title: "Now archive #7", subtitle: "A snapshot in time", slug: "/posts/now-archive-7", dateCreated: "2023-05-17",  content: ""},{ id: 78, directory: "Musings", title: "Now archive #8", subtitle: "A snapshot in time", slug: "/posts/now-archive-8", dateCreated: "2023-07-12",  content: ""},{ id: 79, directory: "Musings", title: "Now archive #9", subtitle: "A snapshot in time", slug: "/posts/now-archive-9", dateCreated: "2023-11-17",  content: ""},{ id: 80, directory: "Musings", title: "Seven Psychopaths (2012)", subtitle: "A tiny note about the movie", slug: "/posts/seven-psychopaths-2012", dateCreated: "2020-11-26",  content: ""},{ id: 81, directory: "Musings", title: "Sicilian notes", subtitle: "A tiny note about our trip in Sicily", slug: "/posts/sicilian-notes", dateCreated: "2023-06-02",  content: ""},{ id: 82, directory: "Musings", title: "Sick days cheatsheet", subtitle: "Or how to help my tired cloudy mind", slug: "/posts/sick-days-cheatsheet", dateCreated: "2022-11-30",  content: ""},{ id: 83, directory: "Musings", title: "Split (2016)", subtitle: "A tiny note about the movie", slug: "/posts/split-2016", dateCreated: "2020-12-21",  content: ""},{ id: 84, directory: "Musings", title: "Styling animated rainbow links", subtitle: "With CSS only", slug: "/posts/styling-animated-rainbow-links", dateCreated: "2022-04-15",  content: ""},{ id: 85, directory: "Musings", title: "The Good, the Bad and the Ugly (1966)", subtitle: "A tiny note about the movie", slug: "/posts/the-good-the-bad-and-the-ugly-1966", dateCreated: "2020-12-05",  content: ""},{ id: 86, directory: "Musings", title: "The Guernsey Literary and Potato Peel Pie Society", subtitle: "A review of the audiobook", slug: "/posts/the-guernsey-literary-and-potato-peel-pie-society", dateCreated: "2022-04-09",  content: ""},{ id: 87, directory: "Musings", title: "The Host (2006)", subtitle: "A tiny note about the movie", slug: "/posts/the-host-2006", dateCreated: "2020-12-12",  content: ""},{ id: 88, directory: "Musings", title: "The Untouchables (1987)", subtitle: "A tiny note about the movie", slug: "/posts/the-untouchables-1987", dateCreated: "2020-12-11",  content: ""},{ id: 89, directory: "Musings", title: "Three years blogging", subtitle: "And apparently still going", slug: "/posts/three-years-blogging", dateCreated: "2023-11-26",  content: ""},{ id: 90, directory: "Musings", title: "UI/UX Learnings", subtitle: "Living notes", slug: "/posts/ui-ux-learnings", dateCreated: "2022-07-13",  content: ""},{ id: 91, directory: "Musings", title: "Victoria (2015)", subtitle: "A tiny note about the movie", slug: "/posts/victoria-2015", dateCreated: "2020-11-26",  content: ""},{ id: 92, directory: "Musings", title: "Waking up in Berlin", subtitle: "A last note", slug: "/posts/waking-up-in-berlin", dateCreated: "2023-01-29",  content: ""},{ id: 93, directory: "Musings", title: "Wonder Woman 1984 (2020)", subtitle: "A tiny note about the movie", slug: "/posts/wonder-woman-1984-2020", dateCreated: "2020-12-26",  content: ""},{ id: 94, directory: "Musings", title: "Words I'm growing suspicious of", subtitle: "Ranting (in a list format)", slug: "/posts/words-i-m-growing-suspicious-of", dateCreated: "2023-12-12",  content: ""},{ id: 95, directory: "Learning", title: "Stars or planets?", slug: "/learning/astro-1-stars-or-planets", dateCreated: "2023-11-15",  content: ""},{ id: 96, directory: "Learning", title: "Colloids: From mayonnaise to hairspray", slug: "/learning/cooking-1-colloids-from-mayonnaise-to-hairspray", dateCreated: "2023-09-30",  content: ""},{ id: 97, directory: "Learning", title: "Food pigments: Keep green food green", slug: "/learning/cooking-2-food-pigments-keep-green-food-green", dateCreated: "2023-09-30",  content: ""},{ id: 98, directory: "Learning", title: "Food pigments: Purple, blue or red blueberries", slug: "/learning/cooking-3-food-pigments-purple-blue-or-red-blueberries", dateCreated: "2023-09-30",  content: ""},{ id: 99, directory: "Learning", title: "Food pigments: Red beets", slug: "/learning/cooking-4-food-pigments-red-beets", dateCreated: "2023-09-30",  content: ""},{ id: 100, directory: "Learning", title: "How to cut open a pomegranate", slug: "/learning/cooking-5-how-to-cut-open-a-pomegranate", dateCreated: "2023-10-15",  content: ""},{ id: 101, directory: "Learning", title: "Data viz: Data visualization handbook", slug: "/learning/data-viz-1-data-visualization-handbook", dateCreated: "2022-05-05",  content: ""},{ id: 102, directory: "Learning", title: "Data viz: Introduction", slug: "/learning/data-viz-2-introduction", dateCreated: "2022-05-05",  content: ""},{ id: 103, directory: "Learning", title: "Data viz: General principles of visualization design", slug: "/learning/data-viz-3-general-principles-of-visualization-design", dateCreated: "2022-05-05",  content: ""},{ id: 104, directory: "Learning", title: "Géologie: Les familles de roches", slug: "/learning/geologie-1-les-familles-de-roches", dateCreated: "2023-12-09",  content: ""},{ id: 105, directory: "Learning", title: "Géologie : Les roches sédimentaires", slug: "/learning/geologie-2-les-roches-sedimentaires", dateCreated: "2023-12-09",  content: ""},{ id: 106, directory: "Learning", title: "Grande-voie: Le matériel", slug: "/learning/grande-voie-1-le-materiel", dateCreated: "2023-10-01",  content: ""},{ id: 107, directory: "Learning", title: "Grande-voie: Les manips au relai", slug: "/learning/grande-voie-2-les-manips-au-relai", dateCreated: "2023-10-01",  content: ""},{ id: 108, directory: "Learning", title: "Making solid shampoo", slug: "/learning/making-solid-shampoo", dateCreated: "2023-10-16",  content: ""},{ id: 109, directory: "Learning", title: "Middle dot", slug: "/learning/middle-dot", dateCreated: "2023-10-21",  content: ""},{ id: 110, directory: "Learning", title: "Nuages: Les bases", slug: "/learning/nuages-les-bases", dateCreated: "2023-11-27",  content: ""},{ id: 111, directory: "Learning", title: "Russian: Duolingoing", slug: "/learning/russian-1-duolingoing", dateCreated: "2023-09-29",  content: ""},{ id: 112, directory: "Learning", title: "Russian: To be or not to be", slug: "/learning/russian-2-to-be-or-not-to-be", dateCreated: "2023-09-29",  content: ""},{ id: 113, directory: "Learning", title: "Russian: Cases, the basics", slug: "/learning/russian-3-cases-the-basics", dateCreated: "2023-09-29",  content: ""},{ id: 114, directory: "Learning", title: "Russian: To have", slug: "/learning/russian-4-to-have", dateCreated: "2023-09-29",  content: ""},{ id: 115, directory: "Learning", title: "Russian: You, you, you", slug: "/learning/russian-5-you-you-you", dateCreated: "2023-09-29",  content: ""},{ id: 116, directory: "Learning", title: "Russian: About names", slug: "/learning/russian-6-about-names", dateCreated: "2023-09-29",  content: ""},{ id: 117, directory: "Learning", title: "Russian: Frequent words", slug: "/learning/russian-7-frequent-words", dateCreated: "2023-10-02",  content: ""},{ id: 118, directory: "Learning", title: "Russian: Hello", slug: "/learning/russian-8-hello", dateCreated: "2023-10-02",  content: ""},{ id: 119, directory: "Learning", title: "Stop motion: First steps", slug: "/learning/stop-motion-first-steps", dateCreated: "2022-02-13",  content: ""},{ id: 120, directory: "Media", title: "12 Angry Men", slug: "/media/12-angry-men", dateCreated: "2021-01-01",  content: ""},{ id: 121, directory: "Media", title: "1984", slug: "/media/1984", dateCreated: "2021-01-01",  content: ""},{ id: 122, directory: "Media", title: "A Gentleman in Moscow", slug: "/media/a-gentleman-in-moscow", dateCreated: "2022-05-01",  content: ""},{ id: 123, directory: "Media", title: "Adieu Bohème", slug: "/media/adieu-boheme", dateCreated: "2022-01-01",  content: ""},{ id: 124, directory: "Media", title: "Aftersun", slug: "/media/aftersun", dateCreated: "2022-12-01",  content: ""},{ id: 125, directory: "Media", title: "Ali and Nino", slug: "/media/ali-and-nino", dateCreated: "2024-02-20",  content: ""},{ id: 126, directory: "Media", title: "Amélie", slug: "/media/amelie", dateCreated: "2021-01-01",  content: ""},{ id: 127, directory: "Media", title: "And Then There Were None", slug: "/media/and-then-there-were-none", dateCreated: "2021-01-01",  content: ""},{ id: 128, directory: "Media", title: "Basara series", slug: "/media/basara", dateCreated: "2021-01-01",  content: ""},{ id: 129, directory: "Media", title: "Broken Earth trilogy", slug: "/media/broken-earth-trilogy", dateCreated: "2021-01-01",  content: ""},{ id: 130, directory: "Media", title: "Can't We Talk About Something More Pleasant?", slug: "/media/cant-we-talk-about-something-more-pleasant", dateCreated: "2023-07-31",  content: ""},{ id: 131, directory: "Media", title: "Captain Fantastic", slug: "/media/captain-fantastic", dateCreated: "2021-01-01",  content: ""},{ id: 132, directory: "Media", title: "Chroniques birmanes", slug: "/media/chroniques-birmanes", dateCreated: "2021-01-01",  content: ""},{ id: 133, directory: "Media", title: "Circe", slug: "/media/circe", dateCreated: "2021-01-01",  content: ""},{ id: 134, directory: "Media", title: "Coraline", slug: "/media/coraline", dateCreated: "2021-01-01",  content: ""},{ id: 135, directory: "Media", title: "Django Unchained", slug: "/media/django-unchained", dateCreated: "2021-01-01",  content: ""},{ id: 136, directory: "Media", title: "Do the Right Thing", slug: "/media/do-the-right-thing", dateCreated: "2021-01-01",  content: ""},{ id: 137, directory: "Media", title: "Dune", slug: "/media/dune", dateCreated: "2021-12-01",  content: ""},{ id: 138, directory: "Media", title: "Everything Everywhere All At Once", slug: "/media/everything-everywhere-all-at-once", dateCreated: "2022-12-01",  content: ""},{ id: 139, directory: "Media", title: "Ex Machina", slug: "/media/ex-machina", dateCreated: "2021-01-01",  content: ""},{ id: 140, directory: "Media", title: "Fight Club", slug: "/media/fight-club", dateCreated: "2021-01-01",  content: ""},{ id: 141, directory: "Media", title: "Freakonomics", slug: "/media/freakonomics", dateCreated: "2021-01-01",  content: ""},{ id: 142, directory: "Media", title: "Free Solo", slug: "/media/free-solo", dateCreated: "2021-01-01",  content: ""},{ id: 143, directory: "Media", title: "Get Out", slug: "/media/get-out", dateCreated: "2021-01-01",  content: ""},{ id: 144, directory: "Media", title: "Gone Girl", slug: "/media/gone-girl", dateCreated: "2021-01-01",  content: ""},{ id: 145, directory: "Media", title: "Guillermo del Toro's Pinocchio", slug: "/media/guillermo-del-toros-pinocchio", dateCreated: "2023-01-01",  content: ""},{ id: 146, directory: "Media", title: "Half of a Yellow Sun", slug: "/media/half-of-a-yellow-sun", dateCreated: "2021-01-01",  content: ""},{ id: 147, directory: "Media", title: "Harry Potter series", slug: "/media/harry-potter-series", dateCreated: "2021-01-01",  content: ""},{ id: 148, directory: "Media", title: "Her", slug: "/media/her", dateCreated: "2021-01-01",  content: ""},{ id: 149, directory: "Media", title: "His Dark Materials series", slug: "/media/his-dark-materials-series", dateCreated: "2021-01-01",  content: ""},{ id: 150, directory: "Media", title: "Hocus Pocus", slug: "/media/hocus-pocus", dateCreated: "2021-01-01",  content: ""},{ id: 151, directory: "Media", title: "Howl's Moving Castle", slug: "/media/howls-moving-castle", dateCreated: "2023-06-29",  content: ""},{ id: 152, directory: "Media", title: "Hunt for the Wilderpeople", slug: "/media/hunt-for-the-wilderpeople", dateCreated: "2021-01-01",  content: ""},{ id: 153, directory: "Media", title: "In the Mood for Love", slug: "/media/in-the-mood-for-love", dateCreated: "2023-09-20",  content: ""},{ id: 154, directory: "Media", title: "Inception", slug: "/media/inception", dateCreated: "2021-01-01",  content: ""},{ id: 155, directory: "Media", title: "Interstellar", slug: "/media/interstellar", dateCreated: "2021-01-01",  content: ""},{ id: 156, directory: "Media", title: "Joker", slug: "/media/joker", dateCreated: "2021-01-01",  content: ""},{ id: 157, directory: "Media", title: "Jumanji", slug: "/media/jumanji", dateCreated: "2021-01-01",  content: ""},{ id: 158, directory: "Media", title: "Kill Bill", slug: "/media/kill-bill", dateCreated: "2021-01-01",  content: ""},{ id: 159, directory: "Media", title: "Kirikou and the Sorceress", slug: "/media/kirikou-and-the-sorceress", dateCreated: "2021-01-01",  content: ""},{ id: 160, directory: "Media", title: "Kubo and the Two Strings", slug: "/media/kubo-and-the-two-strings", dateCreated: "2021-01-01",  content: ""},{ id: 161, directory: "Media", title: "La Horde de Contrevent", slug: "/media/la-horde-de-contrevent", dateCreated: "2021-01-01",  content: ""},{ id: 162, directory: "Media", title: "La La Land", slug: "/media/la-la-land", dateCreated: "2021-01-01",  content: ""},{ id: 163, directory: "Media", title: "La Quête d'Ewilan series", slug: "/media/la-quete-dewilan-series", dateCreated: "2021-01-01",  content: ""},{ id: 164, directory: "Media", title: "Le Grand Méchant Renard", slug: "/media/le-grand-mechant-renard", dateCreated: "2021-01-01",  content: ""},{ id: 165, directory: "Media", title: "L'échappée belle", slug: "/media/lechappee-belle", dateCreated: "2021-01-01",  content: ""},{ id: 166, directory: "Media", title: "Life of Pi", slug: "/media/life-of-pi-book", dateCreated: "2021-01-01",  content: ""},{ id: 167, directory: "Media", title: "Life of Pi", slug: "/media/life-of-pi", dateCreated: "2021-01-01",  content: ""},{ id: 168, directory: "Media", title: "Loki", slug: "/media/loki", dateCreated: "2021-12-01",  content: ""},{ id: 169, directory: "Media", title: "Lord of the Flies", slug: "/media/lord-of-the-flies", dateCreated: "2021-01-01",  content: ""},{ id: 170, directory: "Media", title: "Maniac", slug: "/media/maniac", dateCreated: "2021-01-01",  content: ""},{ id: 171, directory: "Media", title: "Mary Poppins", slug: "/media/mary-poppins", dateCreated: "2021-01-01",  content: ""},{ id: 172, directory: "Media", title: "Memento", slug: "/media/memento", dateCreated: "2021-01-01",  content: ""},{ id: 173, directory: "Media", title: "Memoirs of a Geisha", slug: "/media/memoirs-of-a-geisha", dateCreated: "2021-01-01",  content: ""},{ id: 174, directory: "Media", title: "Métaphysique des Tubes", slug: "/media/metaphysique-des-tubes", dateCreated: "2021-01-01",  content: ""},{ id: 175, directory: "Media", title: "Midsommar", slug: "/media/midsommar", dateCreated: "2021-01-01",  content: ""},{ id: 176, directory: "Media", title: "Moonrise Kingdom", slug: "/media/moonrise-kingdom", dateCreated: "2021-01-01",  content: ""},{ id: 177, directory: "Media", title: "Mulholland Drive", slug: "/media/mulholland-drive", dateCreated: "2024-05-12",  content: ""},{ id: 178, directory: "Media", title: "Mushi-Shi", slug: "/media/mushi-shi", dateCreated: "2021-01-01",  content: ""},{ id: 179, directory: "Media", title: "Nage libre", slug: "/media/nage-libre", dateCreated: "2021-01-01",  content: ""},{ id: 180, directory: "Media", title: "Never Let Me Go", slug: "/media/never-let-me-go", dateCreated: "2021-01-01",  content: ""},{ id: 181, directory: "Media", title: "Orphan black", slug: "/media/orphan-black", dateCreated: "2021-01-01",  content: ""},{ id: 182, directory: "Media", title: "Our flag means death", slug: "/media/our-flag-means-death", dateCreated: "2022-09-01",  content: ""},{ id: 183, directory: "Media", title: "Parasite", slug: "/media/parasite", dateCreated: "2021-01-01",  content: ""},{ id: 184, directory: "Media", title: "Perfect Days", slug: "/media/perfect-days", dateCreated: "2023-12-03",  content: ""},{ id: 185, directory: "Media", title: "Reservation Dogs", slug: "/media/reservation-dogs", dateCreated: "2022-12-01",  content: ""},{ id: 186, directory: "Media", title: "Schindler's List", slug: "/media/schindlers-list", dateCreated: "2021-01-01",  content: ""},{ id: 187, directory: "Media", title: "Sherlock", slug: "/media/sherlock", dateCreated: "2021-01-01",  content: ""},{ id: 188, directory: "Media", title: "Shutter Island", slug: "/media/shutter-island", dateCreated: "2021-01-01",  content: ""},{ id: 189, directory: "Media", title: "Soul", slug: "/media/soul", dateCreated: "2021-01-01",  content: ""},{ id: 190, directory: "Media", title: "Spotlight", slug: "/media/spotlight", dateCreated: "2021-01-01",  content: ""},{ id: 191, directory: "Media", title: "Still Life", slug: "/media/still-life", dateCreated: "2021-01-01",  content: ""},{ id: 192, directory: "Media", title: "Stories of Your Life and Others", slug: "/media/stories-of-your-life-and-others", dateCreated: "2021-01-01",  content: ""},{ id: 193, directory: "Media", title: "Stranger Things", slug: "/media/stranger-things", dateCreated: "2022-01-01",  content: ""},{ id: 194, directory: "Media", title: "Subaru series", slug: "/media/subaru-series", dateCreated: "2021-01-01",  content: ""},{ id: 195, directory: "Media", title: "The Artist", slug: "/media/the-artist", dateCreated: "2021-01-01",  content: ""},{ id: 196, directory: "Media", title: "The Beach", slug: "/media/the-beach", dateCreated: "2021-01-01",  content: ""},{ id: 197, directory: "Media", title: "The Godfather", slug: "/media/the-godfather", dateCreated: "2023-01-01",  content: ""},{ id: 198, directory: "Media", title: "The Great Escape", slug: "/media/the-great-escape", dateCreated: "2021-01-01",  content: ""},{ id: 199, directory: "Media", title: "The Great Gatsby", slug: "/media/the-great-gatsby", dateCreated: "2021-01-01",  content: ""},{ id: 200, directory: "Media", title: "The Guernsey Literary and Potato Peel Pie Society", slug: "/media/the-guernsey-literary-and-potato-peel-pie-society", dateCreated: "2021-01-01",  content: ""},{ id: 201, directory: "Media", title: "The Handmaiden", slug: "/media/the-handmaiden", dateCreated: "2021-01-01",  content: ""},{ id: 202, directory: "Media", title: "The Handmaid's Tale", slug: "/media/the-handmaids-tale", dateCreated: "2021-01-01",  content: ""},{ id: 203, directory: "Media", title: "The Martian", slug: "/media/the-martian", dateCreated: "2021-01-01",  content: ""},{ id: 204, directory: "Media", title: "The Matrix", slug: "/media/the-matrix", dateCreated: "2021-01-01",  content: ""},{ id: 205, directory: "Media", title: "The Murderbot Diaries", slug: "/media/the-murderbot-diaries", dateCreated: "2021-01-01",  content: ""},{ id: 206, directory: "Media", title: "The Night Circus", slug: "/media/the-night-circus", dateCreated: "2021-01-01",  content: ""},{ id: 207, directory: "Media", title: "The Omnivore's Dilemma", slug: "/media/the-omnivores-dilemma", dateCreated: "2021-01-01",  content: ""},{ id: 208, directory: "Media", title: "The Others", slug: "/media/the-others", dateCreated: "2021-01-01",  content: ""},{ id: 209, directory: "Media", title: "The Pianist", slug: "/media/the-pianist", dateCreated: "2021-01-01",  content: ""},{ id: 210, directory: "Media", title: "The Power of the Dog", slug: "/media/the-power-of-the-dog", dateCreated: "2021-12-01",  content: ""},{ id: 211, directory: "Media", title: "The Prestige", slug: "/media/the-prestige", dateCreated: "2021-01-01",  content: ""},{ id: 212, directory: "Media", title: "The Queen's Gambit", slug: "/media/the-queens-gambit", dateCreated: "2023-01-01",  content: ""},{ id: 213, directory: "Media", title: "The Shawshank Redemption", slug: "/media/the-shawshank-redemption", dateCreated: "2021-01-01",  content: ""},{ id: 214, directory: "Media", title: "The Thirteenth Tale", slug: "/media/the-thirteenth-tale", dateCreated: "2021-01-01",  content: ""},{ id: 215, directory: "Media", title: "The Three-Body Problem trilogy", slug: "/media/the-three-body-problem", dateCreated: "2021-01-01",  content: ""},{ id: 216, directory: "Media", title: "The Truman Show", slug: "/media/the-truman-show", dateCreated: "2021-01-01",  content: ""},{ id: 217, directory: "Media", title: "The Wonderful Story of Henry Sugar", slug: "/media/the-wonderful-story-of-henry-sugar", dateCreated: "2023-10-05",  content: ""},{ id: 218, directory: "Media", title: "Three Billboards Outside Ebbing, Missouri", slug: "/media/three-billboards-outside-ebbing-missouri", dateCreated: "2021-01-01",  content: ""},{ id: 219, directory: "Media", title: "Triangle of Sadness", slug: "/media/triangle-of-sadness", dateCreated: "2022-12-01",  content: ""},{ id: 220, directory: "Media", title: "WALL·E", slug: "/media/wall-e", dateCreated: "2021-01-01",  content: ""},{ id: 221, directory: "Media", title: "WandaVision", slug: "/media/wandavision", dateCreated: "2021-12-01",  content: ""},{ id: 222, directory: "Media", title: "What We Do in the Shadows", slug: "/media/what-we-do-in-the-shadows", dateCreated: "2021-01-01",  content: ""},{ id: 223, directory: "Media", title: "Where the mountain meets the moon", slug: "/media/where-the-mountain-meets-the-moon", dateCreated: "2024-03-13",  content: ""},{ id: 224, directory: "Media", title: "Wizard of Oz", slug: "/media/wizard-of-oz", dateCreated: "2021-01-01",  content: ""},{ id: 225, directory: "Media", title: "World of Tomorrow", slug: "/media/world-of-tomorrow", dateCreated: "2021-01-01",  content: ""},{ id: 226, directory: "Media", title: "Years and Years", slug: "/media/years-and-years", dateCreated: "2021-01-01",  content: ""},{ id: 227, directory: "Media", title: "Yotsuba&! series", slug: "/media/yotsuba-series", dateCreated: "2021-01-01",  content: ""},{ id: 228, directory: "Recipes", title: "Asparagus and baked potatoes", slug: "/recipes/asparagus-and-baked-potatoes", dateCreated: "2022-06-18", content: " 5 pommes de terre, 1 botte d'asperges blanches, 1 poignée de noix ou amandes, Miel, Gros sel, Flocons de piment"},{ id: 229, directory: "Recipes", title: "Banana bread", slug: "/recipes/banana-bread", dateCreated: "2021-02-05", content: "This is a recipe that Niamh recommended to me. And as with most dessert recipes, she was spot on. Definitely one of the best banana breads I have eaten. It has a very strong banana taste, is moist in the inside, but crispy on the outside. 115 g butter, cut into chunks, 160 g brown sugar, 5 medium-large mashed bananas, 2 large eggs, 1 tsp salt, 1.5 tsp ground cinnamon, 0.5 tsp ground nutmeg, 260 g flour, 1 tsp baking soda (Natron), >100 sourdough starter (or 1 tsp baking powder)"},{ id: 230, directory: "Recipes", title: "Banana ice cream", slug: "/recipes/banana-ice-cream", dateCreated: "2023-06-30", content: " 5 bananes bien mûres, Un peu de crème, de lait ou de yaourt"},{ id: 231, directory: "Recipes", title: "Basic millet", slug: "/recipes/basic-millet", dateCreated: "2022-10-06", content: " 200 g de millet (environ une tasse), 500 g d'eau"},{ id: 232, directory: "Recipes", title: "Beetroot carpaccio", slug: "/recipes/beetroot-carpaccio", dateCreated: "2023-03-09", content: " 1 beetroot, Olive oil, Flaked salt"},{ id: 233, directory: "Recipes", title: "Bell pepper spread", slug: "/recipes/bell-pepper-spread", dateCreated: "2022-10-31", content: " 1 bell pepper sliced in 8 pieces, 1 large handful of fresh herbs (e.g. cilantro, parsley, etc.), Olive oil, Salt, Lemon or vinegar"},{ id: 234, directory: "Recipes", title: "Brazilian crispy smashed potatoes", slug: "/recipes/brazilian-crispy-smashed-potatoes", dateCreated: "2022-10-06", content: " 20 small potatoes, 2 tbsp olive oil, 1.5 tsp salt"},{ id: 235, directory: "Recipes", title: "Brownies", slug: "/recipes/brownies", dateCreated: "2021-03-24", content: " 110 g butter melted and cooled, 200 g sugar, 2 large eggs, 25 g all-purpose flour, 60 g cocoa powder, 1 pinch salt, 1 pinch baking powder"},{ id: 236, directory: "Recipes", title: "Bulgarian national salad (shopska)", slug: "/recipes/bulgarian-national-salad", dateCreated: "2022-11-12", content: "I was introduced to this super simple salad recipe when traveling in Eastern Europe. A similar recipe exists in multiple of those countries, and always stars three main ingredients: cucumbers, tomatoes and feta. From there, everything else is optional. 1 cucumber, diced, 3 tomatoes, diced, 1 feta cheese, diced"},{ id: 237, directory: "Recipes", title: "Chinese chili oil", slug: "/recipes/chinese-chili-oil", dateCreated: "2022-04-29", content: " 2 cups of neutral oil, 1 cup or less of chili flakes"},{ id: 238, directory: "Recipes", title: "Clafoutis", slug: "/recipes/clafoutis", dateCreated: "2023-07-18", content: "Recette d'un vieux bouquin de cuisine de France. 500 g de fruits avec leurs noyaux (cerises, mirabelles, prunes, baies), 1 cuillère à soupe bombée de farine, 65 g de sucre, 1 pincée de sel, 120 g (12 cl) de lait, 2 oeufs"},{ id: 239, directory: "Recipes", title: "Cleaning spray for everything", slug: "/recipes/cleaning-spray-for-everything", dateCreated: "2023-02-13", content: " Bouteille à sprayer de 500 ml, 1,5 c-à-c. bicarbonate de soude, Eau chaude, 2 c-à-c. vinaigre blanc, 15–20 gouttes d'huiles essentielles (e.g. citron, eucalyptus, pin sylvestre)"},{ id: 240, directory: "Recipes", title: "Cold spicy noodles", slug: "/recipes/cold-spicy-noodles", dateCreated: "2023-08-25", content: " Wheat noodles, 4–6 tbsp oil, 2–3 tbsp gochugaru, 2 tbsp sesame seeds, 2 garlic cloves, cut small, 2 tsp honey, 2–3 tbsp soy sauce, 2 tbsp white vinegar, 3–4 tbsp sesame or peanut paste, 2+ tbsp water, Cucumber, fresh out of the fridge and julienned"},{ id: 241, directory: "Recipes", title: "Couscous", slug: "/recipes/couscous", dateCreated: "2023-10-29", content: "Recette de couscous de Souad, préparée ensemble pour la première fois à Joux en octobre 2023. Il fait légèrement frais dehors, on allume le poêle et on mange tous et toutes du même plat. Autour de la table : France, Souad, Greg, Karima, Thibaud, Robin et moi. Huile d'olive, 1 kg de semoule de blé, 600 g d'eau froide, salée, 2–3 oignons, Persil, Pois-chiches, trempés la veille, 2 boîtes de tomates, 1 petit pot de concentré de tomate, 1 petit verre de vin rouge, Ras el hanout, Sel et poivre, Légumes cuisson lente : carottes, navets, aubergine (coupés en gros morceaux), Légumes cuisson rapide : courgettes, poivrons (coupés en gros morceaux), Épaule d'agneau coupée en gros morceaux, Cuisses de poulet, Merguez"},{ id: 242, directory: "Recipes", title: "Crispy oven fries", slug: "/recipes/crispy-oven-fries", dateCreated: "2022-04-07", content: " 4 large dry and mealy Russet Potatoes, 1 tbsp vinegar, 4 tbsp peanut oil, Coarse salt"},{ id: 243, directory: "Recipes", title: "Croûtons", slug: "/recipes/croutons", dateCreated: "2023-04-20", content: " Old bread, 2 tbsp olive oil, Salt, 1 garlic clove, pressed"},{ id: 244, directory: "Recipes", title: "Dumpling soup with egg whites", slug: "/recipes/dumpling-soup-with-egg-whites", dateCreated: "2022-11-08", content: " 3 egg whites"},{ id: 245, directory: "Recipes", title: "Eggplant tarte tatin", slug: "/recipes/eggplant-tarte-tatin", dateCreated: "2023-07-29", content: "I love tarte tatin. This recipe is Asian-inspired and just hits the right spots for me. 2 eggplants, 3+ tbsp vegetable oil, 4 tbsp rice vinegar, 6 tbsp soy sauce, 4 tbsp caster sugar, 3 tbsp water, 5 cm piece fresh ginger, peeled and finely grated, 2 green chilies, finely sliced (or flaked chili), 2 garlic cloves, peeled and finely sliced, 1 pie dough, Juice of 1 lime"},{ id: 246, directory: "Recipes", title: "Eggplants with a tahini-miso sauce", slug: "/recipes/eggplants-with-tahini-miso", dateCreated: "2023-09-27", content: " 2 medium eggplants, 1.5 tbsp chili powder, 1.5 tbsp smoked paprika, 1.5 tsp sugar, 1 tsp salt, 1/4 cup olive oil"},{ id: 247, directory: "Recipes", title: "Endive pie", slug: "/recipes/endive-pie", dateCreated: "2022-11-12", content: " Dough for one pie, 3–4 endives, 1 apple, thinly sliced, 1 tsp sugar, 20 cl. milk, 1 egg"},{ id: 248, directory: "Recipes", title: "Fennel salad", slug: "/recipes/fennel-salad", dateCreated: "2022-06-06", content: " 1–2 fennel bulbs, thinly sliced, 1 tbsp mustard, 1 tbsp olive oil, 2 tbsp balsamico vinegar (or replace half with apple cider vinegar)"},{ id: 249, directory: "Recipes", title: "Fig pie", slug: "/recipes/fig-pie", dateCreated: "2023-09-01", content: " Figues, 100 g d'amande en poudre (ou noix, noisette), 40 g de sucre, 40 g de beurre, 1 oeuf"},{ id: 250, directory: "Recipes", title: "Flan", slug: "/recipes/flan", dateCreated: "2023-08-08", content: "L'une des premières recettes que j'ai appris à faire petite. 0.5 l de lait, 50 g de sucre, 3 oeufs"},{ id: 251, directory: "Recipes", title: "Freeze spinach leaves", slug: "/recipes/freeze-spinach-leaves", dateCreated: "2023-06-14", content: "It is currently full spring-summer here in Ardèche. We have a lot of amaranth plants growing as weeds among our garden vegetables. Instead of discarding them, the leaves are great to use as an alternative to spinach. Here is a good way to store them for winter. Lots of spinach leaves (or leaves from other similar plants), Ice cubes"},{ id: 252, directory: "Recipes", title: "Fresh cheese", slug: "/recipes/fresh-cheese", dateCreated: "2019-01-27", content: " 1 l milk, 1.5 tbsp white vinegar"},{ id: 253, directory: "Recipes", title: "Gaspacho", slug: "/recipes/gaspacho", dateCreated: "2023-07-30", content: " 6 tomates mûres, 2 poivrons, 2 oignons, 1 concombre, 1 gousse d'ail, 6 c.à.s de vinaigre, 1/2 verre d'huile d'olive, piment en poudre, poivre et sel"},{ id: 254, directory: "Recipes", title: "German Sauerkraut", slug: "/recipes/german-sauerkraut", dateCreated: "2022-03-06", content: " 4 tsp salt, 1 tsp fennel seeds, 1 tsp black pepper"},{ id: 255, directory: "Recipes", title: "German sweet red cabbage", slug: "/recipes/german-sweet-red-cabbage", dateCreated: "2022-10-06", content: " 1 small head of red cabbage, 1 cup apple cider vinegar, 1.5 cups water, 1/3 cup sugar, 8 cloves"},{ id: 256, directory: "Recipes", title: "German Tyrolean Knödel", slug: "/recipes/german-tyrolean-knoedel", dateCreated: "2022-12-04", content: " 200 g dry bread cubes (white part only), 2 eggs, 180 g milk (or water), 180 g spicy mountain cheese, 150 g butter, Salt, pepper, nutmeg"},{ id: 257, directory: "Recipes", title: "Greek tzatziki", slug: "/recipes/greek-tzatziki", dateCreated: "2023-04-27", content: " 1 cucumber, 4 to 5 garlic cloves, minced, 1 tsp white vinegar, Olive oil, 2 cups plain yogurt, strained (or thick greek yogurt)"},{ id: 258, directory: "Recipes", title: "Green tomato chutney", slug: "/recipes/green-tomato-chutney", dateCreated: "2022-10-06", content: " 1 kg green tomatoes, 1 kg red onions, 150 g raisins (or similar dried fruits), 2 tsp salt, 500 g sugar, 1 liter malt, mix A (3/4 cider vinegar, 1/4 balsamic vinegar) or mix B (50:50 rice and balsamic vinegar)"},{ id: 259, directory: "Recipes", title: "Hollandaise sauce", slug: "/recipes/hollandaise-sauce", dateCreated: "2023-02-04", content: " 1 jaune d'oeuf, 1 cuillère à café d'eau froide, 65 g de beurre (ou compléter avec de l'huile d'olive), Jus de citron (ou vinaigre de cidre), Poivre, Sel"},{ id: 260, directory: "Recipes", title: "Hummus", slug: "/recipes/hummus", dateCreated: "2023-04-27", content: " 1 cup dry chickpeas, 1–2 garlic cloves, 1/3 cup tahini, 1 lemon, juiced, Salt, Olive oil"},{ id: 261, directory: "Recipes", title: "Immunity shots", slug: "/recipes/immunity-shots", dateCreated: "2023-11-17", content: "Prepare these frozen immunity shots early winter to be ready for the cold season. 10 cm fresh ginger, 5 cm fresh tumeric, 2 lemons, roughly peeled, 2 oranges, roughly peeled, Black pepper (anti-inflammatory), Honey"},{ id: 262, directory: "Recipes", title: "Indian dal tadka", slug: "/recipes/indian-dal-tadka", dateCreated: "2020-12-24", content: " 1–2 cups of rinsed lentils (mix different types of lentils for a more diverse texture), 0.5 tsp turmeric powder, 1.5 tsp salt, 3.5 cups water, 4 large garlic cloves, crushed, 1 inch ginger, crushed, 1 green chili, crushed, 1.5 tbsp ghee (or oil), 1 tsp cumin seeds, 1 tbsp coriander seeds, crushed, 2 cloves, 1 medium red onion, chopped, 2 medium tomatoes chopped (or tomato puree / paste), 0.5 tsp coriander powder, ¼ tsp garam masala, ¼ tsp red chili powder, 2 tbsp cilantro, chopped, 1 tsp dried fenugreek leaves, crushed"},{ id: 263, directory: "Recipes", title: "Indian ghee", slug: "/recipes/indian-ghee", dateCreated: "2026-01-01", content: "Love the nutty taste and flavor! Bonus points as it can be made out of stale butter. 1 kg beurre à température ambiante"},{ id: 264, directory: "Recipes", title: "Indian pav bhaji", slug: "/recipes/indian-pav-bhaji", dateCreated: "2020-10-10", content: " 4 tbsp coriander seeds, 2 tbsp cumin seeds, 5 dried red chilies (adjust to taste), 8 cloves, 1 tbsp fennel seeds, 1/2 tsp black cardamon seeds (or 24 seeds), 5&nbsp;cm cinnamon stick, 1.5 tsp turmeric powder, 2 tsp ground black pepper, 1/2 tsp ground ginger, 2 tsp salt"},{ id: 265, directory: "Recipes", title: "Italian arancini", slug: "/recipes/italian-arancini", dateCreated: "2022-12-08", content: " Pre-cooked risotto, ideally from the day before"},{ id: 266, directory: "Recipes", title: "Italian bolognese", slug: "/recipes/italian-bolognese", dateCreated: "2023-01-01", content: "This is Valentina's family recipe. Her great-grandma used to make her own tomato paste from scratch. Her grandma reviewed the recipe with some store-bought tomato paste, which works just fine. The secret is chopping the vegetables into super tiny pieces and using tomato paste instead of tomato sauce. 250 g ground beef, 250 g ground pork, 3–4 celery stalks, finely chopped, 3–4 carrots, finely chopped, 1 big onion, finely chopped, 80–100 ml of white wine, 80 g butter + 2–3 spoons olive oil (or 100 g butter with no oil or just olive oil), Tomato paste, ideally the one with added celery and carrots"},{ id: 267, directory: "Recipes", title: "Italian cacio e pepe", slug: "/recipes/italian-cacio-e-pepe", dateCreated: "2021-02-25", content: "I love the one-pan hack of this super creamy cacio e pepe! Spaghetti, Water, Salt, Pecorino (or parmesan), Black pepper, freshly ground"},{ id: 268, directory: "Recipes", title: "Italian carbonara", slug: "/recipes/italian-carbonara", dateCreated: "2022-11-12", content: "Matteo and Valentina introduced Robin and I to an authentic carbonara recipe. It is apparently only one of many subtle variations, as they argued over many details during the cooking process. Just loved learning it with a glass of red wine! 1 pack of tonnarelli, spaghetti, mezze maniche or rigatoni, 200 g guanciale or rindless bacon, 1 cup grated Pecorino Romano (or Parmigiano Reggiano), 3 egg yolks + 1–2 entire eggs, Salt and black pepper"},{ id: 269, directory: "Recipes", title: "Italian gnocchi", slug: "/recipes/italian-gnocchi", dateCreated: "2019-05-24", content: " 500 g pomme de terre, 250 g farine, 1 jaune d'œuf, Sel, 1 pincée muscade, 1 noisette de beurre"},{ id: 270, directory: "Recipes", title: "Italian lasagna", slug: "/recipes/italian-lasagna", dateCreated: "2023-02-04", content: " 1 paquet de lasagnes sèches, 0.5 l de sauce bolognaise, 0.5 l de sauce béchamel, 1 oignon, émincé, 2 gousses d'ail, Poivre, Sel"},{ id: 271, directory: "Recipes", title: "Italian pesto alla trapanese", slug: "/recipes/italian-pesto-alla-trapanese", dateCreated: "2023-06-10", content: "I discovered this pasta sauce recipe, as I was traveling with friends through Sicily in 2023. It is fresh, tomato-ey, and just perfect for summer. 3 medium cloves garlic, A big handful of almonds, blanched and peeled, A bouquet of basil, leaves only, 100 g Pecorino cheese (plus more for serving), 500 g plum tomatoes (meaty ones), seeded, Olive oil, Busiate pasta (for the traditional Trapanese recipe, but any pasta will do), Salt"},{ id: 272, directory: "Recipes", title: "Italian ravioli", slug: "/recipes/italian-ravioli", dateCreated: "2022-10-11", content: " 220 g flour, 2 medium eggs, egg white-water mixture (only water works too)"},{ id: 273, directory: "Recipes", title: "Italian ribollita", slug: "/recipes/italian-ribollita", dateCreated: "2020-06-24", content: "Robin and I discovered ribollita during our tandem bike trip in the Chianti region. We actually rarely use the traditional ingredients, but we love the concept of using up any kinds of vegetables and old bread you have on hand to create a filling soup. ½ lb. dried cannellini beans (1 ⅓ cups), ⅓ cup side-virgin olive oil, ½ onion, 1 carrot, 1 celery stalk, 1 garlic clove, 2-3 thyme sprigs (leaves only), 1 tsp salt, 1 potato, medium sized, ½ lb. lacinato kale (roughly a bunch), ½ head Savoy cabbage (about ¾ lb.), ⅓ lb. Swiss chard, 1 tbsp tomato paste (or ⅓ cup of canned diced tomatoes or 1 fresh tomato), ½ lb. whole wheat, rustic, stale bread, To taste, freshly ground black pepper"},{ id: 274, directory: "Recipes", title: "Italian taralli", slug: "/recipes/italian-taralli", dateCreated: "2022-12-11", content: "I finally made my own taralli after trying them multiple times in Italian bars with Valentina, Matteo and Robin. 550 g plain flour, 10 g (or 2 tsp) salt, to taste, 120 g side virgin olive oil, 200 g dry white wine"},{ id: 275, directory: "Recipes", title: "Italian tiramisu", slug: "/recipes/italian-tiramisu", dateCreated: "2022-11-20", content: "Tiramisu must be one of my favorite desserts. Beware the day I'll own an espresso machine, I'll be making tiramisu all the time. 20 biscuits à la cuillère, 75 g sucre, 250 g mascarpone, 2–3 oeufs, 50 cl café fort refroidi"},{ id: 276, directory: "Recipes", title: "Italian veggie schiaccata", slug: "/recipes/italian-veggie-schiaccata", dateCreated: "2023-09-26", content: "Great to eat as a side or even as a main dish with some accompanying sauce. Some veggies (e.g. 2 carrots, 2 zucchini), grated, 1 potato, grated, 200 g flour, Salt, Olive oil"},{ id: 277, directory: "Recipes", title: "Jam", slug: "/recipes/jam", dateCreated: "2023-07-31", content: "I love Robin's family jam recipe. The result is not too sweet, syrupy with big pieces of fruits left. 5 kg fruit (e.g. apricots, prunes, cherries, etc.), 2.5 kg brown sugar"},{ id: 278, directory: "Recipes", title: "Japanese pickled ginger", slug: "/recipes/japanese-pickled-ginger", dateCreated: "2022-10-06", content: " 1 kg fresh, young ginger roots, 2 tsp salt, 3 cups rice vinegar, 2 cups sugar"},{ id: 279, directory: "Recipes", title: "Korean kimchi stew (kimchijigae)", slug: "/recipes/korean-kimchi-stew", dateCreated: "2019-01-27", content: " 1 pound kimchi, cut into bite size pieces, 1/4 cup kimchi brine, 1/2 pound pork shoulder (or pork belly), 1/2 package tofu, slices of 1&nbsp;cm thick, 3 green onions, 1 medium onion, sliced, 1 tsp salt, 2 tsp sugar, 2 tsp gochugaru (Korean hot pepper flakes), 1 tbsp gochujang (Korean hot pepper paste), 1 tsp toasted sesame oil, 2 cups anchovy stock"},{ id: 280, directory: "Recipes", title: "Korean kimchi", slug: "/recipes/korean-kimchi", dateCreated: "2022-02-10", content: "I love fermenting stuff. Kimchi is one of the first fermenting recipes I learned. It is the perfect addition to any dish you want to brighten or spice a bit. 700 g cabbage, chopped, 3 tbsp salt, water, ginger, minced, garlic, crushed, onion, minced, leek, chopped, green onion, diagonally sliced, carrot, julienned"},{ id: 281, directory: "Recipes", title: "Korean kimchijeon", slug: "/recipes/korean-kimchijeon", dateCreated: "2023-04-05", content: " 200 g kimchi, 2 tbsp kimchi brine, 3 scallions, chopped, 0.5 tsp sugar, 60 g flour, 100 g water (start with less, add if needed)"},{ id: 282, directory: "Recipes", title: "Korean side-spinach (sigeumchi namul)", slug: "/recipes/korean-side-spinach", dateCreated: "2020-11-15", content: " 8 ounces spinach, cleaned and washed, 1 garlic clove, minced, 1 green onion, chopped, 1.5 tsp soy sauce, 1.5 tsp toasted sesame oil, 2 tsp toasted sesame seeds"},{ id: 283, directory: "Recipes", title: "Korean steamed eggs (gyeran-jjim)", slug: "/recipes/korean-steamed-eggs", dateCreated: "2022-01-08", content: " 4 large eggs, 3/4 cup (180 g) water, 1 tsp salt"},{ id: 284, directory: "Recipes", title: "Lacto fermented veggies", slug: "/recipes/lacto-fermentated-veggies", dateCreated: "2022-08-15", content: " Veggies (cut into pieces), Water, Salt"},{ id: 285, directory: "Recipes", title: "Leek pie", slug: "/recipes/leek-pie", dateCreated: "2022-11-30", content: " 1 pâte brisée, 2 poireaux, 200 g crème fraîche ou yaourt, 2–4 oeufs, Poivre, sel, muscade"},{ id: 286, directory: "Recipes", title: "Lemon bars", slug: "/recipes/lemon-bars", dateCreated: "2021-01-21", content: " 250 g flour, 100 g sugar, 230 g butter (diced)"},{ id: 287, directory: "Recipes", title: "Lemon pie", slug: "/recipes/lemon-pie", dateCreated: "2022-03-14", content: " Pâte brisée précuite, 4 gros citrons bio (pour obtenir 150 ml de jus), 3 oeufs, 115 g de sucre, 30 g de beurre doux, 1 cuillère à soupe de maïzena"},{ id: 288, directory: "Recipes", title: "Lemon risotto", slug: "/recipes/lemon-risotto", dateCreated: "2022-03-04", content: " 2 shallots, 1 stick of celery, 60 g unsalted butter, 1 tbsp olive oil, 300 g risotto rice, 1 l vegetable stock, zest and juice of half a lemon, 2 small sprigs of fresh rosemary (finely chopped), 1 large egg yolk, 4 tbsp grated Parmesan cheese (plus more to sprinkle), 60 ml double cream, salt (to taste), pepper (preferably white)"},{ id: 289, directory: "Recipes", title: "Lentil galette", slug: "/recipes/lentil-galette", dateCreated: "2023-10-12", content: " 1 cup red lentils, 1 cup water (+ more for soaking lentils), 1 tbsp oil, Salt"},{ id: 290, directory: "Recipes", title: "Miso butter", slug: "/recipes/miso-butter", dateCreated: "2023-10-12", content: " 6 tbsp butter, cut into pieces (at room temperature), 3 tbsp miso"},{ id: 291, directory: "Recipes", title: "Orange loaf cake", slug: "/recipes/orange-loaf-cake", dateCreated: "2022-02-10", content: " 190 g all-purpose flour, 1 tsp baking powder (or Natron if we use some side sourdough starter), 0.5 tsp salt, 190 g sugar, 170 g butter, at room temperature, 2 large eggs, 5+ tbsp orange juice (or juice of half an orange), 1 orange zest, 250 ml sour cream"},{ id: 292, directory: "Recipes", title: "Overnight oats", slug: "/recipes/overnight-oats", dateCreated: "2022-11-22", content: " Oats, Oat milk or cow milk, Honey (to taste)"},{ id: 293, directory: "Recipes", title: "Pad thai", slug: "/recipes/pad-thai", dateCreated: "2023-09-22", content: " 40 g tamarind pulp, 1/2 cup boiling water, 3 tbsp sugar, 3 tbsp fish sauce, 2 tbsp dark soy sauce, Pepper (to taste)"},{ id: 294, directory: "Recipes", title: "Pain perdu", slug: "/recipes/pain-perdu", dateCreated: "2022-11-09", content: " 25 cl lait, 50 g sucre, 3 oeufs, 6 épaisses tranches de pain"},{ id: 295, directory: "Recipes", title: "Pasta dough", slug: "/recipes/pasta-dough", dateCreated: "2022-10-11", content: " 330 g all-purpose flour or ideally semolina (adjust until right consistency), >3 medium eggs"},{ id: 296, directory: "Recipes", title: "Pasta with bean sauce", slug: "/recipes/pasta-with-bean-sauce", dateCreated: "2020-06-24", content: "A pasta sauce made of beans only? Yep, it is apparently possible and surprisingly good. 0.5 cup side-virgin olive oil, 1 cup walnuts, finely chopped, 0.5 tsp crushed red pepper flakes, 5 garlic cloves — 1 whole, 4 thinly sliced, 1 sprigs rosemary, 1 can white beans, 1.5 cup large tubular pasta, salt, 1 lemon, 2 tbsp finely chopped parsley, ground black pepper"},{ id: 297, directory: "Recipes", title: "Pâte brisée", slug: "/recipes/pate-brisee", dateCreated: "2023-06-30", content: " 210 g blé, 115 g beurre, 0.5 c-à-c sel, 60 ml eau"},{ id: 298, directory: "Recipes", title: "Peach ice tea", slug: "/recipes/peach-ice-tea", dateCreated: "2023-08-08", content: " 3+ peaches, 2 tbsp black tea leaves, 1 l water"},{ id: 299, directory: "Recipes", title: "Persian rice pilaf (sabzi polo)", slug: "/recipes/persian-rice-pilaf", dateCreated: "2022-11-09", content: "One of my go-to recipes for when I have a lot of parsley I want to finish. 2 cups basmati rice, 8.25 cups water (to cook the rice), 5 cups fresh herbs (dill, parsley, cilantro), thinly chopped, 12 cloves garlic, coarsely chopped, 3 tbsp salt, 4 tbsp vegetable oil"},{ id: 300, directory: "Recipes", title: "Pie dough", slug: "/recipes/pie-dough", dateCreated: "2021-01-21", content: " 250 g farine T55 ou autre, 50 g huile, 80 g eau tiède, 1 pincée de sel"},{ id: 301, directory: "Recipes", title: "Polish vegetarian cabbage rolls", slug: "/recipes/polish-vegetarian-cabbage-rolls", dateCreated: "2021-01-19", content: " 1 large cabbage, 2 cups brown rice, cooked, 1 lb veggie crumbles, 0.5 cup onion, minced, 1 garlic clove, minced, 1 tsp ground ginger, 1 tbsp sweet paprika, 1 tbsp parsley, chopped, 1 tsp black pepper, 2 eggs, 0.5 cup vegetable broth, 2 tbsp olive oil, 1 can diced tomatoes, 1 can tomato soup"},{ id: 302, directory: "Recipes", title: "Potato gratin", slug: "/recipes/potato-gratin", dateCreated: "2022-10-31", content: " Potatoes, Milk, Flour, Butter, Cheese"},{ id: 303, directory: "Recipes", title: "Potato soup", slug: "/recipes/potato-soup", dateCreated: "2023-03-01", content: " 8 pommes de terre de taille moyenne, 5 carottes de taille moyenne, 40 g de chorizo (ou autre saucisse séchée), 4 poignées de feuilles d'épinards, 1 oignon, 1 gousse d'ail pressée, Bouillon, Huile, Sel et poivre"},{ id: 304, directory: "Recipes", title: "Puff pastry", slug: "/recipes/puff-pastry", dateCreated: "2022-12-15", content: "This puff pastry recipe doesn't result in as many layers as a store-bought dough, but it does the job with minimal effort. 250 g plain flour, 280 g butter, very cold, 0.5 tsp salt, 120 ml of ice-cold water"},{ id: 305, directory: "Recipes", title: "Rice pilaf", slug: "/recipes/rice-pilaf", dateCreated: "2023-02-03", content: " Long grain rice, 1 onion, sliced, 2 garlic cloves, sliced, Vegetables, diced, 0.5 tsp coriander, 1 tsp paprika, 1 tsp Aleppo pepper, 1 tsp ground turmeric"},{ id: 306, directory: "Recipes", title: "Roasted vegetables", slug: "/recipes/roasted-vegetables", dateCreated: "2022-11-22", content: " vegetables, Olive oil"},{ id: 307, directory: "Recipes", title: "Shortbread", slug: "/recipes/shortbread", dateCreated: "2021-02-14", content: " 240 g butter at room temperature, 120 g sugar, 360 g flour, 1 pod vanilla, seeds scraped out, 0.5 tsp flaky sea salt"},{ id: 308, directory: "Recipes", title: "Slovenian Sauerkraut stew", slug: "/recipes/slovenian-sauerkraut-stew", dateCreated: "2022-04-03", content: " 1 tbsp paprika, 1 tsp chili powder, 500 g sauerkraut, 4 potatoes, peeled and chopped bite-size, 2 carrots, peeled and chopped bite-size, 3 cloves garlic, sliced, 1 onion, diced, 2–3 pickled cucumbers, sliced, 1 stock cube, oil, pepper and salt"},{ id: 309, directory: "Recipes", title: "Solid deodorant", slug: "/recipes/solid-deodorant", dateCreated: "2021-04-11", content: "This is the DIY deodorant recipe I have been using for many years now. Simple and effective. 60 g huile de coco bio, 48 g bicarbonate de soude, 30 g fécule de maïs, 24 gouttes huiles essentielles (palmarosa, tea tree, lavender)"},{ id: 310, directory: "Recipes", title: "Solid shampoo", slug: "/recipes/solid-shampoo", dateCreated: "2023-10-15", content: " 50 g de SCI (Sodium Cocoyl Isethionate), 15 g  de poudre de plante adaptée à vos cheveux (ex : poudre de bhringaraj), 17 g de phase huileuse (huiles végétales ou beurres végétaux, ex : huile de coco), 15 g de phase aqueuse (hydrolat, infusion, macérat aqueux ou eau, ex : hydrolat de romarin à verbénone), 2 g d'acide citrique"},{ id: 311, directory: "Recipes", title: "Sourdough bagels", slug: "/recipes/sourdough-bagels", dateCreated: "2021-01-22", content: " 125 g sourdough starter, 225 g water, 40 g honey (or sugar), 10 g fine salt, 500 g flour, 1 tbsp sugar (for boiling water)"},{ id: 312, directory: "Recipes", title: "Sourdough burger buns", slug: "/recipes/sourdough-burger-buns", dateCreated: "2021-01-22", content: " 200 g levain, 300 g farine T45, 200 g farine semi-complète T110, 8 g sel, 16 g sucre, 200 g lait à température ambiante (ou eau), 60 g beurre mou, 1 oeuf"},{ id: 313, directory: "Recipes", title: "Sourdough naan", slug: "/recipes/sourdough-naan", dateCreated: "2023-09-18", content: " 150 g water, 80 g yogurt, 75 g sourdough starter, 350 g flour, 0.5 tsp salt, 1.5 tbs sugar, 1 tbsp olive oil"},{ id: 314, directory: "Recipes", title: "Stuffed vine leaves", slug: "/recipes/stuffed-vine-leaves", dateCreated: "2023-07-31", content: "France's recipe, perfect for quick out-of-the-freezer starters. 1 tasse d'huile d'olive, 2 c.à.s. de pignons de pin, 2 oignons émincés, 1/2 tasse de riz blanc, 1/4 tasse de raisins secs, sel et poivre, un peu de menthe, 1/2 tasse d'eau chaude, 30 feuilles de vigne fraîches, 1 citron"},{ id: 315, directory: "Recipes", title: "Sushi / gimbap", slug: "/recipes/sushi-gimbap", dateCreated: "2024-05-12", content: "On fait une pause de deux mois à Hanoi lors de notre voyage de l'Europe à l'Asie. On y loue un appart pas cher avec une toute petite cuisine à peine fonctionnelle. Que peut-on préparer avec un minimum de 'cuisine' nécessaire et avec des ingrédients disponibles ? Je me lance dans des rouleaux inspirés des sushis japonais et gimbaps coréens. 3 bols de riz, 3 c-à-s vinaigre de riz, 1.5 c-à-s sucre, Sel"},{ id: 316, directory: "Recipes", title: "Tarte tatin with summer veggies", slug: "/recipes/tarte-tatin-with-summer-veggies", dateCreated: "2023-07-24", content: " Pie dough, 2 tomatoes, 1 zucchini, thinly sliced, 1 onion, thinly sliced"},{ id: 317, directory: "Recipes", title: "Tarte tatin", slug: "/recipes/tarte-tatin", dateCreated: "2023-09-13", content: " 1 pâte à tarte (feuilletée de préférence), 90 g de sucre, 1 c-à-s de miel, 1 c-à-s de vinaigre de cidre ou citron, 30 g de beurre, 1 grosse pincée de sel, 1 kg de pommes, coupées en quartiers, 1 sachet de sucre vanillé, Cannelle"},{ id: 318, directory: "Recipes", title: "Thai red curry paste (prig gang ped)", slug: "/recipes/thai-red-curry-paste", dateCreated: "2022-07-10", content: " 1 tbsp cilantro roots (or stems), 1 tbsp coriander seeds, 0.5 tbsp cumin seeds, 3 dried whole chilis, 1.5 tbsp galangal roots, 35 g garlic, 1 stalk lemongrass, 0.5 tsp peppercorns, 1 tabs. salt, 80 g shallots, 1 tbsp keffir lime zest (or lime zest)"},{ id: 319, directory: "Recipes", title: "Turnips with honey and zaatar", slug: "/recipes/turnip-with-honey-and-zaatar", dateCreated: "2023-03-06", content: "France's recipe: great as a side, can be eaten hot or cold. 1 turnip, Honey, Zaatar"},{ id: 320, directory: "Recipes", title: "Vegetarian gravy (upgraded version)", slug: "/recipes/vegetarian-gravy-upgraded", dateCreated: "2022-12-06", content: " 1 tbsp oil, 3 cloves garlic, chopped, 1 onion, chopped, 1–2 small pre-boiled potatoes, 7 g dried mushrooms, 1/2 cup red wine, 1/4 tsp cayenne pepper, 1/4 tsp cumin, 1 tsp smoked paprika, 1.5 tbsp soy sauce, 1 cup vegetable broth, Shiitake mushrooms, to taste, 1 pinch of salt"},{ id: 321, directory: "Recipes", title: "Vegetarian gravy", slug: "/recipes/vegetarian-gravy", dateCreated: "2022-02-10", content: " 8 tbsp oil, 1 onion, chopped, 3 cloves garlic, minced, 4 tbsp all-purpose flour, 2 tbsp light soy sauce, 1 cup vegetable broth, salt, black pepper, ground"},{ id: 322, directory: "Recipes", title: "Veggie loaf", slug: "/recipes/veggie-loaf", dateCreated: "2022-11-20", content: "Recette de Souad, toute simple et goûtue. 110 g farine blanche, 110 g farine d'épeautre complète, 2 oeufs, 150 g lait, 40 g huile végétale, 0.5 cc bicarbonate, 1 cc vinaigre blanc, Sel et poivre"},{ id: 323, directory: "Recipes", title: "Vietnamese cháo gà", slug: "/recipes/vietnamese-chao-ga", dateCreated: "2022-11-30", content: "Reminescent of childhood memories. Chao was my mother go-to recipe whenever someone in the family would get sick. 1 yellow onion, 1 whole chicken, 1 tbsp salt, 1.5 tbsp sugar, 250 g uncooked jasmine rice, 5&nbsp;cm knob fresh ginger, peeled and finely chopped, 1.5 tbsp fish sauce, 2 tsp freshly ground black pepper"},{ id: 324, directory: "Recipes", title: "Vietnamese đồ chua", slug: "/recipes/vietnamese-do-chua", dateCreated: "2023-02-25", content: " Carrots, peeled and julienned, 2 cups water, 1 cup vinegar (can be a mix of rice, white or apple vinegar), 0.5 cup sugar"},{ id: 325, directory: "Recipes", title: "Watermelon salad", slug: "/recipes/watermelon-salad", dateCreated: "2023-09-14", content: "A refreshing sweet and savory salad for hot summer days. Watermelon, Feta, Fresh mint"},{ id: 326, directory: "Recipes", title: "Whole roasted cauliflower", slug: "/recipes/whole-roasted-cauliflower", dateCreated: "2022-10-06", content: " 2 small heads cauliflower, 2 btsp olive oil, 1.25 tsp flake salt"},{ id: 327, directory: "Recipes", title: "Zucchini soup", slug: "/recipes/zucchini-soup", dateCreated: "2023-07-06", content: " Huile d'olive, 1 kg courgette, 2 oignons, 2 verres de lait et / ou eau, Sel et poivre"},];
