+++ scripts/prebuild.js
@@ -4,9 +4,11 @@ import generateRSS from "./generateRSS.js";
 import generateRecipePost from "./generateRecipePost.js";
 import generatePostIndex from "./generatePostIndex.js";
 
-const posts = await getAllPosts();
+const { posts, learn, media, recipes } = await getAllPosts();
+console.log(recipes);
 
 Promise.all([
   generateRSS(posts),
-  // , generateRecipePost(), generatePostIndex()
+  // , generateRecipePost()
+  generatePostIndex(posts, learn, media, recipes),
 ]);
